The circle group

LieGroups.CircleGroup — Method

CircleGroup

The circle group $𝕊^1$ is the unit circle together with composing points on the circle by adding angles. The circle itself is a one dimensional Riemannian manifold. Hence the Lie algebra is the real line.

The elements of the circle group can be represented in three different ways.

As complex numbers

Elements of the circle group can be represented as complex numbers of absolute value one, that is

\[𝕊¹ = \bigl\{ z ∈ ℂ\ \big|\ |z| = 1\bigr \} = \bigl\{ a + b\mathrm{i} ∈ ℂ\ \big|\ a^2+b^2 = 1\bigr \},\]

where $\mathrm{i}$ denotes the imaginary unit. It is equipped with the group operation of complex multiplication AbelianMultiplicationGroupOperation. That operation is given by

\[(a + b\mathrm{i}) ∘ (c + d\mathrm{i}) := (ac - bd) + (ad + bc)\mathrm{i},\]

for complex numbers $a + b\mathrm{i}, c + d\mathrm{i} ∈ ℂ$.

As angles in $[-π,π)$

Elements of the circle group can be represented by the angle on the unit circle that they correspond to. In that case the elements are represented by real numbers $x ∈ [-π,π)$ and the circle group is identified with a quotient space of the real numbers

\[ 𝕊¹ = ℝ / 2πℤ = \bigl\{ [x] ∈ ℝ / 2πℤ\ \big|\ x ∈ [-π,π)\bigr \}.\]

It is equipped with the group operation of adding angles $\mathrm{mod\, } 2π$ via AdditionGroupOperation.

As part of the 2D plane $ℝ^2$

Elements of the circle group can be represented as two dimensional real valued vectors $x ∈ ℝ$ of length 1. In that case the circle group is identified with the unit circle in $ℝ^2$, that is the one dimensional Sphere

\[𝕊^1 = \bigl\{ (x, y) ∈ ℝ^2\ \big|\ x^2 + y^2 = 1\bigr \}.\]

It is equipped with the group operation of adding the angles of two points on the unit circle which corresponds to complex multiplication

\[(x_1, y_1) ∘ (x_2, y_2) := ( x_1x_2 - y_1y_2, x_1y_2 + x_2y_1),\]

for real valued vectors $(x_1, y_1)^\mathrm{T}, (x_2, y_2)^\mathrm{T} ∈ ℝ^2$ via AbelianMultiplicationGroupOperation.

Constructors

CircleGroup(Circle(ℂ))
CircleGroup(ℂ)
CircleGroup()

Generate the circle group represented as complex numbers.

CircleGroup(Circle(ℝ))
CircleGroup(ℝ)

Generate the circle group represented as real valued angles $x ∈ [-π, π)$.

CircleGroup(Sphere(1))
CircleGroup(ℝ^2)

Generate the circle group represented as two dimensional real valued vectors of unit norm.

The default representation is by complex numbers and can be constructed with CircleGroup().

Representation on the real line

Base.exp — Method

exp(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X)
exp!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g, X)

Compute the Lie group exponential of a vector X of the LieAlgebra of the circle group, represented as angles in $[-π, π)$. In that case, the Lie algebra is the real line and the Lie group exponential of a vector $X ∈ ℝ$ is its equivalence class

\[ \exp(X) = [X] ∈ \bigl\{ [x] ∈ ℝ / 2πℤ\ \big|\ x ∈ [-π,π)\bigr \}.\]

This can be computed in-place of g.

Base.log — Method

log(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g)
log!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X, g)

Compute the Lie group logarithm on the CircleGroup, represented as angles in $[-π,π)$. The LieAlgebra is the real line and $\log$ is given by the identity map.

Formally $\log$ promotes an equivalence class $[X]$ to a representative $X∈ℝ$.

This can be computed in-place of X.

LieGroups.conjugate! — Method

conjugate(CircleGroup, g, h)
conjugate!(CircleGroup, k, g, h)

Compute the conjugation map $c_g: \mathcal G → \mathcal G$ given by $c_g(h) = g∘h∘g^{-1} = h$. It simplifies to the identity since the group operation on the circle group is abelian.

This can be computed in-place of k if k is mutable.

LieGroups.conjugate — Method

conjugate(CircleGroup, g, h)
conjugate!(CircleGroup, k, g, h)

Compute the conjugation map $c_g: \mathcal G → \mathcal G$ given by $c_g(h) = g∘h∘g^{-1} = h$. It simplifies to the identity since the group operation on the circle group is abelian.

This can be computed in-place of k if k is mutable.

LieGroups.diff_conjugate — Method

diff_conjugate(G::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g, h, X)
diff_conjugate!(G::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, Y, g, h, X)

Compute the differential of the conjugation map $c_g(h) = g∘h∘g^{-1}=h$. On the circle group represented as part of the real line, this simplifies to $D(c_g(h))[X] = X$.

This can be computed in-place of Y if Y is mutable.

LieGroups.diff_inv! — Method

diff_inv(G::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g, X)
diff_inv!(G::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, Y, g, X)

Compute the the differential $\mathrm{d}ι_{\mathcal G}([g])[X]$ of the inversion $ι_{\mathcal G}([g]) := [g]^{-1} = [-g]$ at $X ∈ 𝔤$ in the LieAlgebra $𝔤$ of the real CircleGroup G $=\mathcal G$.

\[\mathrm{d}ι_{\mathcal G}([g])[X] = -X.\]

This can be computed in-place of Y if Y is mutable.

LieGroups.diff_inv — Method

diff_inv(G::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g, X)
diff_inv!(G::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, Y, g, X)

Compute the the differential $\mathrm{d}ι_{\mathcal G}([g])[X]$ of the inversion $ι_{\mathcal G}([g]) := [g]^{-1} = [-g]$ at $X ∈ 𝔤$ in the LieAlgebra $𝔤$ of the real CircleGroup G $=\mathcal G$.

\[\mathrm{d}ι_{\mathcal G}([g])[X] = -X.\]

This can be computed in-place of Y if Y is mutable.

LieGroups.sym_rem — Method

sym_rem(x,[T=π])

Compute symmetric remainder of x with respect to the interall 2*T, i.e. (x+T)%2T, where the default for T is $π$.

ManifoldsBase.exp! — Method

exp(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X)
exp!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g, X)

Compute the Lie group exponential of a vector X of the LieAlgebra of the circle group, represented as angles in $[-π, π)$. In that case, the Lie algebra is the real line and the Lie group exponential of a vector $X ∈ ℝ$ is its equivalence class

\[ \exp(X) = [X] ∈ \bigl\{ [x] ∈ ℝ / 2πℤ\ \big|\ x ∈ [-π,π)\bigr \}.\]

This can be computed in-place of g.

ManifoldsBase.log! — Method

log(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g)
log!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X, g)

Compute the Lie group logarithm on the CircleGroup, represented as angles in $[-π,π)$. The LieAlgebra is the real line and $\log$ is given by the identity map.

Formally $\log$ promotes an equivalence class $[X]$ to a representative $X∈ℝ$.

This can be computed in-place of X.

Representation on the complex plane

Base.exp — Method

exp(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X)
exp!(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g, X)

Computes the Lie group exponential on the complex CircleGroup, which coincides with the ordinary complex exponential.

The Lie algebra is precisely the imaginary axis of the complex plane.

This can be computed in-place of g.

\[\exp (\mathrm{i}t) = \cos(t) + \mathrm{i}\sin(t)\]

Base.log — Method

log(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g)
log!(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X, g)

Compute the Lie group logarithm on the complex CircleGroup, which coincides with the ordinary complex logarithm.

This can be computed in-place of X.

LieGroups.diff_left_compose — Method

diff_left_compose(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g, h, X)
diff_left_compose(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, Y, g, h, X)

Compute the differential of the group operation $gh$ with respeect to the lefy argument g.

The formula reads

\[ \mathrm{d} ρ_h(g) = X.\]

since the group operation is abelian. This can be computed in-place of Y if Y is mutable.

LieGroups.diff_right_compose — Method

diff_right_compose(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g, h, X)
diff_right_compose(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, Y, g, h, X)

Compute the differential of the group operation $g∘h$, on an AbstractLieGroup G with respect to its second (right) argument h.

Another interpretation is to consider a function where we do a fixed multiplication from the left with g. i..e. the left group multiplication function $λ_g(h) = g∘h$.

It simplifies for the AbelianMultiplicationGroupOperation to $\mathrm{d}λ_g(h)[X] = X$. This can be computed in-place of Y.

ManifoldsBase.exp! — Method

exp(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X)
exp!(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g, X)

Computes the Lie group exponential on the complex CircleGroup, which coincides with the ordinary complex exponential.

The Lie algebra is precisely the imaginary axis of the complex plane.

This can be computed in-place of g.

\[\exp (\mathrm{i}t) = \cos(t) + \mathrm{i}\sin(t)\]

ManifoldsBase.log! — Method

log(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g)
log!(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X, g)

Compute the Lie group logarithm on the complex CircleGroup, which coincides with the ordinary complex logarithm.

This can be computed in-place of X.

Representation on the 2D-plane

Base.exp — Method

exp(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X)
exp!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g, X)

Compute the Lie group exponential on the CircleGroup, represented as two dimensional vectors in the real plane. It coincides with the ordinary complex exponential after canonical identification of the real plane with the complex plane.

This can be computed in-place of g.

\[\exp \begin{pmatrix} 0\\ t\end{pmatrix} = \begin{pmatrix} \cos(t)\\ \sin(t)\end{pmatrix}\]

Base.log — Method

log(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g)
log!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X, g)

Compute the Lie group logarithm on the CircleGroup, represented as two dimensional vectors in the real plane. It coincides with the ordinary complex logarithm after canonical identification of the real plane with the complex plane.

This can be computed in-place of X.

LieGroups.diff_left_compose! — Method

diff_left_compose(G::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g, h, X)
diff_left_compose!(G::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, Y, g, h, X)

Compute the differential of the group operation $gh$ with respect to the left argument $g$.

Here it simplifies for AbelianMultiplicationGroupOperation to $\mathrm{d}ρ_h(g)[X] = X$.

This can be computed in-place of Y.

LieGroups.diff_left_compose — Method

diff_left_compose(G::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g, h, X)
diff_left_compose!(G::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, Y, g, h, X)

Compute the differential of the group operation $gh$ with respect to the left argument $g$.

Here it simplifies for AbelianMultiplicationGroupOperation to $\mathrm{d}ρ_h(g)[X] = X$.

This can be computed in-place of Y.

LieGroups.diff_right_compose! — Method

diff_right_compose(G::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g, h, X)
diff_right_compose!(G::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, Y, g, h, X)

Compute the differential of the group operation $g∘h$, on an AbstractLieGroup G with respect to its second (right) argument h.

Another interpretation is to consider a function where we do a fixed multiplication from the left with g. i..e. the left group multiplication function $λ_g(h) = g∘h$.

It simplifies for the AbelianMultiplicationGroupOperation to $\mathrm{d}λ_g(h)[X] = X$,

ManifoldsBase.exp! — Method

exp(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X)
exp!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g, X)

Compute the Lie group exponential on the CircleGroup, represented as two dimensional vectors in the real plane. It coincides with the ordinary complex exponential after canonical identification of the real plane with the complex plane.

This can be computed in-place of g.

\[\exp \begin{pmatrix} 0\\ t\end{pmatrix} = \begin{pmatrix} \cos(t)\\ \sin(t)\end{pmatrix}\]

ManifoldsBase.log! — Method

log(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g)
log!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X, g)

Compute the Lie group logarithm on the CircleGroup, represented as two dimensional vectors in the real plane. It coincides with the ordinary complex logarithm after canonical identification of the real plane with the complex plane.

This can be computed in-place of X.