The general linear group

LieGroups.GeneralLinearGroup — Type

GeneralLinearGroup{𝔽,T}

The general linear group $\mathrm{GL}(n)$ is the set of all invertible matrices

\[\mathrm{GL}(n) = \bigl\{ M ∈ 𝔽^{n×n}\ \big|\ \mathrm{det}(M) ≠ 0\bigr \}, \qquad 𝔽 ∈ \{ ℝ, ℂ \},\]

equipped with the MatrixMultiplicationGroupOperation as the group operation.

The set of invertible matrices is a Riemannian manifold, since it inherits its structure from the embedding as an open subset of the space of matrices $ℝ^{n×n}$.

Constructor

GeneralLinearGroup(n::Int; field=ℝ, kwargs...)

Generate the general linear group on $𝔽^{n×n}$. All keyword arguments in kwargs... are passed on to InvertibleMatrices.

Base.exp — Method

exp(::GeneralLinearGroup, X)
exp!(::GeneralLinearGroup, g, X)

Compute the Lie group exponential on the GeneralLinearGroup, which is given by the matrix exponential

\[\exp X = \sum_{k=0}^{∞} \frac{1}{k!}X^k\]

see also [HN12, Example 9.2.3 (b)]

ManifoldsBase.exp! — Method

exp(::GeneralLinearGroup, X)
exp!(::GeneralLinearGroup, g, X)

Compute the Lie group exponential on the GeneralLinearGroup, which is given by the matrix exponential

\[\exp X = \sum_{k=0}^{∞} \frac{1}{k!}X^k\]

see also [HN12, Example 9.2.3 (b)]

exp(::GeneralLinearGroup, X) exp!(::GeneralLinearGroup, g, X)
Compute the Lie group exponential on the GeneralLinearGroup, which is given by the matrix exponential

GeneralLinearGroup{𝔽,T}
The general linear group is the set of all invertible matrices

exp(::GeneralLinearGroup, X) exp!(::GeneralLinearGroup, g, X)
Compute the Lie group exponential on the GeneralLinearGroup, which is given by the matrix exponential

exp(G::AbstractLieGroup, X::T) exp!(G::AbstractLieGroup, g, X)
Compute the (Lie group) exponential function

exp(G::AbstractLieGroup, X::T) exp!(G::AbstractLieGroup, g, X)
Compute the (Lie group) exponential function

exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, X) exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, g, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, g, X)
Compute the Lie group exponential on a LieGroup at a point g or the Identity with an AbelianMultiplicationGroupOperation.

exp(G::LieGroup{𝔽,AdditionGroupOperation}, X) exp!(G::LieGroup{𝔽,AdditionGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AdditionGroupOperation.

exp(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, X) exp!(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AbstractMultiplicationGroupOperation, which simplifies to the matrix exponential.

exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, X) exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, g, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, g, X)
Compute the Lie group exponential on a LieGroup at a point g or the Identity with an AbelianMultiplicationGroupOperation.

exp(G::LieGroup{𝔽,AdditionGroupOperation}, X) exp!(G::LieGroup{𝔽,AdditionGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AdditionGroupOperation.

exp(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, X) exp!(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AbstractMultiplicationGroupOperation, which simplifies to the matrix exponential.

exp(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X) exp!(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g, X)
Computes the Lie group exponential on the complex CircleGroup, which coincides with the ordinary complex exponential.

exp(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X) exp!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g, X)
Compute the Lie group exponential on the CircleGroup, represented as two dimensional vectors in the real plane.

exp(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X) exp!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g, X)
Compute the Lie group exponential of a vector X of the LieAlgebra of the circle group, represented as angles in .

exp(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X) exp!(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g, X)
Computes the Lie group exponential on the complex CircleGroup, which coincides with the ordinary complex exponential.

exp(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X) exp!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g, X)
Compute the Lie group exponential on the CircleGroup, represented as two dimensional vectors in the real plane.

exp(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X) exp!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g, X)
Compute the Lie group exponential of a vector X of the LieAlgebra of the circle group, represented as angles in .

exp(G::HeisenbergGroup, X) exp!(G::HeisenbergGroup, g, X)
Compute the Lie group exponential for the HeisenbergGroup G of the vector X.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{2}} for dispatch.

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{3}} for dispatch.

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{3}} for dispatch.

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{2}} for dispatch.

LieGroups.exp(M::SpecialGalileanGroup, X) LieGroups.exp!(M::SpecialGalileanGroup, h, X)
Compute the Lie group exponential function on the SpecialGalileanGroup(3), where X is an element of the Lie algebra.

LieGroups.exp(M::SpecialGalileanGroup, X) LieGroups.exp!(M::SpecialGalileanGroup, h, X)
Compute the Lie group exponential function on the SpecialGalileanGroup(3), where X is an element of the Lie algebra.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.