The Heisenberg group

LieGroups.HeisenbergGroup — Type

HeisenbergGroup{T}

The HeisenbergGroup(n) is the group of $(n+2)×(n+2)$ matrices, see also [BP08] or Heisenberg group where T specifies the eltype of the matrix entries.

\[\begin{pmatrix} 1 & \mathbf{a}^{\mathrm{T}} & c\\ \mathbf{0}_n & I_n & \mathbf{b}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 1\end{pmatrix},\]

where $I_n$ is the $n×n$ unit matrix, $\mathbf{a}, \mathbf{b} ∈ ℝ^n$ are vectors of length $n$, $\mathbf{0}_n$ is the zero vector of length $n$, and $c ∈ ℝ$ is a real number. The group operation is matrix multiplication.

The Lie algebra consists of the elements

\[\begin{pmatrix} 0 & \mathbf{a}^{\mathrm{T}} & c\\ \mathbf{0}_n & Z_n & \mathbf{b}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 0\end{pmatrix},\]

where additionally $Z_n$ denotes the $n×n$ zero matrix.

Base.exp — Method

exp(G::HeisenbergGroup, g, X)

Exponential map on the HeisenbergGroup G with the left-invariant metric.

We denote by g a point on the Heisenberg group and by $X$ a vector from the Lie algebra. These are of the form

\[g = \begin{pmatrix} 1 & \mathbf{a}^{\mathrm{T}} & c\\ \mathbf{0}_n & I_n & \mathbf{b}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 1\end{pmatrix} \qquad X = \begin{pmatrix} 0 & \mathbf{d}^{\mathrm{T}} & f\\ \mathbf{0}_n & Z_n & \mathbf{e}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 0\end{pmatrix},\]

where $I_n$ is the $n×n$ unit matrix, $Z_n$ is the $n×n$ zero matrix, $\mathbf{a}, \mathbf{b}, \mathbf{d}, \mathbf{e} ∈ ℝ^n$ are vectors of length $n$, $\mathbf{0}_n$ is the zero vector of length $n$, and $c,f ∈ ℝ$ are real numbers.

Then the formula reads

\[\exp_g(X) = \begin{pmatrix} 1 & (\mathbf{a}+\mathbf{d})^{\mathrm{T}} & c+f+\frac{1}{2}\mathbf{d}^{\mathrm{T}}\mathbf{e} + \mathbf{a}^{\mathrm{T}}\mathbf{e}\\ \mathbf{0}_n & I_n & \mathbf{b}+\mathbf{e}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 1\end{pmatrix}.\]

Base.exp — Method

exp(G::HeisenbergGroup, X)
exp!(G::HeisenbergGroup, g, X)

Compute the Lie group exponential for the HeisenbergGroup G of the vector X.

For $X = \begin{pmatrix} 0 & \mathbf{a}^{\mathrm{T}} & c\\ \mathbf{0}_n & Z_n & \mathbf{b}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 0\end{pmatrix}$ from the Lie algebra of the Heisenberg group, where $\mathbf{a}, \mathbf{b} ∈ ℝ^n$ vectors of length $n$, $\mathbf{0}_n$ is the zero vector of length $n$, $c ∈ ℝ$, and $Z_n$ denotes the $n×n$ zero matrix.

Then the

\[\exp_{\mathcal G}(X) = \begin{pmatrix} 1 & \mathbf{a}^{\mathrm{T}} & c + \frac{1}{2}\mathbf{a}^{\mathrm{T}}\mathbf{b}\\ \mathbf{0}_n & I_n & \mathbf{b}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 1\end{pmatrix},\]

where $I_n$ is the $n×n$ unit matrix.

This can be computed in-place of the Lie group element g.

Base.log — Method

log(G::HeisenbergGroup, g, h)

Compute the logarithmic map on the HeisenbergGroup group.

We denote two points $g, h$ from the Heisenberg by

\[g = \begin{pmatrix} 1 & \mathbf{a}^{\mathrm{T}} & c\\ \mathbf{0}_n & I_n & \mathbf{b}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 1\end{pmatrix} \qquad h = \begin{pmatrix} 1 & \mathbf{d}^{\mathrm{T}} & f\\ \mathbf{0}_n & I_n & \mathbf{e}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 1\end{pmatrix},\]

where $I_n$ is the $n×n$ unit matrix, $\mathbf{a}, \mathbf{b}, \mathbf{d}, \mathbf{e} ∈ ℝ^n$ are vectors of length $n$, $\mathbf{0}_n$ is the zero vector of length $n$, and $c,f ∈ ℝ$ are real numbers.

Then formula reads

\[\log_g(h) = \begin{pmatrix} 0 & (\mathbf{d}-\mathbf{q})^{\mathrm{T}} & f - c + \mathbf{a}^{\mathrm{T}}\mathbf{b} - \mathbf{d}^{\mathrm{T}}\mathbf{e} - \frac{1}{2}(\mathbf{d}-\mathbf{a})^{\mathrm{T}}(\mathbf{e}-\mathbf{b})\\ \mathbf{0}_n & Z_n & \mathbf{e} - \mathbf{b}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 0\end{pmatrix},\]

where additionally $Z_n$ denotes the $n×n$ zero matrix.

Base.log — Method

log(G::HeisenbergGroup, g)
log!(G::HeisenbergGroup, X, g)

Compute the Lie group logarithm for the HeisenbergGroup G.

For $g = \begin{pmatrix} 1 & \mathbf{a}^{\mathrm{T}} & c\\ \mathbf{0}_n & I_n & \mathbf{b}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 1\end{pmatrix}$ from the Lie algebra of the Heisenberg group, where $\mathbf{a}, \mathbf{b} ∈ ℝ^n$ vectors of length $n$, $\mathbf{0}_n$ is the zero vector of length $n$, $c ∈ ℝ$, and $I_n$ is the $n×n$ unit matrix.

Then the

\[\log_{\mathcal G}(g) = \begin{pmatrix} 1 & \mathbf{a}^{\mathrm{T}} & c - \frac{1}{2}\mathbf{a}^{\mathrm{T}}\mathbf{b}\\ \mathbf{0}_n & Z_n & \mathbf{b}\\ 0 & \mathbf{0}_n^{\mathrm{T}} & 1\end{pmatrix},\]

where $Z_n$ denotes the $n×n$ zero matrix.

This can be computed in-place of the Lie algebra vector X.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(G::HeisenbergGroup)

Return the injectivity radius on the HeisenbergGroup G, which is $∞$.

Literature

[BP08]

E. Biny and S. Pods. The Geometry of Heisenberg Groups: With Applications in Signal Theory, Optics, Quantization, and Field Quantization (American Mathematical Society, 2008).

HeisenbergGroup{T}
The HeisenbergGroup(n) is the group of matrices, see also [BP08] or Heisenberg group where T specifies the eltype of the matrix entries.

exp(G::AbstractLieGroup, X::T) exp!(G::AbstractLieGroup, g, X)
Compute the (Lie group) exponential function

log(G::AbstractLieGroup, g, h) log(G::AbstractLieGroup, g) log(G::AbstractLieGroup, g::Identity, T) log!(G::AbstractLieGroup, X::T, g)
Compute the (Lie group) logarithmic function , which is the inverse of the Lie group exponential function.

exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, X) exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, g, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, g, X)
Compute the Lie group exponential on a LieGroup at a point g or the Identity with an AbelianMultiplicationGroupOperation.

exp(G::LieGroup{𝔽,AdditionGroupOperation}, X) exp!(G::LieGroup{𝔽,AdditionGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AdditionGroupOperation.

exp(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, X) exp!(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AbstractMultiplicationGroupOperation, which simplifies to the matrix exponential.

log(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h) log(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, g, h) log!(G::LieGroup{𝔽,<:AbeliantMultiplicationGroupOperation}, X, g) log!(G::LieGroup{𝔽,<:AbeliantMultiplicationGroupOperation}, X, g, h)
Compute the Lie group logarithm on a LieGroup at a point g or the Identity with a concrete instance of AbelianMultiplicationGroupOperation.

log(G::LieGroup{𝔽,AdditionGroupOperation}, g) log!(G::LieGroup{𝔽,AdditionGroupOperation}, X, g)
Compute the Lie group logarithm on a LieGroup with an AdditionGroupOperation.

log(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, g) log!(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, X, g)
Compute the Lie group logarithm on a LieGroup with a concrete instance of AbstractMultiplicationGroupOperation, which simplifies to the (matrix) logarithm.

exp(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X) exp!(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g, X)
Computes the Lie group exponential on the complex CircleGroup, which coincides with the ordinary complex exponential.

exp(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X) exp!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g, X)
Compute the Lie group exponential on the CircleGroup, represented as two dimensional vectors in the real plane.

exp(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X) exp!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g, X)
Compute the Lie group exponential of a vector X of the LieAlgebra of the circle group, represented as angles in .

log(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g) log!(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X, g)
Compute the Lie group logarithm on the complex CircleGroup, which coincides with the ordinary complex logarithm.

log(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g) log!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X, g)
Compute the Lie group logarithm on the CircleGroup, represented as two dimensional vectors in the real plane.

log(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g) log!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X, g)
Compute the Lie group logarithm on the CircleGroup, represented as angles in .

exp(::GeneralLinearGroup, X) exp!(::GeneralLinearGroup, g, X)
Compute the Lie group exponential on the GeneralLinearGroup, which is given by the matrix exponential

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{3}} for dispatch.

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{2}} for dispatch.

log(G::SpecialEuclidean, g) log!(G::SpecialEuclidean, X, g)
Compute the Lie group logarithm function on the SpecialEuclideanGroup G, and G uses a TypeParameter{Tuple{2}} for dispatch.

log(G::SpecialEuclidean, g) log!(G::SpecialEuclidean, X, g)
Compute the Lie group logarithm function on the SpecialEuclideanGroup G, where e is the Identity on G uses a TypeParameter{Tuple{3}} for dispatch.

LieGroups.exp(M::SpecialGalileanGroup, X) LieGroups.exp!(M::SpecialGalileanGroup, h, X)
Compute the Lie group exponential function on the SpecialGalileanGroup(3), where X is an element of the Lie algebra.

LieGroups.log(M::SpecialGalileanGroup, g) LieGroups.log!(M::SpecialGalileanGroup, X, g)
Compute the Lie group logarithm function on the SpecialGalileanGroup(3), where g is a group element.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.