The product Lie group

LieGroups.AbstractProductGroupOperation — Type

AbstractProductGroupOperation <: AbstractGroupOperation

An abstract type to model group operations on a product manifold

LieGroups.ProductGroupOperation — Type

ProductGroupOperation{O<:NTuple{N,AbstractGroupOperation} where N} <: AbstractProductGroupOperation

A struct do model a tuple of group operations, one for each factor of a product group, that together forms a new group operation.

Access to the single operations can be done by pgo[i].

Constructor

ProductGroupOperation(o::AbstractGroupOperation...)
×(o::AbstractGroupOperation...) = ProductGroupOperation(o...)

LieGroups.ProductLieGroup — Method

ProductLieGroup(G, H, ...)

Return the LieGroup of the product of Lie groups G and H.

Alternatively, the short hand G × H can be used.

LinearAlgebra.cross — Method

cross(O1::AbstractGroupOperation, O2::AbstractGroupOperation)
O1 × O2
O1 × O2 × O3 × ...

Return the ProductGroupOperation For two AbstractGroupOperation` O1 and O2, where for the case that one of them is a ProductGroupOperation itself, the other is either prepended (if O1 is a product) or appended (if O2 is). If both are product operations, they are combined into one, keeping the order of operations.

For the case that more than two are concatenated with × this is iterated.

LinearAlgebra.cross — Method

cross(G::LieGroup, H::LieGroup)
G × H
G1 × G2 × G3 × ...

Return the ProductLieGroup For two LieGroups G and H, where for the case that one of them is a ProductLieGroup itself, the other is either prepended (if H is a product) or appended (if G is). If both are product Lie groups, they are combined into one, keeping the order of operations.

For the case that more than two are concatenated with × this is iterated.