The special linear group

LieGroups.SpecialLinearGroup — Type

SpecialLinear{𝔽,T}

The special linear group $\mathrm{SL}(n,𝔽)$ is the group of all invertible matrices with unit determinant in $𝔽^{n×n}$ and the MatrixMultiplicationGroupOperation as group operation.

The Lie algebra $\mathfrak sl(n, 𝔽) = T_e \mathrm{SL}(n,𝔽)$ is the set of all matrices in $𝔽^{n×n}$ with trace of zero. By default, tangent vectors $X_p ∈ T_p \mathrm{SL}(n,𝔽)$ for $p ∈ \mathrm{SL}(n,𝔽)$ are represented with their corresponding Lie algebra vector $X_e = p^{-1}X_p ∈ 𝔰𝔩(n, 𝔽)$.

Constructor

GeneralLinearGroup(n::Int, field=ℝ; kwargs...)

Generate the general linear group group on $𝔽^{n×n}$. All keyword arguments in kwargs... are passed on to DeterminantOneMatrices.

ManifoldsBase.hat! — Method

X = hat(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c)
hat!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X, c)

Compute the hat map $(⋅)^{\wedge} : ℝ^{n^2-1} → 𝔤$ that turns a vector of coordinates c into a tangent vector in the Lie algebra.

The formula on the Lie algebra $𝔤$ of the SpecialLinearGroup(n) is given by reshaping $c ∈ ℝ^{n^2-1}$ into an $n$-by$n$ matrix $X$ with the final entry X[n,n] initialised to zero and then set to the trace of this initial matrix.

This can be computed in-place of X.

ManifoldsBase.hat — Method

X = hat(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c)
hat!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X, c)

Compute the hat map $(⋅)^{\wedge} : ℝ^{n^2-1} → 𝔤$ that turns a vector of coordinates c into a tangent vector in the Lie algebra.

The formula on the Lie algebra $𝔤$ of the SpecialLinearGroup(n) is given by reshaping $c ∈ ℝ^{n^2-1}$ into an $n$-by$n$ matrix $X$ with the final entry X[n,n] initialised to zero and then set to the trace of this initial matrix.

This can be computed in-place of X.

ManifoldsBase.vee! — Method

c = vee(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X)
vee!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c, X)

Compute the vee map $(⋅)^{\vee}: \mathfrak g → ℝ^{n^2-1}$ that maps a tangent vector from the Lie algebra to a vector of coordinates c.

The formula on the Lie algebra $𝔤$ of the SpecialLinearGroup(n) is given by reshaping $X ∈ ℝ^{n×n}$ into a vector and omitting the last entry, since that can be reconstructed by considering that $X$ has to be of trace zero.

This can be computed in-place of c.

ManifoldsBase.vee — Method

c = vee(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X)
vee!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c, X)

Compute the vee map $(⋅)^{\vee}: \mathfrak g → ℝ^{n^2-1}$ that maps a tangent vector from the Lie algebra to a vector of coordinates c.

The formula on the Lie algebra $𝔤$ of the SpecialLinearGroup(n) is given by reshaping $X ∈ ℝ^{n×n}$ into a vector and omitting the last entry, since that can be reconstructed by considering that $X$ has to be of trace zero.

This can be computed in-place of c.

SpecialLinear{𝔽,T}
The special linear group is the group of all invertible matrices with unit determinant in and the MatrixMultiplicationGroupOperation as group operation.

X = hat(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c) hat!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X, c)
Compute the hat map that turns a vector of coordinates c into a tangent vector in the Lie algebra.

X = hat(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c) hat!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X, c)
Compute the hat map that turns a vector of coordinates c into a tangent vector in the Lie algebra.

c = vee(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X) vee!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c, X)
Compute the vee map that maps a tangent vector from the Lie algebra to a vector of coordinates c.

c = vee(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X) vee!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c, X)
Compute the vee map that maps a tangent vector from the Lie algebra to a vector of coordinates c.

LieAlgebra{𝔽, G} <: AbstractManifold{𝔽}
Represent the Lie algebra , that is a vector space with an associated lie_bracket which fulfills

hat(G::LieAlgebra, c) hat(G::LieAlgebra, c, T::Type) hat!(G::LieAlgebra, X::T, c)
Compute the hat map that maps a vector of coordinates , to a tangent vector .

hat(G::LieAlgebra, c) hat(G::LieAlgebra, c, T::Type) hat!(G::LieAlgebra, X::T, c)
Compute the hat map that maps a vector of coordinates , to a tangent vector .

vee(𝔤::LieAlgebra, X) vee!(𝔤::LieAlgebra, c, X)
Compute the vee map that maps a tangent vector X from the LieAlgebra to its coordinates with respect to the DefaultLieAlgebraOrthogonalBasis basis in the Lie algebra

vee(𝔤::LieAlgebra, X) vee!(𝔤::LieAlgebra, c, X)
Compute the vee map that maps a tangent vector X from the LieAlgebra to its coordinates with respect to the DefaultLieAlgebraOrthogonalBasis basis in the Lie algebra

MatrixMultiplicationGroupOperation <: AbstractMultiplicationGroupOperation
A group operation that is realised by a matrix multiplication.

X = hat(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, c) hat!(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, X, c)
Compute the hat map that turns a vector of coordinates c into a tangent vector in the Lie algebra.

X = hat(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, c) hat!(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, X, c)
Compute the hat map that turns a vector of coordinates c into a tangent vector in the Lie algebra.

c = vee(𝔤::LieAlgebra{ℝ,<:LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, X) vee!(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, c, X)
Compute the vee map that maps a tangent vector from the Lie algebra to a vector of coordinates c.

c = vee(𝔤::LieAlgebra{ℝ,<:LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, X) vee!(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, c, X)
Compute the vee map that maps a tangent vector from the Lie algebra to a vector of coordinates c.