The special orthogonal group

LieGroups.SpecialOrthogonalGroup — Type

SpecialOrthogonalGroup{T}

The special orthogonal group $\mathrm{SO}(n)$ is the Lie group consisting of the MatrixMultiplicationGroupOperation on the manifold of rotations Rotations.

Constructor

SpecialOrthogonalGroup(n; kwargs...)

Generate special orthogonal group $\mathrm{SO}(n)$. All keyword arguments in kwargs... are passed on to Rotations as well.

For this Lie group, a lot of implementations are already covered by the defaults in the generic (matrix) multiplication operation.

Functions

Base.exp — Method

exp(G::OrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, X)
exp(G::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, X)
exp!(G::OrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, g, X)
exp!(G::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, g, X)

Compute the Lie group exponential function on the OrthogonalGroup\mathrm{O}(2)or [SpecialOrthogonalGroup](@ref)\mathrm{SO}(2)`.

Since the Lie algebra of both groups agrees and consist of the set of skew symmetric matrices, these simplify for the case of $2×2$ matrices to $X=\begin{pmatrix}\n 0 & -α\\\n α & 0\n\end{pmatrix}{pmatrix}$, for some $α∈ℝ$.

Their exponential is

\[\exp_{\mathcal G}(X) = \begin{pmatrix}\n \cos(α) & -\sin(α)\\\n \sin(α) & \cos(α)\n\end{pmatrix}.\]

This result can be computed in-place of g.

Note that since $\mathrm{SO}(2)$ consists of two disjoint connected components and the exponential map is smooth, the result $g$ always lies in the connected component of the identity.

Base.log — Method

log(G::OrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, g)
log(G::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, g)
log!(G::OrthogonalGroup{TypeParameter{Tuple{2}}}, X, ::Identity{MatrixMultiplicationGroupOperation}, g)
log!(G::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, X, ::Identity{MatrixMultiplicationGroupOperation}, g)

Compute the Lie group logarithm function on the OrthogonalGroup\mathrm{O}(2)or [SpecialOrthogonalGroup](@ref)\mathrm{SO}(2)`.

For the two-dimensional case, any rotation matrix $g$ can be represented as $\begin{pmatrix}\n \cos(α) & -\sin(α)\\\n \sin(α) & \cos(α)\n\end{pmatrix}$. For the orthogonal group, $g$ might also include one further reflection.

The logarithm is then

\[\log_{\mathcal G}(g) = \begin{pmatrix}\n 0 & -α &\\\n α & 0\n\end{pmatrix}.\]

This result can be computed in-place of X

Note the logarithmic map is only locally around the identity uniquely determined. Especially, since $\mathrm{SO}(2)$ consists of two disjoint connected components and the exponential map is smooth, for any $g$ in the other component, the logarithmic map is defined, but not the inverse of the exponential map.

ManifoldsBase.exp! — Method

exp(G::OrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, X)
exp(G::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, X)
exp!(G::OrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, g, X)
exp!(G::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, g, X)

Compute the Lie group exponential function on the OrthogonalGroup\mathrm{O}(2)or [SpecialOrthogonalGroup](@ref)\mathrm{SO}(2)`.

Since the Lie algebra of both groups agrees and consist of the set of skew symmetric matrices, these simplify for the case of $2×2$ matrices to $X=\begin{pmatrix}\n 0 & -α\\\n α & 0\n\end{pmatrix}{pmatrix}$, for some $α∈ℝ$.

Their exponential is

\[\exp_{\mathcal G}(X) = \begin{pmatrix}\n \cos(α) & -\sin(α)\\\n \sin(α) & \cos(α)\n\end{pmatrix}.\]

This result can be computed in-place of g.

Note that since $\mathrm{SO}(2)$ consists of two disjoint connected components and the exponential map is smooth, the result $g$ always lies in the connected component of the identity.

ManifoldsBase.log! — Method

log(G::OrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, g)
log(G::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, ::Identity{MatrixMultiplicationGroupOperation}, g)
log!(G::OrthogonalGroup{TypeParameter{Tuple{2}}}, X, ::Identity{MatrixMultiplicationGroupOperation}, g)
log!(G::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, X, ::Identity{MatrixMultiplicationGroupOperation}, g)

Compute the Lie group logarithm function on the OrthogonalGroup\mathrm{O}(2)or [SpecialOrthogonalGroup](@ref)\mathrm{SO}(2)`.

For the two-dimensional case, any rotation matrix $g$ can be represented as $\begin{pmatrix}\n \cos(α) & -\sin(α)\\\n \sin(α) & \cos(α)\n\end{pmatrix}$. For the orthogonal group, $g$ might also include one further reflection.

The logarithm is then

\[\log_{\mathcal G}(g) = \begin{pmatrix}\n 0 & -α &\\\n α & 0\n\end{pmatrix}.\]

This result can be computed in-place of X

Note the logarithmic map is only locally around the identity uniquely determined. Especially, since $\mathrm{SO}(2)$ consists of two disjoint connected components and the exponential map is smooth, for any $g$ in the other component, the logarithmic map is defined, but not the inverse of the exponential map.