The interface for Lie Groups

LieGroups.LieGroup — Type

LieGroup{𝔽,M<:AbstractManifold{𝔽}, O<:AbstractGroupOperation}

Represent a Lie Group $\mathcal G$.

A Lie Group $\mathcal G$ is a group endowed with the structure of a manifold such that the group operations $∘: \mathcal G×\mathcal G → \mathcal G$, see compose and the inverse operation $⋅^{-1}: \mathcal G → \mathcal G$, see inv are smooth, see for example [HN12, Definition 9.1.1].

Lie groups are named after the Norwegian mathematician Marius Sophus Lie (1842–1899).

Fields

manifold: an AbstractManifold $\mathcal M$

op: an AbstractGroupOperation $∘$ on that manifold

Constructor

LieGroup(M::AbstractManifold, op::AbstractGroupOperation; vectors=LeftInvariantRepresentation())

Generate a Lie group based on a manifold M and a group operation op, where vectors by default are stored in the Lie Algebra.

LieGroups.LieAlgebra — Type

LieAlgebra{𝔽, G} <: AbstractManifold{𝔽}

Represent the Lie Algebra $\mathfrak g$, that is a $𝔽$vector space with an associated lie_bracket $[⋅,⋅]: \mathfrak g×\mathfrak g → \mathfrak g$ which fulfills

$[X,X] = 0$ for all $X ∈ \mathfrak g$

The Jacobi identity holds $[X, [Y,Z]] = [[X,Y],Z] = [Y, [X,Z]]$ holds for all $X, Y, Z ∈ \mathfrak g$.

The Lie algebras considered here are those related to a LieGroup $\mathcal G$, namely the tangent space the tangent space $T_{\mathrm e}\mathcal G$ at the Identity, this is internally just a const of the corresponding TangentSpace.

Constructor

LieAlgebra(G::LieGroup)

Return the Lie Algebra belonging to the LieGroup G.

Group Operations

LieGroups.AbstractGroupOperation — Type

AbstractGroupOperation

Represent a type of group operation for a LieGroup $\mathcal G$, that is a smooth binary operation $∘ : \mathcal G × \mathcal G → \mathcal G$ on elements of a Lie group $\mathcal G$.

LieGroups.Identity — Type

Identity{O<:AbstractGroupOperation}

Represent the group identity element $e ∈ \mathcal G$ on a LieGroup $\mathcal G$ with AbstractGroupOperation of type O.

Similar to the philosophy that points are agnostic of their group at hand, the identity does not store the group $\mathcal G$ it belongs to. However it depends on the type of the AbstractGroupOperation used.

See also identity_element on how to obtain the corresponding AbstractManifoldPoint or array representation.

Constructors

Identity(::LieGroup{𝔽,M,O}) where {𝔽,M,O<:AbstractGroupOperation}
Identity(o::AbstractGroupOperation)
Identity(::Type{AbstractGroupOperation})

create the identity of the corresponding subtype O<:AbstractGroupOperation

Functions on Lie Groups

Base.adjoint — Function

adjoint(G::LieGroup, g X)
adjoint!(G::LieGroup, Y, g, X)

Compute the adjoint $\mathrm{Ad}(g): \mathfrak g → \mathfrak g$, which is defined as the differential diff_conjugate of the conjugate $c_g(h) = g∘h∘g^{-1}$ evaluated at the Identity $h=\mathrm e$. The operation can be performed in-place of Y.

\[ \mathrm{Ad}(g)[X] = D c_g(\mathrm e) [X], \qquad X ∈ \mathfrak g.\]

see [HN12, Section 9.2.3].

On matrix Lie groups the adjoint reads $\mathrm{Ad}(g)[X] = g∘X∘g^{-1}$.

LieGroups.adjoint! — Function

adjoint(G::LieGroup, g X)
adjoint!(G::LieGroup, Y, g, X)

Compute the adjoint $\mathrm{Ad}(g): \mathfrak g → \mathfrak g$, which is defined as the differential diff_conjugate of the conjugate $c_g(h) = g∘h∘g^{-1}$ evaluated at the Identity $h=\mathrm e$. The operation can be performed in-place of Y.

\[ \mathrm{Ad}(g)[X] = D c_g(\mathrm e) [X], \qquad X ∈ \mathfrak g.\]

see [HN12, Section 9.2.3].

On matrix Lie groups the adjoint reads $\mathrm{Ad}(g)[X] = g∘X∘g^{-1}$.

ManifoldsBase.base_manifold — Function

base_manifold(G::LieGroup)

Return the manifold stored within the LieGroup G.

base_manifold(A::AbstractGroupAction)

Return the AbstractManifold of the AbstractGroupAction.

LieGroups.compose — Function

compose(G::LieGroup, g, h)
compose!(G::LieGroup, k, g, h)

Perform the group oepration $g ∘ h$ for two $g, h ∈ \mathcal G$ on the LieGroup G. This can also be done in-place of h.

LieGroups.compose! — Function

compose(G::LieGroup, g, h)
compose!(G::LieGroup, k, g, h)

Perform the group oepration $g ∘ h$ for two $g, h ∈ \mathcal G$ on the LieGroup G. This can also be done in-place of h.

LieGroups.diff_left_compose — Function

diff_left_compose(G::LieGroup, g, h, X)
diff_left_compose!(G::LieGroup, Y, g, h, X)

Compute the differential of the left group multiplication $λ_g(h) = g∘h$, on the LieGroup G, that is Compute $Dλ_g(h)[X]$, $X ∈ 𝔤$. This can be done in-place of Y.

LieGroups.diff_left_compose! — Function

diff_left_compose(G::LieGroup, g, h, X)
diff_left_compose!(G::LieGroup, Y, g, h, X)

Compute the differential of the left group multiplication $λ_g(h) = g∘h$, on the LieGroup G, that is Compute $Dλ_g(h)[X]$, $X ∈ 𝔤$. This can be done in-place of Y.

LieGroups.diff_right_compose — Function

diff_right_compose(G::LieGroup, h, g, X)
diff_right_compose!(G::LieGroup, Y, h, g, X)

Compute the differential of the right group multiplication $ρ_g(h) = h∘g$, on the LieGroup G, that is Compute $Dρ_g(h)[X]$, $X ∈ 𝔤$ This can be done in-place of Y.

LieGroups.diff_right_compose! — Function

diff_right_compose(G::LieGroup, h, g, X)
diff_right_compose!(G::LieGroup, Y, h, g, X)

Compute the differential of the right group multiplication $ρ_g(h) = h∘g$, on the LieGroup G, that is Compute $Dρ_g(h)[X]$, $X ∈ 𝔤$ This can be done in-place of Y.

LieGroups.inv_left_compose — Function

inv_left_compose(G::LieGroup, g, h)
inv_left_compose!(G::LieGroup, k, g, h)

Compute the inverse of the left group multiplication $λ_g(h) = g∘h$, on the LieGroup G, that is, compute $λ_g^{-1}(h) = g^{-1}∘h$. This can be done in-place of k.

LieGroups.inv_left_compose! — Function

compose(G::LieGroup, g, h)
compose!(G::LieGroup, k, g, h)

Perform the group oepration $g ∘ h$ for two $g, h ∈ \mathcal G$ on the LieGroup G. This can also be done in-place of h.

LieGroups.inv_right_compose — Function

inv_right_compose(G::LieGroup, h, g)
inv_right_compose!(G::LieGroup, k, h, g)

Compute the inverse of the right group multiplication $ρ_g(h) = h∘g$, on the LieGroup G, that is Compute $ρ_g^{-1}(h) = h∘g^{-1}$. This can be done in-place of k.

LieGroups.inv_right_compose! — Function

inv_right_compose(G::LieGroup, h, g)
inv_right_compose!(G::LieGroup, k, h, g)

Compute the inverse of the right group multiplication $ρ_g(h) = h∘g$, on the LieGroup G, that is Compute $ρ_g^{-1}(h) = h∘g^{-1}$. This can be done in-place of k.

LieGroups.conjugate — Function

conjugate(G::LieGroup, g, h)
conjugate!(G::LieGroup, k, g, h)

Compute the conjugation map $c_g: \mathcal G → \mathcal G$ given by $c_g(h) = g∘h∘g^{-1}$. This can be done in-place of k.

LieGroups.conjugate! — Function

conjugate(G::LieGroup, g, h)
conjugate!(G::LieGroup, k, g, h)

Compute the conjugation map $c_g: \mathcal G → \mathcal G$ given by $c_g(h) = g∘h∘g^{-1}$. This can be done in-place of k.

LieGroups.diff_conjugate — Function

diff_conjugate(G::LieGroup, g, h, X)
diff_conjugate!(G::LieGroup, Y, g, h, X)

Compute the differential of the conjugate $c_g(h) = g∘h∘g^{-1}$, which can be performed in-place of Y.

\[ D(c_g(h))[X], \qquad X ∈ \mathfrak g.\]

LieGroups.diff_conjugate! — Function

diff_conjugate(G::LieGroup, g, h, X)
diff_conjugate!(G::LieGroup, Y, g, h, X)

Compute the differential of the conjugate $c_g(h) = g∘h∘g^{-1}$, which can be performed in-place of Y.

\[ D(c_g(h))[X], \qquad X ∈ \mathfrak g.\]

Base.exp — Function

exp(G::LieGroup, g, X, t::Number=1)
exp!(G::LieGroup, h, g, X, t::Number=1)

Compute the Lie group exponential map given by

\[\exp_g X = g∘\exp_{\mathcal G}(X)\]

where X can be scaled by t, the computation can be performed in-place of h, and $\exp_{\mathcal G}$ denotes the Lie group exponential function.

Note

If g is the Identity the Lie group exponential function is computed directly. Implementing the Lie group exponential function introduces a default implementation for this function.

Note

The Lie group exponential map is usually different from the exponential map with respect to the metric of the underlying Riemannian manifold $\mathcal M$. To access the (Riemannian) exponential map, use exp(base_manifoldG, g, X).

exp(G::LieGroup, e::Identity, X)
exp!(G::LieGroup, h, e::Identity, X)

Compute the (Lie group) exponential function

\[\exp_{\mathcal G}: \mathfrak g → \mathcal G,\qquad \exp_{\mathcal G}(X) = γ_X(1),\]

where $γ_X$ is the unique solution of the initial value problem

\[γ(0) = \mathrm e, \quad γ'(t) = γ(t)⋅X.\]

The computation can be performed in-place of h.

See also [HN12, Definition 9.2.2].

ManifoldsBase.exp! — Function

exp(G::LieGroup, g, X, t::Number=1)
exp!(G::LieGroup, h, g, X, t::Number=1)

Compute the Lie group exponential map given by

\[\exp_g X = g∘\exp_{\mathcal G}(X)\]

where X can be scaled by t, the computation can be performed in-place of h, and $\exp_{\mathcal G}$ denotes the Lie group exponential function.

Note

If g is the Identity the Lie group exponential function is computed directly. Implementing the Lie group exponential function introduces a default implementation for this function.

Note

The Lie group exponential map is usually different from the exponential map with respect to the metric of the underlying Riemannian manifold $\mathcal M$. To access the (Riemannian) exponential map, use exp(base_manifoldG, g, X).

exp(G::LieGroup, e::Identity, X)
exp!(G::LieGroup, h, e::Identity, X)

Compute the (Lie group) exponential function

\[\exp_{\mathcal G}: \mathfrak g → \mathcal G,\qquad \exp_{\mathcal G}(X) = γ_X(1),\]

where $γ_X$ is the unique solution of the initial value problem

\[γ(0) = \mathrm e, \quad γ'(t) = γ(t)⋅X.\]

The computation can be performed in-place of h.

See also [HN12, Definition 9.2.2].

LieGroups.identity_element — Function

identity_element(G::LieGroup)
identity_element!(G::LieGroup, g)

Return a point representation of the Identity on the LieGroup G. By default this representation is the default array or number representation. It should return the corresponding default representation of $e$ as a point on G if points are not represented by arrays. This can be performed in-place of g.

LieGroups.identity_element! — Function

identity_element(G::LieGroup)
identity_element!(G::LieGroup, g)

Return a point representation of the Identity on the LieGroup G. By default this representation is the default array or number representation. It should return the corresponding default representation of $e$ as a point on G if points are not represented by arrays. This can be performed in-place of g.

LieGroups.is_identity — Function

is_identity(G::LieGroup, q; kwargs)

Check whether q is the identity on the LieGroup $\mathcal G$. This means it is either the Identity{O} with the respect to the corresponding AbstractGroupOperation O, or (approximately) the correct point representation.

See also

identity_element, identity_element!

Base.inv — Method

inv(G::LieGroup, g)
inv!(G::LieGroup, h, g)

Compute the inverse group element $g^{-1}$ with respect to the AbstractGroupOperation $∘$ on the LieGroup $\mathcal G$, that is, return the unique element $h=g^{-1}$ such that $h∘g=\mathrm e$, where $\mathrm e$ denotes the Identity.

This can be done in-place of h, without side effects, that is you can do inv!(G, g, g).

LieGroups.inv! — Function

inv(G::LieGroup, g)
inv!(G::LieGroup, h, g)

Compute the inverse group element $g^{-1}$ with respect to the AbstractGroupOperation $∘$ on the LieGroup $\mathcal G$, that is, return the unique element $h=g^{-1}$ such that $h∘g=\mathrm e$, where $\mathrm e$ denotes the Identity.

This can be done in-place of h, without side effects, that is you can do inv!(G, g, g).

LieGroups.diff_inv — Function

diff_inv(G::LieGroup, g, X)
diff_inv!(G::LieGroup, Y, g, X)

Compute the differential of the function $ι_{\mathcal G}(p) = p^-1$, where `Dι_{\mathcal G}(p): \mathfrak g → \mathfrak g. This can be done in-place ofY`.

LieGroups.diff_inv! — Function

diff_inv(G::LieGroup, g, X)
diff_inv!(G::LieGroup, Y, g, X)

Compute the differential of the function $ι_{\mathcal G}(p) = p^-1$, where `Dι_{\mathcal G}(p): \mathfrak g → \mathfrak g. This can be done in-place ofY`.

Base.log — Function

log(G::LieGroup, g, h)
log!(G::LieGroup, X, g, h)

Compute the Lie group logarithmic map

\[\log_g h = ∘\log_{\mathcal G}(g^{-1}∘h)\]

where $\log_{\mathcal G}$ denotes the Lie group logarithmic function The computation can be performed in-place of X.

Note

If g is the Identity the Lie group logarithmic function is computed directly. Implementing the Lie group logarithmic function introduces a default implementation for this function.

Note

The Lie group logarithmic map is usually different from the logarithmic map with respect to the metric of the underlying Riemannian manifold $\mathcal M$. To access the (Riemannian) logarithmic map, use log(base_manifoldG, g, h).

log(G::LieGroup, e::Identity, g)
log!(G::LieGroup, X, e::Identity, g)

Compute the (Lie group) logarithmic function $\log_{\mathcal G}: \mathcal G → \mathfrak g$, which is the inverse of the Lie group exponential function The compuation can be performed in-place of X.

ManifoldsBase.log! — Function

log(G::LieGroup, g, h)
log!(G::LieGroup, X, g, h)

Compute the Lie group logarithmic map

\[\log_g h = ∘\log_{\mathcal G}(g^{-1}∘h)\]

where $\log_{\mathcal G}$ denotes the Lie group logarithmic function The computation can be performed in-place of X.

Note

If g is the Identity the Lie group logarithmic function is computed directly. Implementing the Lie group logarithmic function introduces a default implementation for this function.

Note

The Lie group logarithmic map is usually different from the logarithmic map with respect to the metric of the underlying Riemannian manifold $\mathcal M$. To access the (Riemannian) logarithmic map, use log(base_manifoldG, g, h).

log(G::LieGroup, e::Identity, g)
log!(G::LieGroup, X, e::Identity, g)

Compute the (Lie group) logarithmic function $\log_{\mathcal G}: \mathcal G → \mathfrak g$, which is the inverse of the Lie group exponential function The compuation can be performed in-place of X.

Actions on Lie groups

LieGroups.AbstractGroupAction — Type

AbstractGroupAction{T<:AbstractGroupActionType, L<:LieGroup, M<:AbstractManifold}

An abstract group action of AbstractGroupActionType T of a LieGroup of type L acting on an AbstractManifold of type M.

See [HN12, Section 9.1.3] for more details.

LieGroups.AbstractGroupActionType — Type

AbstractGroupActionType

An abstract supertype for group actions.

LieGroups.AbstractLeftGroupActionType — Type

AbstractLeftGroupActionType <: AbstractGroupActionType

A type representing a (smooth) group action $σ: \mathcal G × \mathcal M → \mathcal M$ of a LieGroup $\mathcal G$ acting (from the left) on an AbstractManifold $\mathcal M$. with the following properties

$σ(\mathrm e, p) = p$ holds for all $p ∈ \mathcal M$

$σ(g, σ(h, p)) = σ(g∘h, p)$ holds for all $g,h ∈ \mathcal G$, $p ∈ \mathcal M$

where $∘$ denotes the group operation of the LieGroup $\mathcal G$. See also [HN12, Definition 9.1.11].

The type of action can be seen a bit better when writing the action as a family $σ_g(p)$: we obtain from the second property as

\[ σ_g(σ_h(p)) = σ_{gh}(p)\]

and see that $g$ appears on the left.

To emphasize the side the group operation is acting from, we sometimes write $σ^{\mathrm L}$. If the action is clear from context we write $σ(g, p) = g ⋅ p$.

One notable example of a left action is the inverse of an action of AbstractRightGroupActionType $σ^{\mathrm R}$, which is given by $τ_g = (σ^{\mathrm R}_g)^{-1} = σ^{\mathrm R}_{g^{-1}}$. We obtain

\[τ_g(τ_h(p)) = σ^{\mathrm R}_{g^{-1}}(σ^{\mathrm R}_{h^{-1}}(p)) = σ^{\mathrm R}_{h^{-1}g^{-1}}(p) = σ^{\mathrm R}_{(gh)^{-1}}(p) τ_{gh}(p).\]

Note

In function definitions we follow the idea of the family of actions and use the order (M, g, p) in function signatures.

LieGroups.AbstractRightGroupActionType — Type

AbstractRightGroupActionType <: AbstractGroupActionType

A type representing a (smooth) group action $σ: \mathcal M × \mathcal G → \mathcal M$ of a LieGroup $\mathcal G$ acting (from the right) on an AbstractManifold $\mathcal M$. with the following properties

$σ(p, \mathrm e) = p$ holds for all $p ∈ \mathcal M$

$σ(σ(p, g), h) = σ(p, g∘h)$ holds for all $g,h ∈ \mathcal G$, $p ∈ \mathcal M$

where $∘$ denotes the group operation of the LieGroup $\mathcal G$. See also [HN12, Remark 9.1.12].

The type of action can be seen a bit better when writing the action as a family $σ_g(p)$: we obtain from the second property as

\[ σ_g(σ_h(p)) = σ_{hg}(p)\]

and see that $g$ appears on the right.

To emphasize the side the group operation is acting from, we sometimes write $σ^{\mathrm R}$. If the action is clear from context we write $σ(p, g) = p ⋅ g$.

One notable example of a right action is the inverse of an action of AbstractLeftGroupActionType $σ^{\mathrm L}$, which is given by $τ_g = (σ^{\mathrm L}_g)^{-1} = σ^{\mathrm L}_{g^{-1}}$. We obtain

\[τ_g(τ_h(p)) = σ^{\mathrm L}_{g^{-1}}(σ^{\mathrm L}_{h^{-1}}(p)) = σ^{\mathrm L}_{g^{-1}h^{-1}}(p) = σ^{\mathrm L}_{(hg)^{-1}}(p) τ_{hg}(p).\]

Note

In function definitions we follow the idea of the family of actions and use the order (M, g, p) in function signatures.

Functions for Lie group actions

Base.inv — Method

 inv(T::AbstractGroupActionType)

Return the type representing the inverse of an AbstractGroupActionType.

Base.inv — Method

 inv(A::AbstractGroupAction{T})

Return the tuple representing the inverse of an AbstractGroupAction of AbstractGroupActionType T. This is usually done by returning the group action with the inverse type of T.

LieGroups.base_lie_group — Method

base_Lie_group(A::AbstractGroupAction)

Return the LieGroup of the AbstractGroupAction.

LieGroups.diff_apply! — Method

diff_apply(A::AbstractGroupAction{T, L, M}, g, p, X)
diff_apply!(A::AbstractGroupAction{T, L, M}, Y, g, p, X)

Compute the differential $D_p σ_g(p): T_p\mathcal M → T_{σ_g(p)}\mathcal M$, where for a left group action we have $σ_g(p) = σ(g,p)$, for a right action $σ_g(p) = σ(p, g)$.

LieGroups.diff_apply — Method

diff_apply(A::AbstractGroupAction{T, L, M}, g, p, X)
diff_apply!(A::AbstractGroupAction{T, L, M}, Y, g, p, X)

Compute the differential $D_p σ_g(p): T_p\mathcal M → T_{σ_g(p)}\mathcal M$, where for a left group action we have $σ_g(p) = σ(g,p)$, for a right action $σ_g(p) = σ(p, g)$.

LieGroups.diff_group_apply! — Method

diff_group_apply(A::AbstractGroupAction{T, L, M}, g, p, X)
diff_group_apply!(A::AbstractGroupAction{T, L, M}, Y, g, p, X)

Compute the differential $D_g σ_p(g): T_g\mathcal G → T_{σ_p(g)}\mathcal M$, where we use the short hand notation $σ_p(g) = σ(g,p)$ for a left action, and for a right action $σ_p(g) = σ(p, g)$.

LieGroups.diff_group_apply — Method

diff_group_apply(A::AbstractGroupAction{T, L, M}, g, p, X)
diff_group_apply!(A::AbstractGroupAction{T, L, M}, Y, g, p, X)

Compute the differential $D_g σ_p(g): T_g\mathcal G → T_{σ_p(g)}\mathcal M$, where we use the short hand notation $σ_p(g) = σ(g,p)$ for a left action, and for a right action $σ_p(g) = σ(p, g)$.

LieGroups.switch — Method

switch(T::AbstractGroupActionType)

Return the object representing an AbstractGroupActionType related to a group operation action that switched the side, that is it turns a left action type into its corresponding right action type.

LieGroups.switch — Method

 switch(A::AbstractGroupAction{T})

Return the group operation action representing the similar AbstractGroupAction of AbstractGroupActionType T but acting from the other side. It switches left to right and vice versa. This is done by returning the group action with the “switched” type of T.

Manifolds.apply! — Method

apply(A::AbstractGroupAction{T, L, M}, g, p)
apply!(A::AbstractGroupAction{T, L, M}, q, g, p)

Apply the group action induced by $g ∈ \mathcal G$ to $p ∈ \mathcal M$, where the kind of group action is indicated by the AbstractGroupActionType T. This can be perfomed in-place of q.

Manifolds.apply — Method

apply(A::AbstractGroupAction{T, L, M}, g, p)
apply!(A::AbstractGroupAction{T, L, M}, q, g, p)

Apply the group action induced by $g ∈ \mathcal G$ to $p ∈ \mathcal M$, where the kind of group action is indicated by the AbstractGroupActionType T. This can be perfomed in-place of q.

ManifoldsBase.base_manifold — Method

base_manifold(A::AbstractGroupAction)

Return the AbstractManifold of the AbstractGroupAction.

Specific Lie group actions

LieGroups.GroupOperationAction — Type

GroupOperationAction{T<:AbstractLeftGroupActionType,G<:LieGroup} <: AbstractGroupAction{T,G,G}

The AbstractGroupAction.

LieGroups.InverseLeftGroupOperation — Type

InverseLeftGroupOperation <: AbstractRightGroupActionType

A type for the AbstractLeftGroupActionType when acting on the group itself given by the inverse of a LeftGroupOperation $σ_h$ as

\[τ_h(g) \coloneqq σ_h^{-1}(g) = σ(h^{-1},g) = h^{-1}∘g\]

Note that while in naming it is the inverse of the left action, it's properties yield that is is an AbstractRightGroupActionType, since

\[τ_g(τ_h(p)) = σ^{\mathrm L}_{g^{-1}}(σ^{\mathrm L}_{h^{-1}}(p)) = σ^{\mathrm L}_{g^{-1}h^{-1}}(p) = σ^{\mathrm L}_{(hg)^{-1}}(p) τ_{hg}(p).\]

for its inverse $(σ_h)^{-1}$ see InverseLeftGroupOperation.

Note

Some literature also calls this by itself the right group operation action.

LieGroups.InverseRightGroupOperation — Type

InverseRightGroupOperation <: AbstractLeftGroupActionType

A type for the AbstractLeftGroupActionType when acting on the group itself given by the inverse of a RightGroupOperation $σ_h$ as

\[τ_h(g) \coloneqq σ_h^{-1}(g) = σ(h^{-1},g) = g∘h^{-1}\]

Note that while in naming it is the inverse of the right action, it's properties yield that is is an AbstractLeftGroupActionType, since

\[τ_g(τ_h(p)) = σ^{\mathrm R}_{g^{-1}}(σ^{\mathrm R}_{h^{-1}}(p)) = σ^{\mathrm R}_{h^{-1}g^{-1}}(p) = σ^{\mathrm R}_{(gh)^{-1}}(p) τ_{gh}(p).\]

for its inverse $(σ_h)^{-1}$ see InverseLeftGroupOperation.

LieGroups.LeftGroupOperation — Type

LeftGroupOperation <: AbstractLeftGroupActionType

A type for the AbstractLeftGroupActionType when acting on the group itself from the left, that is

\[σ_h(g) = σ(h,g) = h∘g\]

for its inverse $(σ_h)^{-1}$ see InverseLeftGroupOperation.

LieGroups.RightGroupOperation — Type

RightGroupOperation <: AbstractRightGroupActionType

A type for the AbstractLeftGroupActionType when acting on the group itself gfrom the right

\[σ_h(g) = σ(h,g) = g∘h\]

for its inverse $(σ_h)^{-1}$ see InverseRightGroupOperation.

Base.inv — Method

inv(::GroupOperationAction)

return the inverse group operation action, that is, use the type representing the inverse operation.

Base.inv — Method

inv(::InverseLeftGroupOperation)

Return the inverse of the InverseLeftGroupOperation, that is the LeftGroupOperation.

Base.inv — Method

inv(::InverseRightGroupOperation)

Return the inverse of the InverseRightGroupOperation, that is the RightGroupOperation.

Base.inv — Method

inv(::LeftGroupOperation)

Return the inverse of the LeftGroupOperation, that is the InverseLeftGroupOperation.

Base.inv — Method

inv(::RightGroupOperation)

Return the inverse of the RightGroupOperation, that is the InverseRightGroupOperation.

LieGroups.diff_apply! — Method

Compute the differential $D_g σ_p(g): T_g\mathcal G → T_{σ_p(g)}\mathcal M$ of a group operation action, that is

for the LeftGroupOperation this calls diff_left_compose(G, g, h, X)

for the RightGroupOperation this calls diff_right_compose(G, g, h, X)

for the InverseLeftGroupOperation this calls diff_left_compose with $g^{-1}$

for the InverseRightGroupOperation this calls diff_right_compose with $g^{-1}$

LieGroups.diff_apply — Method

Compute the differential $D_g σ_p(g): T_g\mathcal G → T_{σ_p(g)}\mathcal M$ of a group operation action, that is

for the LeftGroupOperation this calls diff_left_compose(G, g, h, X)

for the RightGroupOperation this calls diff_right_compose(G, g, h, X)

for the InverseLeftGroupOperation this calls diff_left_compose with $g^{-1}$

for the InverseRightGroupOperation this calls diff_right_compose with $g^{-1}$

LieGroups.switch — Method

switch(::InverseLeftGroupOperation)

Return the InverseRightGroupOperation, that is, turns $σ_g = g^{-1}∘h$ into $τ_g(h) = h∘g^{-1}$

LieGroups.switch — Method

switch(::InverseRightGroupOperation)

Return the InverseLeftGroupOperation, that is, turns $σ_g = h∘g^{-1}$ into $τ_g(h) = g^{-1}∘h$

LieGroups.switch — Method

switch(::LeftGroupOperation)

Return the RightGroupOperation, that is, turns $σ_g = g∘h$ into $τ_g(h) = h∘g$

LieGroups.switch — Method

switch(::RightGroupOperation)

Return the LeftGroupOperation, that is, turns $σ_g = h∘g$ into $τ_g(h) = g∘h$

Manifolds.apply! — Method

apply(A::GroupOperationAction, g, h)
apply!(A::GroupOperationAction, k, g, h)

apply the stored group operation action, using compose $∘$. this can be done in-place of k.

Manifolds.apply — Method

apply(A::GroupOperationAction, g, h)
apply!(A::GroupOperationAction, k, g, h)

apply the stored group operation action, using compose $∘$. this can be done in-place of k.

Functions on Lie Algebras

LieGroups.lie_bracket — Function

lie_bracket!(𝔤::LieAlgebra, X, Y)
lie_bracket!(𝔤::LieAlgebra, Z, X, Y)

Compute the Lie bracket $[⋅,⋅]: \mathfrak g×\mathfrak g → \mathfrak g$ which fulfills

$[X,X] = 0$ for all $X ∈ \mathfrak g$

The Jacobi identity holds $[X, [Y,Z]] = [[X,Y],Z] = [Y, [X,Z]]$ holds for all $X, Y, Z ∈ \mathfrak g$.

The computation can be done in-place of Z.

LieGroups.lie_bracket! — Function

lie_bracket!(𝔤::LieAlgebra, X, Y)
lie_bracket!(𝔤::LieAlgebra, Z, X, Y)

Compute the Lie bracket $[⋅,⋅]: \mathfrak g×\mathfrak g → \mathfrak g$ which fulfills

$[X,X] = 0$ for all $X ∈ \mathfrak g$

The Jacobi identity holds $[X, [Y,Z]] = [[X,Y],Z] = [Y, [X,Z]]$ holds for all $X, Y, Z ∈ \mathfrak g$.

The computation can be done in-place of Z.

Specific Group Operations

For some generic group operations, an easy generic implementation can be provided. This section lists these specific cases and what is implemented for these.

Additive Group operation

LieGroups.AdditionGroupOperation — Type

AdditionGroupOperation <: AbstractGroupOperation

A group operation that is realised introducing defaults that fall back to + and - being overloaded, for example compose(G::LieGroup{𝔽,M,AdditionGroupOperation}, a,b) = a+b

Multiplicative Group operation

Literature

[HN12]

J. Hilgert and K.-H. Neeb. Structure and Geometry of Lie Groups (Springer Monographs in Mathematics, 2012).