An Interface for Lie group actions

LieGroups.AbstractActionActsOnType — Type

AbstractActionActsOnType

An abstract type representing what an action acts on, Most notably these are the ActionActsOnLeft and ActionActsOnRight, see their documentations for more details.

For its practical use see the SemidirectProductGroupOperation.

LieGroups.AbstractGroupActionType — Type

AbstractGroupActionType

An abstract supertype for group action types, which are used within a GroupAction.

LieGroups.AbstractLeftGroupActionType — Type

AbstractLeftGroupActionType <: AbstractGroupActionType

A type representing a (smooth) group action $σ: \mathcal G × \mathcal M → \mathcal M$ of a AbstractLieGroup $\mathcal G$ acting (from the left) on an AbstractManifold $\mathcal M$. with the following properties

$σ(\mathrm{e}, p) = p$ holds for all $p ∈ \mathcal M$

$σ(g, σ(h, p)) = σ(g∘h, p)$ holds for all $g,h ∈ \mathcal G$, $p ∈ \mathcal M$

where $∘$ denotes the group operation of the AbstractLieGroup $\mathcal G$. See also [HN12, Definition 9.1.11].

The type of action can be seen a bit better when writing the action as a family $σ_g(p)$: we obtain from the second property as

\[ σ_g(σ_h(p)) = σ_{gh}(p)\]

and see that $g$ appears on the left.

When writing about general group actions, the symbol $α$ is often used. The order of arguments then follows the same as the one of the left action. Most often we use the index notation $α_g(p)$. The definition of functions also follows this notation, i.e. we use e.g. apply(A, g, p)

One notable example of a left action is the inverse of an action of AbstractRightGroupActionType $τ$, which is given by $σ_g = (τ_g)^{-1} = τ_{g^{-1}}$. We obtain

\[σ_g(σ_h(p)) = τ_{g^{-1}}(τ_{h^{-1}}(p)) = τ_{h^{-1}g^{-1}}(p) = τ_{(gh)^{-1}}(p) σ_{gh}(p).\]

LieGroups.AbstractRightGroupActionType — Type

AbstractRightGroupActionType <: AbstractGroupActionType

A type representing a (smooth) group action $τ: \mathcal M × \mathcal G → \mathcal M$ of a AbstractLieGroup $\mathcal G$ acting (from the right) on an AbstractManifold $\mathcal M$. with the following properties

$τ(p, \mathrm{e}) = p$ holds for all $p ∈ \mathcal M$

$τ(τ(p, g), h) = τ(τ(p, g), h)$ holds for all $g,h ∈ \mathcal G$, $p ∈ \mathcal M$

where $∘$ denotes the group operation of the AbstractLieGroup $\mathcal G$. See also [HN12, Remark 9.1.12].

The type of action can be seen a bit better when writing the action as a family $τ_g(p)$: we obtain from the second property as

\[ τ_g(τ_h(p)) = τ_{hg}(p)\]

and see that $g$ appears on the right.

When writing about general group actions, the symbol $α$ is often used. In that case the order of arguments follows either the one from the left action, but most often we use the index notation. The definition of functions also follows this notation, i.e. we use e.g. apply(A, g, p)

One notable example of a right action is the inverse of an action of AbstractLeftGroupActionType $σ$, which is given by $τ_g = (σ_g)^{-1} = σ_{g^{-1}}$. We obtain

\[τ_g(τ_h(p)) = σ_{g^{-1}}(σ_{h^{-1}}(p)) = σ_{g^{-1}h^{-1}}(p) = σ_{(hg)^{-1}}(p) τ_{hg}(p).\]

LieGroups.ActionActsOnLeft — Type

ActionActsOnLeft <: AbstractActionActsOnType

An AbstractActionActsOnType representing that an action acts on the left.

This is meant in the following way: Given a GroupAction $α: \mathcal G × \mathcal H → \mathcal H$ where a Lie group $\mathcal G$ acts on another Lie group $\mathcal H$ with an arbitrary action.

Then, e.g. within the definition of the LeftSemidirectProductGroupOperation or RightSemidirectProductGroupOperation, we have two choices where the group action $α$ acts on, namely: Let $g ∈ \mathcal G$ and $h_1,h_2 ∈ \mathcal H$ be given, then this type represents the variant

\[α_g(h_1) ⋅ h_2,\]

where $⋅$ denotes the group operation on $\mathcal H$. The Left in the name of this type refers to the fact that the action is applied to the left element $h_1$.

Note that this is independent of both the type of action (left or right) and whether the semidirect product is a left or a right semidirect one.

For its practical use see the SemidirectProductGroupOperation.

LieGroups.ActionActsOnRight — Type

ActionActsOnRight <: AbstractActionActsOnType

An AbstractActionActsOnType representing that an action acts on the right.

This is meant in the following way: Given a GroupAction $α: \mathcal G × \mathcal H → \mathcal H$ where a Lie group $\mathcal G$ acts on another Lie group $\mathcal H$ with an arbitrary action.

Then, e.g. within the definition of the LeftSemidirectProductGroupOperation or RightSemidirectProductGroupOperation, we have two choices where the group action $α$ acts on, namely: Let $g ∈ \mathcal G$ and $h_1,h_2 ∈ \mathcal H$ be given, then this type represents the variant

\[h_1 ⋅ α_g(h_2),\]

where $⋅$ denotes the group operation on $\mathcal H$. The Right in the name of this type refers to the fact that the action is applied to the right element $h_2$.

Note that this is independent of both the type of action (left or right) and whether the semidirect product is a left or a right semidirect one.

For its practical use see the SemidirectProductGroupOperation.

LieGroups.GroupAction — Type

GroupAction{T<:GroupActionType, L<:LieGroup, M<:AbstractManifold}

Specify a group action of AbstractGroupActionType T of a AbstractLieGroup G acting on an AbstractManifold M.

Let $\mathcal M$ be a AbstractManifold and $\mathcal G$ be a AbstractLieGroup with group operation $∘$.

A (smooth) action of the group $\mathcal G$ on the manifold $\mathcal M$ is a map

\[α: \mathcal G × \mathcal M → \mathcal M\]

with the properties

Identity. $α(\mathrm{e}, p) = p$ holds for all $p ∈ \mathcal M$

Compatibility. If $α$ is a $AbstractLeftGroupActionType$ we usually denote it by $σ$ and the compatibility reads

\[σ_g(σ_h(p)) = σ_{g∘h}(p) \text{ holds for all} g,h ∈ \mathcal G \text{ and} p ∈ \mathcal M\]

If $α$ is a $AbstractRightGroupActionType$ we usually denote it by $τ$ and the compatibility reads

\[τ_g(τ_h(p)) = τ_{h∘g}(p)`` holds for all ``g,h ∈ \mathcal G \text{ for all} p ∈ \mathcal M\]

intuitively the left/right property of an action specifies on which side the “outer” group actions element $g$ gets “appended” in the composition.

Fields

type::AbstractGroupActionType: The type of the group action.

group::AbstractLieGroup: The group acting.

manifold::AbstractManifold: The manifold the group acts upon.

See [HN12, Section 9.1.3] for more details.

Constructors

GroupAction(
    group::AbstractLieGroup, manifold::ManifoldsBase.AbstractManifold, type::AbstractGroupActionType
)

Generate a group action where the type of the action and what it acts on are keyword arguments. They default to the most common choice, that the ActionActsOnRight.

Functions for Lie group actions

Base.inv — Method

inv(::AbstractGroupActionType)

return the inverse group operation action, that is, use the type representing the inverse operation.

Base.inv — Method

 inv(A::GroupAction{T})

Return the GroupAction representing the inverse of an GroupAction of AbstractGroupActionType T. This is usually done by returning the group action with the inverse type of T.

LieGroups.apply! — Method

apply(A::GroupAction{T}, g, p)
apply!(A::GroupAction{T}, q, g, p)

Apply the group action induced by $g ∈ \mathcal G$ to $p ∈ \mathcal M$, where the kind of group action is indicated by the AbstractGroupActionType T. This can be performed in-place of q.

LieGroups.apply — Method

apply(A::GroupAction{T}, g, p)
apply!(A::GroupAction{T}, q, g, p)

Apply the group action induced by $g ∈ \mathcal G$ to $p ∈ \mathcal M$, where the kind of group action is indicated by the AbstractGroupActionType T. This can be performed in-place of q.

LieGroups.base_lie_group — Method

base_lie_group(A::GroupAction)

Return the AbstractLieGroup of the GroupAction specifying the action.

LieGroups.default_left_action — Method

default_left_action(G::AbstractLieGroup, M::AbstractManifold)

Return the default left action for a Lie group G acting on a manifold M.

LieGroups.default_right_action — Method

default_right_action(G::AbstractLieGroup, M::AbstractManifold)

Return the default right action for a Lie group G acting on a manifold M.

LieGroups.diff_apply! — Method

diff_apply(A::GroupAction, g, p, X)
diff_apply!(A::GroupAction, Y, g, p, X)

Compute the differential $\mathrm{D}_p α_g(p): T_p\mathcal M → T_{σ_g(p)}\mathcal M$.

LieGroups.diff_apply — Method

diff_apply(A::GroupAction, g, p, X)
diff_apply!(A::GroupAction, Y, g, p, X)

Compute the differential $\mathrm{D}_p α_g(p): T_p\mathcal M → T_{σ_g(p)}\mathcal M$.

LieGroups.diff_group_apply! — Method

diff_group_apply(A::GroupAction, g, p, X)
diff_group_apply!(A::GroupAction, Y, g, p, X)

Compute the differential $\mathrm{d}_{\mathcal G} σ_g(p): \mathfrak g → \mathfrak g$, where we use the short hand notation $σ_p(g) = σ(g,p)$ for a left action, and for a right action $τ_p(g) = τ(g,p)$.

LieGroups.diff_group_apply — Method

diff_group_apply(A::GroupAction, g, p, X)
diff_group_apply!(A::GroupAction, Y, g, p, X)

Compute the differential $\mathrm{d}_{\mathcal G} σ_g(p): \mathfrak g → \mathfrak g$, where we use the short hand notation $σ_p(g) = σ(g,p)$ for a left action, and for a right action $τ_p(g) = τ(g,p)$.

LieGroups.switch — Method

switch(T::AbstractGroupActionType)

Return the object representing an AbstractGroupActionType related to a group operation action that switched the side, that is it turns a left action type into its corresponding right action type.

LieGroups.switch — Method

 switch(A::GroupAction{T})

Return the group operation action representing the similar GroupAction of AbstractGroupActionType T but acting from the other side. It switches left to right and vice versa. This is done by returning the group action with the “switched” type of T.

ManifoldsBase.base_manifold — Method

base_manifold(A::GroupAction)

Return the AbstractManifold the group action acts upon.

Generic Lie group actions

LieGroups.AdditionGroupAction — Type

AdditionGroupAction <: AbstractLeftGroupActionType

Specify that in a GroupAction with Lie group $\mathcal G$ and manifold $\mathcal M$, where the group action is given by addition

Given an element $g ∈ \mathcal G$ and a point $p ∈ \mathcal M$, the family of actions $σ_g: \mathcal M → \mathcal M$ is given by

\[σ_g(p) = g + p\]

where $+$ denotes the vector or matrix addition.

LieGroups.InverseLeftGroupOperationAction — Type

InverseLeftGroupOperationAction <: AbstractRightGroupActionType

A type for the AbstractLeftGroupActionType when acting on the group itself given by the inverse of a LeftGroupOperationAction $σ_h$ as

\[τ_h(g) \coloneqq σ_h^{-1}(g) = σ(h^{-1},g) = h^{-1}∘g\]

Note that while in naming it is the inverse of the left action, it's properties yield that is is an AbstractRightGroupActionType, since

\[τ_g(τ_h(p)) = σ_{g^{-1}}(σ_{h^{-1}}(p)) = σ_{g^{-1}h^{-1}}(p) = σ_{(hg)^{-1}}(p) τ_{hg}(p).\]

for its inverse $(σ_h)^{-1}$ see InverseLeftGroupOperationAction.

Note

Some literature also calls this by itself the right group operation action.

LieGroups.InverseRightGroupOperationAction — Type

InverseRightGroupOperationAction <: AbstractLeftGroupActionType

A type for the AbstractLeftGroupActionType when acting on the group itself given by the inverse of a RightGroupOperationAction $σ_h$ as

\[τ_h(g) \coloneqq σ_h^{-1}(g) = σ(h^{-1},g) = g∘h^{-1}\]

Note that while in naming it is the inverse of the right action, it's properties yield that is is an AbstractLeftGroupActionType, since

\[σ_g(σ_h(p)) = τ_{g^{-1}}(τ_{h^{-1}}(p)) = τ_{h^{-1}g^{-1}}(p) = τ_{(gh)^{-1}}(p) σ_{gh}(p).\]

for its inverse $(σ_h)^{-1}$ see InverseLeftGroupOperationAction.

LieGroups.LeftGroupOperationAction — Type

LeftGroupOperationAction <: AbstractLeftGroupActionType

A type for the AbstractLeftGroupActionType when acting on the group itself from the left, that is

\[σ_h(g) = σ(h,g) = h∘g\]

for its inverse $(σ_h)^{-1}$ see InverseLeftGroupOperationAction.

LieGroups.RightGroupOperationAction — Type

RightGroupOperationAction <: AbstractRightGroupActionType

A type for the AbstractRightGroupActionType when acting on the group itself from the right.

\[σ_h(g) = σ(h,g) = g∘h\]

for its inverse $(σ_h)^{-1}$ see InverseRightGroupOperationAction.

Base.inv — Method

inv(::InverseLeftGroupOperationAction)

Return the inverse of the InverseLeftGroupOperationAction, that is the LeftGroupOperationAction.

Base.inv — Method

inv(::InverseRightGroupOperationAction)

Return the inverse of the InverseRightGroupOperationAction, that is the RightGroupOperationAction.

Base.inv — Method

inv(::LeftGroupOperationAction)

Return the inverse of the LeftGroupOperationAction, that is the InverseLeftGroupOperationAction.

Base.inv — Method

inv(::RightGroupOperationAction)

Return the inverse of the RightGroupOperationAction, that is the InverseRightGroupOperationAction.

LieGroups.GroupOperationAction — Function

GroupOperationAction(group::LieGroup, action=LeftGroupOperationAction(), on=ActionActsOnLeft())

Return a GroupAction for an AbstractGroupActionType action representing the group operation as an action of the group on itself.

LieGroups.switch — Method

switch(::InverseLeftGroupOperationAction)

Return the InverseRightGroupOperationAction, that is, turns $σ_g = g^{-1}∘h$ into $τ_g(h) = h∘g^{-1}$

LieGroups.switch — Method

switch(::InverseRightGroupOperationAction)

Return the InverseLeftGroupOperationAction, that is, turns $τ_g = h∘g^{-1}$ into $σ_g(h) = g^{-1}∘h$

LieGroups.switch — Method

switch(::LeftGroupOperationAction)

Return the RightGroupOperationAction, that is, turns $σ_g(h) = g∘h$ into $τ_g(h) = h∘g$

LieGroups.switch — Method

switch(::RightGroupOperationAction)

Return the LeftGroupOperationAction, that is, turns $τ_g(h) = h∘g$ into $τ_g = g∘h$

LieGroups.ColumnwiseGroupAction — Type

ColumnwiseGroupAction{A<:AbstractGroupActionType} <: AbstractGroupActionType

A type for an action that applies a group action column-wise on a manifold, where a column-wise interpretation makes sense, e.g. a matrix-manifold.

Fields

action::A: The group action to be applied column-wise.

Constructor

ColumnwiseGroupAction(action::AbstractGroupActionType)

LieGroups.LeftMultiplicationGroupAction — Type

LeftMultiplicationGroupAction <: AbstractLeftGroupActionType

Specify that in a GroupAction with Lie group $\mathcal G$ and manifold $\mathcal M$, where the group action is given by multiplication from the left:

Given an element $g ∈ \mathcal G$ and a point $p ∈ \mathcal M$, the family of actions $σ_g: \mathcal M → \mathcal M$ is given by

\[σ_g(p) = g*p\]

where $*$ denotes the matrix(-vector) multiplication.

LieGroups.RotationAroundAxisAction — Type

RotationAroundAxisAction(axis::AbstractVector)

Space of actions of the circle group CircleGroup on $ℝ^3$ around given axis.

LieGroups.apply — Method

apply(A::GroupAction{<:RotationAroundAxisAction}, θ, p)

Rotate point p from Euclidean manifold around axis A.axis by angle θ. The formula reads

\[p_{rot} = (\cos(θ))p + (k×p) \sin(θ) + k (k⋅p) (1-\cos(θ)),\]

where $k$ is the vector A.type.axis and ⋅ is the dot product.

LieGroups.RowwiseGroupAction — Type

RowwiseGroupAction{A<:AbstractGroupActionType} <: AbstractGroupActionType

A type for an action that applies a group action row-wise on a manifold, where a row-wise interpretation makes sense, e.g. a matrix-manifold.

Fields

action::A: The group action to be applied row-wise.

Constructor

RowwiseGroupAction(action::AbstractGroupActionType)

Literature

[HN12]

J. Hilgert and K.-H. Neeb. Structure and Geometry of Lie Groups (Springer Monographs in Mathematics, 2012).