An interface for Lie algebras

LieGroups.LieAlgebra — Type

LieAlgebra{𝔽, G} <: AbstractManifold{𝔽}

Represent the Lie algebra $\mathfrak g$, that is a $𝔽$ vector space with an associated lie_bracket $[⋅,⋅]: \mathfrak g×\mathfrak g → \mathfrak g$ which fulfills

$[X,X] = 0$ for all $X ∈ \mathfrak g$

The Jacobi identity $[X, [Y,Z]] = [[X,Y],Z] = [Y, [X,Z]]$ holds for all $X, Y, Z ∈ \mathfrak g$.

The Lie algebras considered here are those related to a AbstractLieGroup $\mathcal G$, namely the tangent space $T_{\mathrm{e}}\mathcal G$ at the Identity, this is internally just a const of the corresponding TangentSpace.

Convention for representing tangent vectors in the Lie algebra

A vector field $\mathcal X: \mathcal G → T\mathcal G$, $X(g) ∈ T_g\mathcal G$ is called a left-invariant vector field if it satisfies

\[\mathcal X(λ_g(h)) = Dλ_g(h)[\mathcal X(h)], \quad\text{for all}\quad g, h ∈ \mathcal G,\]

where $λ_g: \mathcal G → \mathcal G$ is the left multiplication by $g$. Hence $\mathcal X$ is determined already when $X ∈ \mathfrak g$ is given, since $\mathcal X(g) = Dλ_g(e)[X]$, cf [HN12, Definition 9.1.7].

Throughout LieGroups.jl, we use this left-invariant convention to store tangent vectors at points on a Lie group as elements of the corresponding Lie algebra.

Constructor

LieAlgebra(G::AbstractLieGroup)

Return the Lie Algebra belonging to the AbstractLieGroup G.

LieGroups.DefaultLieAlgebraOrthogonalBasis — Type

DefaultLieAlgebraOrthogonalBasis{𝔽} <: ManifoldsBase.AbstractOrthogonalBasis{𝔽,ManifoldsBase.TangentSpaceType}

Specify an orthogonal basis for a Lie algebra. This is used as the default within hat and vee.

If not specifically overwritten/implemented for a Lie group, the DefaultOrthogonalBasis at the identity_element on the `base_manifold acts as a fallback.

Note

In order to implement the corresponding get_coordinates and get_vector functions, define get_coordinates_lie(::AbstractLieAlgebra, X, B) and get_vector_lie(::AbstractLieAlgebra, X, B), resp.

Functions on Lie algebras

Base.rand — Method

rand(::LieGroup; vector_at=nothing, σ=1.0, kwargs...)
rand(::LieAlgebra; σ=1.0, kwargs...)
rand!(::AbstractLieGroup, gX; vector_at=nothing, kwargs...)
rand!(::LieAlgebra, X; σ=1.0, kwargs...)

Compute a random point or tangent vector on a Lie group.

For points this just means to generate a random point on the underlying manifold itself.

For tangent vectors, an element in the Lie Algebra is generated, see also rand(::LieAlgebra; kwargs...)

LieGroups.base_lie_group — Method

base_lie_group(𝔤::LieAlgebra)

Return the base_lie_group of the given LieAlgebra belongs to.

LieGroups.lie_bracket! — Method

lie_bracket(𝔤::LieAlgebra, X, Y)
lie_bracket!(𝔤::LieAlgebra, Z, X, Y)

Compute the Lie bracket $[⋅,⋅]: \mathfrak g×\mathfrak g → \mathfrak g$ which fulfills

$[X,X] = 0$ for all $X ∈ \mathfrak g$

The Jacobi identity $[X, [Y,Z]] = [[X,Y],Z] = [Y, [X,Z]]$ holds for all $X, Y, Z ∈ \mathfrak g$.

The computation can be done in-place of Z.

LieGroups.lie_bracket — Method

lie_bracket(𝔤::LieAlgebra, X, Y)
lie_bracket!(𝔤::LieAlgebra, Z, X, Y)

Compute the Lie bracket $[⋅,⋅]: \mathfrak g×\mathfrak g → \mathfrak g$ which fulfills

$[X,X] = 0$ for all $X ∈ \mathfrak g$

The Jacobi identity $[X, [Y,Z]] = [[X,Y],Z] = [Y, [X,Z]]$ holds for all $X, Y, Z ∈ \mathfrak g$.

The computation can be done in-place of Z.

ManifoldsBase.base_manifold — Method

base_manifold(𝔤::LieAlgebra)

Return the base_manifold the AbstractLieGroup of the given LieAlgebra is based on.

ManifoldsBase.get_coordinates — Function

get_coordinates(𝔤::LieAlgebra, X::T, B::AbstractBasis)
get_coordinates!(𝔤::LieAlgebra, c, X::T, B::AbstractBasis)

Return the vector of coordinates to the decomposition of X with respect to an AbstractBasis of the LieAlgebra 𝔤. The operation can be performed in-place of c.

By default this function requires that identity_element(G, T) is available and calls the corresponding get_coordinates function of the Riemannian manifold the Lie group is build on.

The inverse operation is get_vector.

See also vee.

ManifoldsBase.get_coordinates! — Function

get_coordinates(𝔤::LieAlgebra, X::T, B::AbstractBasis)
get_coordinates!(𝔤::LieAlgebra, c, X::T, B::AbstractBasis)

Return the vector of coordinates to the decomposition of X with respect to an AbstractBasis of the LieAlgebra 𝔤. The operation can be performed in-place of c.

By default this function requires that identity_element(G, T) is available and calls the corresponding get_coordinates function of the Riemannian manifold the Lie group is build on.

The inverse operation is get_vector.

See also vee.

ManifoldsBase.get_vector — Function

get_vector(G::AbstractLieGroup, c, B::AbstractBasis; kwargs...)
get_vector(𝔤::LieAlgebra, c, B::AbstractBasis; kwargs...)
get_vector!(G::AbstractLieGroup, X::T, c, B::AbstractBasis; kwargs...)
get_vector!(𝔤::LieAlgebra, X::T, c, B::AbstractBasis; kwargs...)

Return the vector corresponding to a set of coefficients in an AbstractBasis of the LieAlgebra 𝔤. Since all tangent vectors are assumed to be represented in the Lie algebra, both signatures are equivalent. The operation can be performed in-place of a tangent vector X of type ::T.

By default this function requires identity_element(G) and calls the corresponding get_vector function of the Riemannian manifold the Lie group is build on.

The inverse operation is get_coordinates.

Keyword arguments

tangent_vector_type specify the tangent vector type to use for the allocating variants.

See also hat

ManifoldsBase.get_vector! — Function

get_vector(G::AbstractLieGroup, c, B::AbstractBasis; kwargs...)
get_vector(𝔤::LieAlgebra, c, B::AbstractBasis; kwargs...)
get_vector!(G::AbstractLieGroup, X::T, c, B::AbstractBasis; kwargs...)
get_vector!(𝔤::LieAlgebra, X::T, c, B::AbstractBasis; kwargs...)

Return the vector corresponding to a set of coefficients in an AbstractBasis of the LieAlgebra 𝔤. Since all tangent vectors are assumed to be represented in the Lie algebra, both signatures are equivalent. The operation can be performed in-place of a tangent vector X of type ::T.

By default this function requires identity_element(G) and calls the corresponding get_vector function of the Riemannian manifold the Lie group is build on.

The inverse operation is get_coordinates.

Keyword arguments

tangent_vector_type specify the tangent vector type to use for the allocating variants.

See also hat

ManifoldsBase.hat! — Method

hat(G::LieAlgebra, c)
hat(G::LieAlgebra, c, T::Type)
hat!(G::LieAlgebra, X::T, c)

Compute the hat map $(⋅)^{\wedge}: \mathcal V → 𝔤$ that maps a vector of coordinates $\mathbf{c} ∈ \mathcal V$, to a tangent vector $X ∈ \mathfrak g$. The coefficients are given with respect to a specific basis to a tangent vector in the Lie algebra

\[X = \sum_{i∈\mathcal I} c_iB_i,\]

where $\{ B_i \}_{i∈\mathcal I}$ is a basis of the Lie algebra and $\mathcal I$ a corresponding index set, which is usually $\mathcal I=\{ 1,\ldots,n \}$. Then $\mathcal V = ℝ^n$.

For the allocating variant, you can specify the type T of the tangent vector to obtain, in case there are different representations. The first signature produces the default representation.

The computation can be performed in-place of X. The inverse of hat is vee. Technically, hat is a specific case of get_vector and is implemented using the DefaultLieAlgebraOrthogonalBasis.

ManifoldsBase.hat — Method

hat(G::LieAlgebra, c)
hat(G::LieAlgebra, c, T::Type)
hat!(G::LieAlgebra, X::T, c)

Compute the hat map $(⋅)^{\wedge}: \mathcal V → 𝔤$ that maps a vector of coordinates $\mathbf{c} ∈ \mathcal V$, to a tangent vector $X ∈ \mathfrak g$. The coefficients are given with respect to a specific basis to a tangent vector in the Lie algebra

\[X = \sum_{i∈\mathcal I} c_iB_i,\]

where $\{ B_i \}_{i∈\mathcal I}$ is a basis of the Lie algebra and $\mathcal I$ a corresponding index set, which is usually $\mathcal I=\{ 1,\ldots,n \}$. Then $\mathcal V = ℝ^n$.

For the allocating variant, you can specify the type T of the tangent vector to obtain, in case there are different representations. The first signature produces the default representation.

The computation can be performed in-place of X. The inverse of hat is vee. Technically, hat is a specific case of get_vector and is implemented using the DefaultLieAlgebraOrthogonalBasis.

ManifoldsBase.hat — Method

hat(G::LieAlgebra, c)
hat(G::LieAlgebra, c, T::Type)
hat!(G::LieAlgebra, X::T, c)

Compute the hat map $(⋅)^{\wedge}: \mathcal V → 𝔤$ that maps a vector of coordinates $\mathbf{c} ∈ \mathcal V$, to a tangent vector $X ∈ \mathfrak g$. The coefficients are given with respect to a specific basis to a tangent vector in the Lie algebra

\[X = \sum_{i∈\mathcal I} c_iB_i,\]

where $\{ B_i \}_{i∈\mathcal I}$ is a basis of the Lie algebra and $\mathcal I$ a corresponding index set, which is usually $\mathcal I=\{ 1,\ldots,n \}$. Then $\mathcal V = ℝ^n$.

For the allocating variant, you can specify the type T of the tangent vector to obtain, in case there are different representations. The first signature produces the default representation.

The computation can be performed in-place of X. The inverse of hat is vee. Technically, hat is a specific case of get_vector and is implemented using the DefaultLieAlgebraOrthogonalBasis.

ManifoldsBase.inner — Method

inner(𝔤::LieAlgebra, X, Y)

Compute the inner product $⟨⋅,⋅⟩: \mathfrak g×\mathfrak g → ℝ$ on the Lie algebra $𝔤$. By default this uses the inner product on the underlying manifold of the AbstractLieGroup of 𝔤 at the identity_element(G). Note that this method allocates an appropriate identity element.

ManifoldsBase.is_point — Method

is_point(𝔤::LieAlgebra, X; kwargs...)

Check whether X is a valid point on the Lie Algebra 𝔤. This falls back to checking whether X is a valid point on the tangent space at the identity_element(G) on the base_manifold(G) on the AbstractLieGroup of 𝔤

ManifoldsBase.vee! — Method

vee(𝔤::LieAlgebra, X)
vee!(𝔤::LieAlgebra, c, X)

Compute the vee map $(⋅){\vee}: \mathfrak g → \mathcal V$ that maps a tangent vector X from the LieAlgebra $\mathfrak g$ to its coordinates with respect to the DefaultLieAlgebraOrthogonalBasis basis in the Lie algebra

\[X = \sum_{i∈\mathcal I} c_iB_i,\]

where $\{ B_i \}_{i∈\mathcal I}$ is a basis of the Lie algebra and $\mathcal I$ a corresponding index set, which is usually $\mathcal I=\{ 1,\ldots,n \}$. Then $\mathcal V = ℝ^n$

The computation can be performed in-place of c. The inverse of vee is hat. Technically, vee is a specific case of get_coordinates and is implemented using the DefaultLieAlgebraOrthogonalBasis.

ManifoldsBase.vee — Method

vee(𝔤::LieAlgebra, X)
vee!(𝔤::LieAlgebra, c, X)

Compute the vee map $(⋅){\vee}: \mathfrak g → \mathcal V$ that maps a tangent vector X from the LieAlgebra $\mathfrak g$ to its coordinates with respect to the DefaultLieAlgebraOrthogonalBasis basis in the Lie algebra

\[X = \sum_{i∈\mathcal I} c_iB_i,\]

where $\{ B_i \}_{i∈\mathcal I}$ is a basis of the Lie algebra and $\mathcal I$ a corresponding index set, which is usually $\mathcal I=\{ 1,\ldots,n \}$. Then $\mathcal V = ℝ^n$

The computation can be performed in-place of c. The inverse of vee is hat. Technically, vee is a specific case of get_coordinates and is implemented using the DefaultLieAlgebraOrthogonalBasis.

ManifoldsBase.zero_vector — Method

zero_vector(𝔤::LieAlgebra)
zero_vector(𝔤::LieAlgebra, T::Type)
zero_vector!(𝔤::LieAlgebra, X::T)

Generate a zero_vector of type T in the LieAlgebra $𝔤$ of the AbstractLieGroup G. By default this calls zero_vector on the manifold of G at the identity_element(G,T)

For the allocating variant the type T of the zero vector can be specified.

Random.rand! — Method

rand(::LieGroup; vector_at=nothing, σ=1.0, kwargs...)
rand(::LieAlgebra; σ=1.0, kwargs...)
rand!(::AbstractLieGroup, gX; vector_at=nothing, kwargs...)
rand!(::LieAlgebra, X; σ=1.0, kwargs...)

Compute a random point or tangent vector on a Lie group.

For points this just means to generate a random point on the underlying manifold itself.

For tangent vectors, an element in the Lie Algebra is generated, see also rand(::LieAlgebra; kwargs...)

Internal functions and macros

LieGroups.@default_lie_algebra_fallbacks — Macro

default_lie_algebra_fallbacks(TG, TF, Op, TV, Xfield::Symbol)

Introduce default fallbacks for all basic functions on Lie algebras, for Lie group of type TG with number system TF, an group operation Op, tangent vectors of type TV, with forwarding to fields Xfield and tangent vector functions

Literature

[HN12]

J. Hilgert and K.-H. Neeb. Structure and Geometry of Lie Groups (Springer Monographs in Mathematics, 2012).

rand(::LieGroup; vector_at=nothing, σ=1.0, kwargs...) rand(::LieAlgebra; σ=1.0, kwargs...) rand!(::AbstractLieGroup, gX; vector_at=nothing, kwargs...) rand!(::LieAlgebra, X; σ=1.0, kwargs...)
Compute a random point or tangent vector on a Lie group.

LieAlgebra{𝔽, G} <: AbstractManifold{𝔽}
Represent the Lie algebra , that is a vector space with an associated lie_bracket which fulfills

base_lie_group(𝔤::LieAlgebra)
Return the base_lie_group of the given LieAlgebra belongs to.

lie_bracket(𝔤::LieAlgebra, X, Y) lie_bracket!(𝔤::LieAlgebra, Z, X, Y)
Compute the Lie bracket which fulfills

lie_bracket(𝔤::LieAlgebra, X, Y) lie_bracket!(𝔤::LieAlgebra, Z, X, Y)
Compute the Lie bracket which fulfills

base_manifold(𝔤::LieAlgebra)
Return the base_manifold the AbstractLieGroup of the given LieAlgebra is based on.

get_coordinates(𝔤::LieAlgebra, X::T, B::AbstractBasis) get_coordinates!(𝔤::LieAlgebra, c, X::T, B::AbstractBasis)
Return the vector of coordinates to the decomposition of X with respect to an AbstractBasis of the LieAlgebra 𝔤.

get_coordinates(𝔤::LieAlgebra, X::T, B::AbstractBasis) get_coordinates!(𝔤::LieAlgebra, c, X::T, B::AbstractBasis)
Return the vector of coordinates to the decomposition of X with respect to an AbstractBasis of the LieAlgebra 𝔤.

get_vector(G::AbstractLieGroup, c, B::AbstractBasis; kwargs...) get_vector(𝔤::LieAlgebra, c, B::AbstractBasis; kwargs...) get_vector!(G::AbstractLieGroup, X::T, c, B::AbstractBasis; kwargs...) get_vector!(𝔤::LieAlgebra, X::T, c, B::AbstractBasis; kwargs...)
Return the vector corresponding to a set of coefficients in an AbstractBasis of the LieAlgebra 𝔤.

get_vector(G::AbstractLieGroup, c, B::AbstractBasis; kwargs...) get_vector(𝔤::LieAlgebra, c, B::AbstractBasis; kwargs...) get_vector!(G::AbstractLieGroup, X::T, c, B::AbstractBasis; kwargs...) get_vector!(𝔤::LieAlgebra, X::T, c, B::AbstractBasis; kwargs...)
Return the vector corresponding to a set of coefficients in an AbstractBasis of the LieAlgebra 𝔤.

hat(G::LieAlgebra, c) hat(G::LieAlgebra, c, T::Type) hat!(G::LieAlgebra, X::T, c)
Compute the hat map that maps a vector of coordinates , to a tangent vector .

hat(G::LieAlgebra, c) hat(G::LieAlgebra, c, T::Type) hat!(G::LieAlgebra, X::T, c)
Compute the hat map that maps a vector of coordinates , to a tangent vector .

hat(G::LieAlgebra, c) hat(G::LieAlgebra, c, T::Type) hat!(G::LieAlgebra, X::T, c)
Compute the hat map that maps a vector of coordinates , to a tangent vector .

inner(𝔤::LieAlgebra, X, Y)
Compute the inner product on the Lie algebra .

is_point(𝔤::LieAlgebra, X; kwargs...)
Check whether X is a valid point on the Lie Algebra 𝔤.

vee(𝔤::LieAlgebra, X) vee!(𝔤::LieAlgebra, c, X)
Compute the vee map that maps a tangent vector X from the LieAlgebra to its coordinates with respect to the DefaultLieAlgebraOrthogonalBasis basis in the Lie algebra

vee(𝔤::LieAlgebra, X) vee!(𝔤::LieAlgebra, c, X)
Compute the vee map that maps a tangent vector X from the LieAlgebra to its coordinates with respect to the DefaultLieAlgebraOrthogonalBasis basis in the Lie algebra

zero_vector(𝔤::LieAlgebra) zero_vector(𝔤::LieAlgebra, T::Type) zero_vector!(𝔤::LieAlgebra, X::T)
Generate a zero_vector of type T in the LieAlgebra of the AbstractLieGroup G.

rand(::LieGroup; vector_at=nothing, σ=1.0, kwargs...) rand(::LieAlgebra; σ=1.0, kwargs...) rand!(::AbstractLieGroup, gX; vector_at=nothing, kwargs...) rand!(::LieAlgebra, X; σ=1.0, kwargs...)
Compute a random point or tangent vector on a Lie group.

get_coordinates(𝔤::OrthogonalLieAlgebra, X, ::DefaultLieAlgebraOrthogonalBasis) get_coordinates(G::SpecialOrthogonalLieAlgebra, X, ::DefaultLieAlgebraOrthogonalBasis) get_coordinates!(G::OrthogonalLieAlgebra, c, X ::DefaultLieAlgebraOrthogonalBasis) get_coordinates!(G::SpecialOrthogonalLieAlgebra, c, X ::DefaultLieAlgebraOrthogonalBasis)
Compute the vector of coordinates from the Lie algebra tangent vector of the OrthogonalGroup O(n) in the DefaultLieAlgebraOrthogonalBasis, where is the dimension of the Lie algebra.

get_coordinates(𝔤::OrthogonalLieAlgebra, X, ::DefaultLieAlgebraOrthogonalBasis) get_coordinates(G::SpecialOrthogonalLieAlgebra, X, ::DefaultLieAlgebraOrthogonalBasis) get_coordinates!(G::OrthogonalLieAlgebra, c, X ::DefaultLieAlgebraOrthogonalBasis) get_coordinates!(G::SpecialOrthogonalLieAlgebra, c, X ::DefaultLieAlgebraOrthogonalBasis)
Compute the vector of coordinates from the Lie algebra tangent vector of the OrthogonalGroup O(n) in the DefaultLieAlgebraOrthogonalBasis, where is the dimension of the Lie algebra.

get_vector(G::OrthogonalLieAlgebra, e, c, ::DefaultLieAlgebraOrthogonalBasis) get_vector(G::SpecialOrthogonalLieAlgebra, e, c, ::DefaultLieAlgebraOrthogonalBasis) get_vector!(G::OrthogonalLieAlgebra, X, e, c, ::DefaultLieAlgebraOrthogonalBasis) get_vector!(G::SpecialOrthogonalLieAlgebra, X, e, c, ::DefaultLieAlgebraOrthogonalBasis)
Compute the tangent vector from a vector of coordinates .

lie_bracket(𝔰𝔢::LieAlgebra{ℝ, SpecialEuclideanGroupOperation, SpecialEuclideanGroup}, X::ArrayPartition, Y::ArrayPartition) lie_bracket!(𝔰𝔢::LieAlgebra{ℝ, SpecialEuclideanGroupOperation, SpecialEuclideanGroup}, Z::ArrayPartition, X::ArrayPartition, Y::ArrayPartition)
Calculate the Lie bracket between elements X and Y of the Lie algebra of the SpecialEuclideanGroup.

lie_bracket(𝔰𝔢::LieAlgebra{ℝ, SpecialEuclideanGroupOperation, SpecialEuclideanGroup}, X, Y) lie_bracket!(𝔰𝔢::LieAlgebra{ℝ, SpecialEuclideanGroupOperation, SpecialEuclideanGroup}, Z, X, Y)
Calculate the Lie bracket between elements X and Y of the Lie algebra of the SpecialEuclideanGroup.

lie_bracket(𝔰𝔢::LieAlgebra{ℝ, SpecialEuclideanGroupOperation, SpecialEuclideanGroup}, X, Y) lie_bracket!(𝔰𝔢::LieAlgebra{ℝ, SpecialEuclideanGroupOperation, SpecialEuclideanGroup}, Z, X, Y)
Calculate the Lie bracket between elements X and Y of the Lie algebra of the SpecialEuclideanGroup.

lie_bracket(𝔰𝔢::LieAlgebra{ℝ, SpecialEuclideanGroupOperation, SpecialEuclideanGroup}, X::ArrayPartition, Y::ArrayPartition) lie_bracket!(𝔰𝔢::LieAlgebra{ℝ, SpecialEuclideanGroupOperation, SpecialEuclideanGroup}, Z::ArrayPartition, X::ArrayPartition, Y::ArrayPartition)
Calculate the Lie bracket between elements X and Y of the Lie algebra of the SpecialEuclideanGroup.

lie_bracket(𝔰𝔤𝔞𝔩::LieAlgebra{ℝ,<:LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, X, Y) lie_bracket!(𝔰𝔤𝔞𝔩::LieAlgebra{ℝ,<:LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, Z, X, Y)
Compute the Lie bracket of two tangent vectors X, Y of the Lie algebra of the SpecialGalileanGroup(3), i.e. the matrix commutator of their hat representations.

lie_bracket(𝔰𝔤𝔞𝔩::LieAlgebra{ℝ,<:LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, X, Y) lie_bracket!(𝔰𝔤𝔞𝔩::LieAlgebra{ℝ,<:LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, Z, X, Y)
Compute the Lie bracket of two tangent vectors X, Y of the Lie algebra of the SpecialGalileanGroup(3), i.e. the matrix commutator of their hat representations.

X = hat(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, c) hat!(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, X, c)
Compute the hat map that turns a vector of coordinates c into a tangent vector in the Lie algebra.

X = hat(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, c) hat!(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, X, c)
Compute the hat map that turns a vector of coordinates c into a tangent vector in the Lie algebra.

c = vee(𝔤::LieAlgebra{ℝ,<:LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, X) vee!(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, c, X)
Compute the vee map that maps a tangent vector from the Lie algebra to a vector of coordinates c.

c = vee(𝔤::LieAlgebra{ℝ,<:LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, X) vee!(𝔤::LieAlgebra{ℝ,LeftSpecialGalileanGroupOperation,<:SpecialGalileanGroup}, c, X)
Compute the vee map that maps a tangent vector from the Lie algebra to a vector of coordinates c.

X = hat(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c) hat!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X, c)
Compute the hat map that turns a vector of coordinates c into a tangent vector in the Lie algebra.

X = hat(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c) hat!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X, c)
Compute the hat map that turns a vector of coordinates c into a tangent vector in the Lie algebra.

c = vee(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X) vee!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c, X)
Compute the vee map that maps a tangent vector from the Lie algebra to a vector of coordinates c.

c = vee(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, X) vee!(𝔤::LieAlgebra{ℝ,MatrixMultiplicationGroupOperation,<:SpecialLinearGroup}, c, X)
Compute the vee map that maps a tangent vector from the Lie algebra to a vector of coordinates c.

base_lie_group(A::GroupAction)
Return the AbstractLieGroup of the GroupAction specifying the action.

base_manifold(A::GroupAction)
Return the AbstractManifold the group action acts upon.

rand(::AbstractLieGroup; vector_at=nothing, σ::Real=1.0, kwargs...) rand(::AbstractLieGroup, PT::Type; vector_at=nothing, σ::Real=1.0, kwargs...) rand!(::LieAlgebra, T::Type; σ::Real=1.0, kwargs...) rand!(::AbstractLieGroup, gX::PT; vector_at=nothing, σ::Real=1.0, kwargs...) rand!(::LieAlgebra, X::T; σ::Real=1.0, kwargs...)
Compute a random point or tangent vector on a Lie group.

AbstractLieGroup{𝔽, O<:AbstractGroupOperation, M<:AbstractManifold{𝔽}} <: AbstractManifold{𝔽}
An abstract type to represent Lie groups.

LieGroup{𝔽, O<:AbstractGroupOperation, M<:AbstractManifold{𝔽}} <: AbstractLieGroup{𝔽, O, M}
Represent a Lie Group .

base_manifold(G::AbstractLieGroup)
Return the manifold stored within the AbstractLieGroup G.

is_point(G::AbstractLieGroup, g; kwargs...)
Check whether g is a valid point on the Lie Group G.

rand(::AbstractLieGroup; vector_at=nothing, σ::Real=1.0, kwargs...) rand(::AbstractLieGroup, PT::Type; vector_at=nothing, σ::Real=1.0, kwargs...) rand!(::LieAlgebra, T::Type; σ::Real=1.0, kwargs...) rand!(::AbstractLieGroup, gX::PT; vector_at=nothing, σ::Real=1.0, kwargs...) rand!(::LieAlgebra, X::T; σ::Real=1.0, kwargs...)
Compute a random point or tangent vector on a Lie group.

lie_bracket(𝔤::LieAlgebra{𝔽,AdditionGroupOperation}, X, Y) lie_bracket!(𝔤::LieAlgebra{𝔽,AdditionGroupOperation}, Z, X, Y)
Compute the Lie bracket , which for the for the AdditionGroupOperation simplifies to the corresponding zero_vector.

lie_bracket(::LieAlgebra{𝔽,MatrixMultiplicationGroupOperation}, X, Y) lie_bracket!(::LieAlgebra{𝔽,MatrixMultiplicationGroupOperation}, Z, X, Y)
Compute the Lie bracket , which for the for the MatrixMultiplicationGroupOperation yields the commutator bracket

lie_bracket(::LieAlgebra{𝔽,AbelianMultiplicationGroupOperation}, X, Y) lie_bracket!(::LieAlgebra{𝔽,AbelianMultiplicationGroupOperation}, Z, X, Y)
Compute the Lie bracket , which for the for the AbelianMultiplicationGroupOperation yields the zero vector of the LieAlgebra due to commutativity.

lie_bracket(::LieAlgebra{𝔽,AbelianMultiplicationGroupOperation}, X, Y) lie_bracket!(::LieAlgebra{𝔽,AbelianMultiplicationGroupOperation}, Z, X, Y)
Compute the Lie bracket , which for the for the AbelianMultiplicationGroupOperation yields the zero vector of the LieAlgebra due to commutativity.

lie_bracket(𝔤::LieAlgebra{𝔽,AdditionGroupOperation}, X, Y) lie_bracket!(𝔤::LieAlgebra{𝔽,AdditionGroupOperation}, Z, X, Y)
Compute the Lie bracket , which for the for the AdditionGroupOperation simplifies to the corresponding zero_vector.

lie_bracket(::LieAlgebra{𝔽,MatrixMultiplicationGroupOperation}, X, Y) lie_bracket!(::LieAlgebra{𝔽,MatrixMultiplicationGroupOperation}, Z, X, Y)
Compute the Lie bracket , which for the for the MatrixMultiplicationGroupOperation yields the commutator bracket