An interface for Lie groups

LieGroups.AbstractLieGroup — Type

AbstractLieGroup{𝔽, O<:AbstractGroupOperation, M<:AbstractManifold{𝔽}} <: AbstractManifold{𝔽}

An abstract type to represent Lie groups. For most cases it should suffice to “combine” an AbstractManifold with an AbstractGroupOperation, see LieGroup.

LieGroups.LieGroup — Type

LieGroup{𝔽, O<:AbstractGroupOperation, M<:AbstractManifold{𝔽}} <:  AbstractLieGroup{𝔽, O, M}

Represent a Lie Group $\mathcal G$.

A Lie Group $\mathcal G$ is a group endowed with the structure of a manifold such that the group operations $∘: \mathcal G×\mathcal G → \mathcal G$, see compose and the inverse operation $⋅^{-1}: \mathcal G → \mathcal G$, see inv are smooth, see for example [HN12, Definition 9.1.1].

Lie groups are named after the Norwegian mathematician Marius Sophus Lie (1842–1899).

Fields

manifold: an AbstractManifold $\mathcal M$

op: an AbstractGroupOperation $∘$ on that manifold

Constructor

LieGroup(M::AbstractManifold, op::AbstractGroupOperation)

Generate a Lie group based on a manifold M and a group operation op, where vectors by default are stored in the Lie Algebra.

LieGroups.AbstractLieAlgebraTangentVector — Type

AbstractLieAlgebraTangentVector <: ManifoldsBase.AbstractTangentVector

An abstract type for a tangent vector represented in a LieAlgebra.

While an tangent vectors are usually kept untyped for flexibility, it might be necessary to distinguish different types of points, for example

for complicated representations that require a struct

@ semantic verification

when there exist different representations

By sub-typing the AbstractManifoldPoint, this follows the same idea as in ManifoldsBase.jl.

LieGroups.AbstractLieGroupPoint — Type

AbstractLieGroupPoint <: ManifoldsBase.AbstractManifoldPoint end

An abstract type for a point on an AbstractLieGroup. While an points and tangent vectors are usually kept untyped for flexibility, it might be necessary to distinguish different types of points, for example

for complicated representations that require a struct

semantic verification

when there exist different representations

By sub-typing the AbstractManifoldPoint, this follows the same idea as in ManifoldsBase.jl.

Special types

LieGroups.BaseManifoldInverseRetraction — Type

BaseManifoldInverseRetraction{IRM<:AbstractInverseRetractionMethod} <: AbstractInverseRetractionMethod

Compute an inverse retraction by using the inverse retraction of type IRM on the base manifold of a LieGroup.

Constructor

BaseManifoldInverseRetraction(irm::AbstractInverseRetractionMethod)

Generate the inverse retraction with inverse retraction rm to use on the base manifold.

LieGroups.BaseManifoldRetraction — Type

BaseManifoldRetraction{RM<:AbstractRetractionMethod} <: AbstractRetractionMethod

Compute a retraction by using the retraction of type RM on the base manifold of a LieGroup.

Constructor

BaseManifoldRetraction(rm::AbstractRetractionMethod)

Generate the retraction with retraction rm to use on the base manifold.

LieGroups.BaseManifoldVectorTransportMethod — Type

BaseManifoldVectorTransportMethod{VTM<:AbstractVectorTransportMethod} <:
    AbstractVectorTransportMethod

Compute a vector transport by using the transport of type VTM on the base manifold of a LieGroup.

Constructor

BaseManifoldVectorTransportMethod(vtm::AbstractVectorTransportMethod)

Generate the vector transport with transport vtm to use on the base manifold.

Functions on Lie groups

Base.adjoint — Method

adjoint(G::AbstractLieGroup, g, X)
adjoint!(G::AbstractLieGroup, Y, g, X)

Compute the adjoint $\mathrm{Ad}(g): \mathfrak g → \mathfrak g$, which is defined as the differential diff_conjugate of the conjugate $c_g(h) = g∘h∘g^{-1}$ evaluated at the Identity $h=\mathrm{e}$. The operation can be performed in-place of Y.

\[ \mathrm{Ad}(g)[X] = \mathrm{d} c_g(\mathrm{e}) [X], \qquad X ∈ \mathfrak g.\]

see [HN12, Section 9.2.3].

On matrix Lie groups the adjoint reads $\mathrm{Ad}(g)[X] = gXg^{-1}$.

Base.exp — Method

exp(G::AbstractLieGroup, g, X)
exp!(G::AbstractLieGroup, h, g, X)

Compute the Lie group exponential map for $g∈\mathcal G$ and $X∈\mathfrak g$, where $\mathfrak g$ denotes the LieAlgebra of $\mathcal G$. It is given by

\[\exp_g X = g∘\exp_{\mathcal G}(X)\]

where X can be scaled by t, the computation can be performed in-place of h, and $\exp_{\mathcal G}$ denotes the Lie group exponential function.

If g is the Identity the Lie group exponential function $\exp_{\mathcal G}$ is computed directly. Implementing the Lie group exponential function introduces a default implementation with the formula above.

Note

The Lie group exponential map is usually different from the exponential map with respect to the metric of the underlying Riemannian manifold $\mathcal M$. To access the (Riemannian) exponential map, use exp(base_manifold(G), g, X).

Base.exp — Method

exp(G::AbstractLieGroup, X::T)
exp!(G::AbstractLieGroup, g, X)

Compute the (Lie group) exponential function

\[\exp_{\mathcal G}: \mathfrak g → \mathcal G,\qquad \exp_{\mathcal G}(X) = γ_X(1),\]

where $γ_X$ is the unique solution of the initial value problem

\[γ(0) = \mathrm{e}, \quad γ'(s) = γ(s)⋅X.\]

See also [HN12, Definition 9.2.2]. On matrix Lie groups this is the same as the matrix exponential.

The computation can be performed in-place of g.

Naming convention

There are at least two different objects usually called “exponential” that need to be distinguished

the (Riemannian) exponential map exp(M, p, X) from ManifoldsBase.jl. This can be accessed here using exp(base_manifold(G), p, X)

the exponential map for a (left/right/bi-invariant) Cartan-Schouten (pseudo-)metric exp(G, g, X), which we use as a default within this package

the (matrix/Lie group) exponential function exp(G, g) which agrees with the previous one for g being the identity there.

Base.inv — Method

inv(G::AbstractLieGroup, g)
inv!(G::AbstractLieGroup, h, g)

Compute the inverse group element $g^{-1}$ with respect to the AbstractGroupOperation $∘$ on the AbstractLieGroup $\mathcal G$, that is, return the unique element $h=g^{-1}$ such that $h∘g=\mathrm{e}$, where $\mathrm{e}$ denotes the Identity.

This can be done in-place of h, without side effects, that is you can do inv!(G, g, g).

Info

This function also handles the case where g is the Identity(G). Since this would lead to ambiguities when implementing a new group operations, this function calls _inv and _inv!, respectively, which is meant for the actual computation of group operations on (non-Identity` but maybe its numerical representation) elements.

Base.isapprox — Method

isapprox(M::AbstractLieGroup, g, h; kwargs...)

Check if points g and h from AbstractLieGroup are approximately equal. this function calls the corresponding isapprox on the AbstractManifold after handling the cases where one or more of the points are the Identity. All keyword argments are passed to this function as well.

Base.log — Method

log(G::AbstractLieGroup, g, h)
log!(G::AbstractLieGroup, X, g, h)

Compute the Lie group logarithmic map $\log_g: \mathcal G → \mathfrak g$, where $\mathfrak g$ denotes the LieAlgebra of $\mathcal G$. It is given by

\[\log_g h = \log_{\mathcal G}(g^{-1}∘h)\]

where $\log_{\mathcal G}$ denotes the Lie group logarithmic function The computation can be performed in-place of X.

Naming convention

There are at least two different objects usually called “logarithm” that need to be distinguished

the (Riemannian) logarithmic map log(M, p, X) from ManifoldsBase.jl

the exponential map for a (left/right/bi-invariant) Cartan-Schouten (pseudo-)metric exp(G, g, X), which we use as a default within this package

the (matrix/Lie group) exponential function exp(G, g) which agrees with the previous one for g being the identity there.

Base.log — Method

log(G::AbstractLieGroup, g, h)
log(G::AbstractLieGroup, g)
log(G::AbstractLieGroup, g::Identity, T)
log!(G::AbstractLieGroup, X::T, g)

Compute the (Lie group) logarithmic function $\log_{\mathcal G}: \mathcal G → \mathfrak g$, which is the inverse of the Lie group exponential function. For the allocating variant, you can specify the type T, when the point argument is the identity and hence does not provide the representation used. The computation can be performed in-place of X::T, which then determines the type.

Naming convention

There are at least two different objects usually called “logarithm” that need to be distinguished

the (Riemannian) logarithm map log(M, p, q) from ManifoldsBase.jl. This can be accessed here using log(base_manifold(G), p, q).

the logarithmic map for a (left/right/bi-invariant) Cartan-Schouten (pseudo-)metric log(G, g, h), which we use as a default within this package

the (matrix/Lie group) logarithm function log(G, h) which agrees with the previous one for g being the identity there.

Base.rand — Method

rand(::AbstractLieGroup; vector_at=nothing, σ::Real=1.0, kwargs...)
rand(::AbstractLieGroup, PT::Type; vector_at=nothing, σ::Real=1.0, kwargs...)
rand!(::LieAlgebra, T::Type; σ::Real=1.0, kwargs...)
rand!(::AbstractLieGroup, gX::PT; vector_at=nothing, σ::Real=1.0, kwargs...)
rand!(::LieAlgebra, X::T; σ::Real=1.0, kwargs...)

Compute a random point or tangent vector on a Lie group.

For points this just means to generate a random point on the underlying manifold itself.

For tangent vectors, an element in the Lie Algebra is generated, see also rand(::LieAlgebra; kwargs...)

For both cases, you can provide the type $T$ for the tangent vector and/or point $PT$, if you want to generate a random point in a certain representation. For the in-place variants the type is inferred from pX and X, respectively.

Base.rand — Method

rand(::AbstractLieGroup; vector_at=nothing, σ::Real=1.0, kwargs...)
rand(::AbstractLieGroup, PT::Type; vector_at=nothing, σ::Real=1.0, kwargs...)
rand!(::LieAlgebra, T::Type; σ::Real=1.0, kwargs...)
rand!(::AbstractLieGroup, gX::PT; vector_at=nothing, σ::Real=1.0, kwargs...)
rand!(::LieAlgebra, X::T; σ::Real=1.0, kwargs...)

Compute a random point or tangent vector on a Lie group.

For points this just means to generate a random point on the underlying manifold itself.

For tangent vectors, an element in the Lie Algebra is generated, see also rand(::LieAlgebra; kwargs...)

For both cases, you can provide the type $T$ for the tangent vector and/or point $PT$, if you want to generate a random point in a certain representation. For the in-place variants the type is inferred from pX and X, respectively.

LieGroups.compose! — Method

compose(G::AbstractLieGroup, g, h)
compose!(G::AbstractLieGroup, k, g, h)

Perform the group operation $g ∘ h$ for two $g, h ∈ \mathcal G$ on the AbstractLieGroup G. This can also be done in-place of h.

Info

This function also handles the case where g or/and h are the Identity(G). Since this would lead to ambiguities when implementing a new group operations, this function calls _compose and _compose!, respectively, which is meant for the actual computation of group operations on (non-Identity` but maybe its numerical representation) elements.

LieGroups.compose — Method

compose(G::AbstractLieGroup, g, h)
compose!(G::AbstractLieGroup, k, g, h)

Perform the group operation $g ∘ h$ for two $g, h ∈ \mathcal G$ on the AbstractLieGroup G. This can also be done in-place of h.

Info

This function also handles the case where g or/and h are the Identity(G). Since this would lead to ambiguities when implementing a new group operations, this function calls _compose and _compose!, respectively, which is meant for the actual computation of group operations on (non-Identity` but maybe its numerical representation) elements.

LieGroups.conjugate! — Method

conjugate(G::AbstractLieGroup, g, h)
conjugate!(G::AbstractLieGroup, k, g, h)

Compute the conjugation map $c_g: \mathcal G → \mathcal G$ given by $c_g(h) = g∘h∘g^{-1}$. This can be done in-place of k.

LieGroups.conjugate — Method

conjugate(G::AbstractLieGroup, g, h)
conjugate!(G::AbstractLieGroup, k, g, h)

Compute the conjugation map $c_g: \mathcal G → \mathcal G$ given by $c_g(h) = g∘h∘g^{-1}$. This can be done in-place of k.

LieGroups.diff_conjugate! — Method

diff_conjugate(G::AbstractLieGroup, g, h, X)
diff_conjugate!(G::AbstractLieGroup, Y, g, h, X)

Compute the differential of the conjugate $c_g(h) = g∘h∘g^{-1}$ on the AbstractLieGroup G. The operation can be performed in-place of Y.

\[ \mathrm{d}(c_g(h))[X], \qquad X ∈ \mathfrak g.\]

LieGroups.diff_conjugate — Method

diff_conjugate(G::AbstractLieGroup, g, h, X)
diff_conjugate!(G::AbstractLieGroup, Y, g, h, X)

Compute the differential of the conjugate $c_g(h) = g∘h∘g^{-1}$ on the AbstractLieGroup G. The operation can be performed in-place of Y.

\[ \mathrm{d}(c_g(h))[X], \qquad X ∈ \mathfrak g.\]

LieGroups.diff_inv! — Method

diff_inv(G::AbstractLieGroup, g, X)
diff_inv!(G::AbstractLieGroup, Y, g, X)

Compute the differential of the function $ι_{\mathcal G}(g) = g^{-1}$, where $\mathrm{d}ι_{\mathcal G}(g): \mathfrak g → \mathfrak g$. This can be done in-place of Y. Note that we represent tangent vectors in the Lie algebra $𝔤$.

For example on matrix manifolds this means, we use $X ∈ 𝔤$ and hence $W = gX ∈ T_g\mathcal G$. The (classical) differential $\mathrm{D}ι_{\mathcal G}(g): T_g\mathcal G → T_{g^{-1}}\mathcal G$ reads

\[ \mathrm{D}ι_{\mathcal G}(g)[W] = -g^{-1}Wg^{-1} = -Xg^{-1} = -g^{-1}(gXg^{-1}) = -g^{-1}\mathrm{Ad}(g)[X] = V ∈ T_{g^{-1}}\mathcal G,\]

see e.g. [Gil08]. To bring this back to the Lie algebra, we Write $V = g^{-1}Y ∈ T_{g^{-1}}\mathcal G$ for some $Y ∈ 𝔤$ and obtain

\[ \mathrm{d} ι_{\mathcal G}(g)[X] = -\mathrm{Ad}(g)[X] ∈ 𝔤,\]

where we use $\mathrm{d}$ to denote the differential in the Lie algebra.

LieGroups.diff_inv — Method

diff_inv(G::AbstractLieGroup, g, X)
diff_inv!(G::AbstractLieGroup, Y, g, X)

Compute the differential of the function $ι_{\mathcal G}(g) = g^{-1}$, where $\mathrm{d}ι_{\mathcal G}(g): \mathfrak g → \mathfrak g$. This can be done in-place of Y. Note that we represent tangent vectors in the Lie algebra $𝔤$.

For example on matrix manifolds this means, we use $X ∈ 𝔤$ and hence $W = gX ∈ T_g\mathcal G$. The (classical) differential $\mathrm{D}ι_{\mathcal G}(g): T_g\mathcal G → T_{g^{-1}}\mathcal G$ reads

\[ \mathrm{D}ι_{\mathcal G}(g)[W] = -g^{-1}Wg^{-1} = -Xg^{-1} = -g^{-1}(gXg^{-1}) = -g^{-1}\mathrm{Ad}(g)[X] = V ∈ T_{g^{-1}}\mathcal G,\]

see e.g. [Gil08]. To bring this back to the Lie algebra, we Write $V = g^{-1}Y ∈ T_{g^{-1}}\mathcal G$ for some $Y ∈ 𝔤$ and obtain

\[ \mathrm{d} ι_{\mathcal G}(g)[X] = -\mathrm{Ad}(g)[X] ∈ 𝔤,\]

where we use $\mathrm{d}$ to denote the differential in the Lie algebra.

LieGroups.diff_left_compose! — Method

diff_left_compose(G::AbstractLieGroup, g, h, X)
diff_left_compose!(G::AbstractLieGroup, Y, g, h, X)

Compute the differential of the group operation $g∘h$, on an AbstractLieGroup G with respect to its first (left) argument g.

Another interpretation is to consider a function where we do a fixed multiplication from the right with h. i..e. the right group multiplication function $ρ_h(g) = g∘h$ (where the right refers to the fixed argument $h$).

In this notation, this function computes the differential $\mathrm{d}ρ_h: 𝔤 → 𝔤$.

For example on matrix Lie groups this means that for $X ∈ 𝔤$ we can start with $W = gX ∈ T_g\mathcal G$ and compute the (classical) differential $\mathrm{D}ρ_h(g): T_g\mathcal G → T_{g∘h}\mathcal G$.

It reads

\[ \mathrm{D}ρ_h(g)[W] = Wh = gXh = V ∈ T_{g∘h}\mathcal G.\]

To obtain the Lie group differential $\mathrm{d}ρ_h(g)$ we have to “pull back” $V$ from the tangent space $T_{g∘h}\mathcal G$ to the Lie algebra $𝔤$. We use the same identification, that we can write $V = ghY ∈ T_{g∘h}\mathcal G$. This means in practice that with $V = gXh = gh(h^{-1}Xh)$ differential reads

\[\mathrm{d} ρ_h(g)[X] = h^{-1}Xh = \mathrm{Ad}(h^{-1})[X] ∈ 𝔤,\]

where $\mathrm{Ad}$ denotes the adjoint.

LieGroups.diff_left_compose — Method

diff_left_compose(G::AbstractLieGroup, g, h, X)
diff_left_compose!(G::AbstractLieGroup, Y, g, h, X)

Compute the differential of the group operation $g∘h$, on an AbstractLieGroup G with respect to its first (left) argument g.

Another interpretation is to consider a function where we do a fixed multiplication from the right with h. i..e. the right group multiplication function $ρ_h(g) = g∘h$ (where the right refers to the fixed argument $h$).

In this notation, this function computes the differential $\mathrm{d}ρ_h: 𝔤 → 𝔤$.

For example on matrix Lie groups this means that for $X ∈ 𝔤$ we can start with $W = gX ∈ T_g\mathcal G$ and compute the (classical) differential $\mathrm{D}ρ_h(g): T_g\mathcal G → T_{g∘h}\mathcal G$.

It reads

\[ \mathrm{D}ρ_h(g)[W] = Wh = gXh = V ∈ T_{g∘h}\mathcal G.\]

To obtain the Lie group differential $\mathrm{d}ρ_h(g)$ we have to “pull back” $V$ from the tangent space $T_{g∘h}\mathcal G$ to the Lie algebra $𝔤$. We use the same identification, that we can write $V = ghY ∈ T_{g∘h}\mathcal G$. This means in practice that with $V = gXh = gh(h^{-1}Xh)$ differential reads

\[\mathrm{d} ρ_h(g)[X] = h^{-1}Xh = \mathrm{Ad}(h^{-1})[X] ∈ 𝔤,\]

where $\mathrm{Ad}$ denotes the adjoint.

LieGroups.diff_right_compose! — Method

diff_right_compose(G::AbstractLieGroup, g, h, X)
diff_right_compose!(G::AbstractLieGroup, Y, g, h, X)

Compute the differential of the group operation $g∘h$, on an AbstractLieGroup G with respect to its second (right) argument h.

Another interpretation is to consider a function where we do a fixed multiplication from the left with g. i..e. the left group multiplication function $λ_g(h) = g∘h$ (where the left refers to the fixed argument $g$.).

In this notation, this function $\mathrm{d}λ_g: 𝔤 → 𝔤$.

For example on matrix Lie groups this means that for $X ∈ 𝔤$ we can start with $W = hX ∈ T_h\mathcal G$ and compute the (classical) differential $\mathrm{D}λ_g(h): T_h\mathcal G → T_{g∘h}\mathcal G$.

It reads

\[ \mathrm{D}λ_g(h)[W] = gW = ghX ∈ T_{g∘h}\mathcal G.\]

To obtain the Lie group differential $\mathrm{d}λ_g(h)$ we have to multiply the result with $(gh)^{-1}$ from the left and move from $W$ to $X$. Then the differential just simplifies to the identity. It reads

\[\mathrm{d} λ_g(h)[X] = X ∈ 𝔤.\]

LieGroups.diff_right_compose — Method

diff_right_compose(G::AbstractLieGroup, g, h, X)
diff_right_compose!(G::AbstractLieGroup, Y, g, h, X)

Compute the differential of the group operation $g∘h$, on an AbstractLieGroup G with respect to its second (right) argument h.

Another interpretation is to consider a function where we do a fixed multiplication from the left with g. i..e. the left group multiplication function $λ_g(h) = g∘h$ (where the left refers to the fixed argument $g$.).

In this notation, this function $\mathrm{d}λ_g: 𝔤 → 𝔤$.

For example on matrix Lie groups this means that for $X ∈ 𝔤$ we can start with $W = hX ∈ T_h\mathcal G$ and compute the (classical) differential $\mathrm{D}λ_g(h): T_h\mathcal G → T_{g∘h}\mathcal G$.

It reads

\[ \mathrm{D}λ_g(h)[W] = gW = ghX ∈ T_{g∘h}\mathcal G.\]

To obtain the Lie group differential $\mathrm{d}λ_g(h)$ we have to multiply the result with $(gh)^{-1}$ from the left and move from $W$ to $X$. Then the differential just simplifies to the identity. It reads

\[\mathrm{d} λ_g(h)[X] = X ∈ 𝔤.\]

LieGroups.identity_element! — Method

identity_element(G::AbstractLieGroup)
identity_element(G::AbstractLieGroup, T)
identity_element!(G::AbstractLieGroup, e::T)

Return a point representation of the Identity on the AbstractLieGroup G. By default this representation is the default array or number representation. If there exist several representations, the type T can be used to distinguish between them, and it should be provided for both the AbstractLieGroupPoint as well as the AbstractLieAlgebraTangentVector if they differ, since maybe only one of these types might be available for the second signature.

It returns the corresponding default representation of $e$ as a point on G. This can be performed in-place of e.

LieGroups.identity_element — Method

identity_element(G::AbstractLieGroup)
identity_element(G::AbstractLieGroup, T)
identity_element!(G::AbstractLieGroup, e::T)

Return a point representation of the Identity on the AbstractLieGroup G. By default this representation is the default array or number representation. If there exist several representations, the type T can be used to distinguish between them, and it should be provided for both the AbstractLieGroupPoint as well as the AbstractLieAlgebraTangentVector if they differ, since maybe only one of these types might be available for the second signature.

It returns the corresponding default representation of $e$ as a point on G. This can be performed in-place of e.

LieGroups.inv_left_compose! — Method

inv_left_compose(G::AbstractLieGroup, g, h)
inv_left_compose!(G::AbstractLieGroup, k, g, h)

Compute the inverse of the left group operation $λ_g(h) = g∘h$, on the AbstractLieGroup G, that is, compute $λ_g^{-1}(h) = g^{-1}∘h$. This can be done in-place of k.

LieGroups.inv_left_compose — Method

inv_left_compose(G::AbstractLieGroup, g, h)
inv_left_compose!(G::AbstractLieGroup, k, g, h)

Compute the inverse of the left group operation $λ_g(h) = g∘h$, on the AbstractLieGroup G, that is, compute $λ_g^{-1}(h) = g^{-1}∘h$. This can be done in-place of k.

LieGroups.inv_right_compose! — Method

inv_right_compose(G::AbstractLieGroup, h, g)
inv_right_compose!(G::AbstractLieGroup, k, h, g)

Compute the inverse of the right group operation $ρ_g(h) = h∘g$, on the AbstractLieGroup G, that is compute $ρ_g^{-1}(h) = h∘g^{-1}$. This can be done in-place of k.

LieGroups.inv_right_compose — Method

inv_right_compose(G::AbstractLieGroup, h, g)
inv_right_compose!(G::AbstractLieGroup, k, h, g)

Compute the inverse of the right group operation $ρ_g(h) = h∘g$, on the AbstractLieGroup G, that is compute $ρ_g^{-1}(h) = h∘g^{-1}$. This can be done in-place of k.

LieGroups.is_identity — Method

is_identity(G::AbstractLieGroup, q; kwargs...)

Check whether q is the identity on the AbstractLieGroup $\mathcal G$. This means it is either the Identity{O} with the respect to the corresponding AbstractGroupOperation O, or (approximately) the correct point representation.

See also

identity_element, identity_element!

LieGroups.jacobian_conjugate — Function

jacobian_conjugate(G::AbstractLieGroup, g, h; basis::AbstractBasis=DefaultLieAlgebraOrthogonalBasis(); X=zero_vector(LieAlgebar(G)))
jacobian_conjugate!(G::AbstractLieGroup, J, g, h; basis::AbstractBasis=DefaultLieAlgebraOrthogonalBasis(); X=zero_vector(LieAlgebar(G)))

Compute the Jacobian of the conjugate $c_g(h) = g∘h∘g^{-1}$, with respect to an AbstractBasis of the LieAlgebra.

A default is implemented using diff_conjugate $\mathrm{d}(c_g(h))[X]$: the $j$th column of of the Jacobian matrix $J$ are given by the coefficients of the tangent vector $\mathrm{d}(c_g(h))[X_j]$ with respect to the basis $B$, where $X_j$ is the $j$th basis vector of $B$.

Note

For the case that h is the Identity and the relation of $\mathrm{d}(c_g(h))[X]$ to the adjoint $\mathrm{Ad}(g)$, the Jacobian then sometimes called “adjoint matrix”, e.g. in [SDA21], when choosing as a basis the DefaultLieAlgebraOrthogonalBasis() that is used for hat and vee.

Keyword arguments

X=zero_vector(LieAlgebra(G)) pass an interims memory to store the Lie algebra tangent vector in.

LieGroups.jacobian_conjugate! — Function

jacobian_conjugate(G::AbstractLieGroup, g, h; basis::AbstractBasis=DefaultLieAlgebraOrthogonalBasis(); X=zero_vector(LieAlgebar(G)))
jacobian_conjugate!(G::AbstractLieGroup, J, g, h; basis::AbstractBasis=DefaultLieAlgebraOrthogonalBasis(); X=zero_vector(LieAlgebar(G)))

Compute the Jacobian of the conjugate $c_g(h) = g∘h∘g^{-1}$, with respect to an AbstractBasis of the LieAlgebra.

A default is implemented using diff_conjugate $\mathrm{d}(c_g(h))[X]$: the $j$th column of of the Jacobian matrix $J$ are given by the coefficients of the tangent vector $\mathrm{d}(c_g(h))[X_j]$ with respect to the basis $B$, where $X_j$ is the $j$th basis vector of $B$.

Note

For the case that h is the Identity and the relation of $\mathrm{d}(c_g(h))[X]$ to the adjoint $\mathrm{Ad}(g)$, the Jacobian then sometimes called “adjoint matrix”, e.g. in [SDA21], when choosing as a basis the DefaultLieAlgebraOrthogonalBasis() that is used for hat and vee.

Keyword arguments

X=zero_vector(LieAlgebra(G)) pass an interims memory to store the Lie algebra tangent vector in.

LieGroups.jacobian_exp — Function

jacobian_exp(G::AbstractLieGroup, g, X, b)
jacobian_exp!(G::AbstractLieGroup, J, g, X, b)

Compute the Jacobian of the exp $\exp_g(X)$ with respect to an AbstractBasis of the LieAlgebra.

LieGroups.jacobian_exp! — Function

jacobian_exp(G::AbstractLieGroup, g, X, b)
jacobian_exp!(G::AbstractLieGroup, J, g, X, b)

Compute the Jacobian of the exp $\exp_g(X)$ with respect to an AbstractBasis of the LieAlgebra.

LieGroups.point_type — Method

point_type(G::AbstractLieGroup, tangent_vector_type::Type)

Change tangent_vector_type that is a type of tangent vector type on Lie group G to its matching type for representing points.

By default both these types are assumed to be identical.

LieGroups.pull_back_tangent! — Method

pull_back_tangent(G::AnstractLieGroup, g, X; kwargs...)
pull_back_tangent!(G::AbstractLiegroup, Y, g, X; kwargs...)

Given a tangent vector X on the tangent space at g interpreted as the one on the manifold, this function pulls it back to the LieAlgebra.

By default this function falls back to calling diff_left_compose, but compared to that function, this function also takes care about the change of representation.

For example if a default representation for tangent vectors on a manifold is also the Lie algebra, then this function simplifies to the identity. This is for example the case for the SpecialOrthogonalGroup and its Rotations manifold.

Keyword argument

e = identity_element(G, typeof(g)) – if you have a memory available to store an identity point in, you can pass that memory here.

LieGroups.pull_back_tangent — Method

pull_back_tangent(G::AnstractLieGroup, g, X; kwargs...)
pull_back_tangent!(G::AbstractLiegroup, Y, g, X; kwargs...)

Given a tangent vector X on the tangent space at g interpreted as the one on the manifold, this function pulls it back to the LieAlgebra.

By default this function falls back to calling diff_left_compose, but compared to that function, this function also takes care about the change of representation.

For example if a default representation for tangent vectors on a manifold is also the Lie algebra, then this function simplifies to the identity. This is for example the case for the SpecialOrthogonalGroup and its Rotations manifold.

Keyword argument

e = identity_element(G, typeof(g)) – if you have a memory available to store an identity point in, you can pass that memory here.

LieGroups.push_forward_tangent! — Method

push_forward_tangent(G::AnstractLieGroup, g, X)
push_forward_tangent!(G::AbstractLiegroup, Y, g, X)

Given a Lie algebra vector X on the LieAlgebra, this function pushes the vector forward to the tangent space at g interpreted as the one on the manifold.

By default this function falls back to calling diff_left_compose, but compared to that function, this function also takes care about the change of representation.

For example if a default representation for tangent vectors on a manifold is also the Lie algebra, then this function simplifies to the identity. This is for example the case for the SpecialOrthogonalGroup and its Rotations manifold.

Keyword argument

e = identity_element(G, typeof(g)) – if you have a memory available to store an identity point in, you can pass that memory here.

LieGroups.push_forward_tangent — Method

push_forward_tangent(G::AnstractLieGroup, g, X)
push_forward_tangent!(G::AbstractLiegroup, Y, g, X)

Given a Lie algebra vector X on the LieAlgebra, this function pushes the vector forward to the tangent space at g interpreted as the one on the manifold.

By default this function falls back to calling diff_left_compose, but compared to that function, this function also takes care about the change of representation.

For example if a default representation for tangent vectors on a manifold is also the Lie algebra, then this function simplifies to the identity. This is for example the case for the SpecialOrthogonalGroup and its Rotations manifold.

Keyword argument

e = identity_element(G, typeof(g)) – if you have a memory available to store an identity point in, you can pass that memory here.

LinearAlgebra.adjoint! — Method

adjoint(G::AbstractLieGroup, g, X)
adjoint!(G::AbstractLieGroup, Y, g, X)

Compute the adjoint $\mathrm{Ad}(g): \mathfrak g → \mathfrak g$, which is defined as the differential diff_conjugate of the conjugate $c_g(h) = g∘h∘g^{-1}$ evaluated at the Identity $h=\mathrm{e}$. The operation can be performed in-place of Y.

\[ \mathrm{Ad}(g)[X] = \mathrm{d} c_g(\mathrm{e}) [X], \qquad X ∈ \mathfrak g.\]

see [HN12, Section 9.2.3].

On matrix Lie groups the adjoint reads $\mathrm{Ad}(g)[X] = gXg^{-1}$.

Manifolds.inv! — Method

inv(G::AbstractLieGroup, g)
inv!(G::AbstractLieGroup, h, g)

Compute the inverse group element $g^{-1}$ with respect to the AbstractGroupOperation $∘$ on the AbstractLieGroup $\mathcal G$, that is, return the unique element $h=g^{-1}$ such that $h∘g=\mathrm{e}$, where $\mathrm{e}$ denotes the Identity.

This can be done in-place of h, without side effects, that is you can do inv!(G, g, g).

Info

This function also handles the case where g is the Identity(G). Since this would lead to ambiguities when implementing a new group operations, this function calls _inv and _inv!, respectively, which is meant for the actual computation of group operations on (non-Identity` but maybe its numerical representation) elements.

ManifoldsBase.base_manifold — Method

base_manifold(G::AbstractLieGroup)

Return the manifold stored within the AbstractLieGroup G.

ManifoldsBase.exp! — Method

exp(G::AbstractLieGroup, g, X)
exp!(G::AbstractLieGroup, h, g, X)

Compute the Lie group exponential map for $g∈\mathcal G$ and $X∈\mathfrak g$, where $\mathfrak g$ denotes the LieAlgebra of $\mathcal G$. It is given by

\[\exp_g X = g∘\exp_{\mathcal G}(X)\]

where X can be scaled by t, the computation can be performed in-place of h, and $\exp_{\mathcal G}$ denotes the Lie group exponential function.

If g is the Identity the Lie group exponential function $\exp_{\mathcal G}$ is computed directly. Implementing the Lie group exponential function introduces a default implementation with the formula above.

Note

The Lie group exponential map is usually different from the exponential map with respect to the metric of the underlying Riemannian manifold $\mathcal M$. To access the (Riemannian) exponential map, use exp(base_manifold(G), g, X).

ManifoldsBase.exp! — Method

exp(G::AbstractLieGroup, X::T)
exp!(G::AbstractLieGroup, g, X)

Compute the (Lie group) exponential function

\[\exp_{\mathcal G}: \mathfrak g → \mathcal G,\qquad \exp_{\mathcal G}(X) = γ_X(1),\]

where $γ_X$ is the unique solution of the initial value problem

\[γ(0) = \mathrm{e}, \quad γ'(s) = γ(s)⋅X.\]

See also [HN12, Definition 9.2.2]. On matrix Lie groups this is the same as the matrix exponential.

The computation can be performed in-place of g.

Naming convention

There are at least two different objects usually called “exponential” that need to be distinguished

the (Riemannian) exponential map exp(M, p, X) from ManifoldsBase.jl. This can be accessed here using exp(base_manifold(G), p, X)

the exponential map for a (left/right/bi-invariant) Cartan-Schouten (pseudo-)metric exp(G, g, X), which we use as a default within this package

the (matrix/Lie group) exponential function exp(G, g) which agrees with the previous one for g being the identity there.

ManifoldsBase.inverse_retract — Method

inverse_retract(G::AbstractLieGroup, g, h, m::BaseManifoldInverseRetraction)

Compute the inverse retraction of g and h on the AbstractLieGroup G by using an inverse retraction on the underlying manifold and pulling the result back to the Lie algebra.

ManifoldsBase.is_point — Method

is_point(G::AbstractLieGroup, g; kwargs...)

Check whether g is a valid point on the Lie Group G. This falls back to checking whether g is a valid point on the base_manifoldG. unless g is an Identity. Then, it is checked whether it is the identity element corresponding to G.

ManifoldsBase.log! — Method

log(G::AbstractLieGroup, g, h)
log!(G::AbstractLieGroup, X, g, h)

Compute the Lie group logarithmic map $\log_g: \mathcal G → \mathfrak g$, where $\mathfrak g$ denotes the LieAlgebra of $\mathcal G$. It is given by

\[\log_g h = \log_{\mathcal G}(g^{-1}∘h)\]

where $\log_{\mathcal G}$ denotes the Lie group logarithmic function The computation can be performed in-place of X.

Naming convention

There are at least two different objects usually called “logarithm” that need to be distinguished

the (Riemannian) logarithmic map log(M, p, X) from ManifoldsBase.jl

the exponential map for a (left/right/bi-invariant) Cartan-Schouten (pseudo-)metric exp(G, g, X), which we use as a default within this package

the (matrix/Lie group) exponential function exp(G, g) which agrees with the previous one for g being the identity there.

ManifoldsBase.log! — Method

log(G::AbstractLieGroup, g, h)
log(G::AbstractLieGroup, g)
log(G::AbstractLieGroup, g::Identity, T)
log!(G::AbstractLieGroup, X::T, g)

Compute the (Lie group) logarithmic function $\log_{\mathcal G}: \mathcal G → \mathfrak g$, which is the inverse of the Lie group exponential function. For the allocating variant, you can specify the type T, when the point argument is the identity and hence does not provide the representation used. The computation can be performed in-place of X::T, which then determines the type.

Naming convention

There are at least two different objects usually called “logarithm” that need to be distinguished

the (Riemannian) logarithm map log(M, p, q) from ManifoldsBase.jl. This can be accessed here using log(base_manifold(G), p, q).

the logarithmic map for a (left/right/bi-invariant) Cartan-Schouten (pseudo-)metric log(G, g, h), which we use as a default within this package

the (matrix/Lie group) logarithm function log(G, h) which agrees with the previous one for g being the identity there.

ManifoldsBase.retract — Method

retract(G::AbstractLieGroup, g, h, m::BaseManifoldRetraction)

Compute the retraction of g and X on the AbstractLieGroup G by pushing X forward to the tangent space at g and using a retraction on the underlying manifold.

ManifoldsBase.vector_transport_to — Method

vector_transport_to(G::AbstractLieGroup, g, X, h, m::BaseManifoldVectorTransportMethod)

Compute the vector transport of a Lie algebra X from g to h using a vector transport on the underlying manifold. This is done by pushing X forward to the tangent space at g, then performing the vector transport on the base manifold, and finally pulling the resulting tangent vector back to the Lie algebra.

This method merely exists for experimental reasons, since the parallel transport on Lie groups, where all tangent vectors are represented in the Lie algebra is the identity. Hence any of the methods performed here are more costly than plain parallel transport.

Random.rand! — Method

rand(::AbstractLieGroup; vector_at=nothing, σ::Real=1.0, kwargs...)
rand(::AbstractLieGroup, PT::Type; vector_at=nothing, σ::Real=1.0, kwargs...)
rand!(::LieAlgebra, T::Type; σ::Real=1.0, kwargs...)
rand!(::AbstractLieGroup, gX::PT; vector_at=nothing, σ::Real=1.0, kwargs...)
rand!(::LieAlgebra, X::T; σ::Real=1.0, kwargs...)

Compute a random point or tangent vector on a Lie group.

For points this just means to generate a random point on the underlying manifold itself.

For tangent vectors, an element in the Lie Algebra is generated, see also rand(::LieAlgebra; kwargs...)

For both cases, you can provide the type $T$ for the tangent vector and/or point $PT$, if you want to generate a random point in a certain representation. For the in-place variants the type is inferred from pX and X, respectively.

Internal functions and macros

LieGroups.CommonUnitarySubAlgebra — Type

CommonUnitarySubAlgebra{𝔽,T}

A constant that allows to refer to several sub Algebras of $\mathfrak u(n)$ for implementations where

certain sub algebras real/complex share a common implementation, e.g. for the same sizes T usually via TypeParameter

certain functions are the same for all sizes T as long as the field 𝔽 is the same

LieGroups.@default_lie_group_fallbacks — Macro

default_lie_group_fallbacks(TG, TF, TP, TV, pfield::Symbol, Xfield::Symbol, groupOp)

Introduce default fallbacks for all basic functions on Lie groups, for Lie group of type TG with group operation Op, points of type TP, tangent vectors of type TV, with forwarding to fields pfield and Xfield for point and tangent vector functions, respectively.

A validation Lie group

LieGroups.ValidationLieGroup — Type

ValidationLieGroup{L<:AbstractLieGroup} <: AbstractLieGroup

A Lie group to add tests to input parameters and output values of functions defined for LieGroups.

Using the ignore_contexts keyword allows to specify a single Symbol or a vector of Symbols Of which contexts to ignore.

Current contexts are

:All: disable all checks

:Point: checks for points

:Algebra: checks related to the LieAlgebra

:Output: checks for output

:Input: checks for input variables

Fields

lie_group::L the AbstractLieGroup to be decorated

mode::Symbol: The mode to be used for error handling, either :error or :warn

ignore_contexts::AbstractVector{Symbol}: store contexts to be ignored of validation.

ignore_functions::Dict{<:Function,<:Union{Symbol,<:AbstractVector{Symbol}}: store contexts to be ignored with in a function or its mutating variant.

where all but the first field are analogous to the setups of the ValidationManifold. We refer to those docs for more examples on their meaning.

Constructor

ValidationLieGroup(L::AbstractLieGroup, check_manifold=true; kwargs...)

Generate the Validation Lie Group for the given AbstractLieGroup L. If check_manifold is set to true the inner manifold is additionally wrapped in a ValidationManifold. All suitable keywords are passed to the constructor of the validation manifold as well.

Keyword arguments

error::Symbol=:error: specify how errors in the validation should be reported. this is passed to is_point and is_vector as the error keyword argument. Available values are :error, :warn, :info, and :none. Every other value is treated as :none.

ignore_contexts = Vector{Symbol}() a vector to indicate which validation contexts should not be performed.

ignore_functions=Dict{Function,Union{Symbol,Vector{Symbol}}}() a dictionary to disable certain contexts within functions. The key here is the non-mutating function variant (if it exists). The contexts are the same as in ignore_contexts.

A metric Lie group to change the metric on a Lie Algebra

LieGroups.MetricLieGroup — Type

MetricLieGroup{
    𝔽, O <: AbstractGroupOperation, M <: ManifoldsBase.AbstractManifold{𝔽},
    L <: LieGroup{𝔽, O, M}, G <: AbstractMetric
} <: AbstractLieGroup{𝔽, O, M}}

Equip an AbstractLieGroup $\mathcal G$ explicitly with an AbstractMetric ⟨⋅,⋅⟩.

By default every AbstractLieGroup $\mathcal G$ is assumed to implicitly implement all functions on their LieAlgebra with respect to a metric, that corresponds to a certain “default metric”, either because it is a widely recognized metric or because the first implementation within LieGroups.jl was done with respect to this metric.

This default is usually indicated by checking that metric(L) returns DefaultMetric.

This decorator type allows to explicitly equip the Lie group with a different metric. Note that for any functions unrelated to the metric, this new Lie group will simply forward to the Lie group that is internally stored. This for example includes group operations like [compose], inv, or identity_element, or functions that are already otherwise passed on to the inner manifold like manifold_dimension.

Constructor

MetricLieGroup(G::AbstractLieGroup, metric::AbstractMetric)

Generate the AbstractLieGroup equipped explicitly with the AbstractMetric metric.

Manifolds.metric — Function

metric(G::MetricLieGroup)

Get the AbstractMetric associated with the MetricLieGroup $\mathcal G$.

Internal functions

LieGroups._vLc — Function

_vLc(M::ValidationLieGroup, f::Function, context::Symbol)
_vLc(M::ValidationLieGroup, f::Function, context::NTuple{N,Symbol}) where {N}
_vLc(M::ValidationLieGroup, ::Nothing, context::NTuple{N,Symbol}) where {N}

Return whether a check should be performed within f and the context(s) provided, if the second argument is :Nothing, only the context is checked

This function returns false and hence indicates not to check, when

(one of the) context(s) is in the ignore list for f within ignore_functions

(one of the) context(s) is in the ignore_contexts list

Otherwise the test is active.

!!! Note This function is internal and used very often, so it has a very short name; _vLc stands for "ValidationLieGroup check".

LieGroups._msg — Function

_msg(G::ValidationLieGroup, str; error=:None, within::Union{Nothing,<:Function} = nothing,
context::Union{NTuple{N,Symbol} where N} = NTuple{0,Symbol}())

issue a message str according to the mode mode (as @error, @warn, @info).

Literature

[Gil08]

M. B. Giles. Collected Matrix Derivative Results for Forward and Reverse Mode Algorithmic Differentiation. In: Advances in Automatic Differentiation, Lecture Notes in Computational Science and Engineering, edited by C. H. Bischof, H. M. Bücker, P. Hovland, U. Naumann and J. Utke (Springer, Berlin, Heidelberg, 2008); pp. 35–44.

[HN12]

J. Hilgert and K.-H. Neeb. Structure and Geometry of Lie Groups (Springer Monographs in Mathematics, 2012).

[SDA21]

J. Solà, J. Deray and D. Atchuthan. A micro Lie theory for state estimation in robotics (Dec 2021), arXiv:1812.01537 [cs.RO], arXiv: 1812.01537.