LieGroups.jl

Lie groups and Lie algebras in Julia

Work with Lie groups, their Lie algebras, and group actions

Get started List of Lie groups Lie group interface

🪶

Lightweight Interface

This package provides a lightweight interface to define Lie groups based on the LieGroup combining a AbstractManifold from ManifoldsBase.jl with a GroupOperation. Especially the exp and log maps are here the Lie group ones.

🧮

Lie Algebra and group actions

The LieAlgebra is generically provided including an . Furthermore GroupActions allow Lie groups to act on other manifolds.

⚡️

Efficient

When possible, functions are available working in-place, like exp! or log! to reduce memory allocations. The neutral element, the Identity{<:GroupOperation} provides an allocation free implementation that can “materialise” into the correct actual point on the Lie group when necessary.

Library of Lie groups

This package provides a library of Lie groups. On the one hand there are abstract product- and power- as well as semidirect product Lie groups. On the other hand a library of concrete Lie groups is available as well.

📚

Well-documented and -tested

All Lie groups are documented – both their theoretical foundation and all numerical functionality. The theoretical background also refers to further literature. A test suite provides a comprehensive verification for any Lie group, existing and newly written, as well.

Manifolds.jl

The manifolds from Manifolds.jl build the foundation of the Lie groups implemented here. At the same time every LieGroup is also a manifold (with a different connection), so they can for example be used with Manopt.jl.

LieGroups.LieGroups — Module

LieGroups.jl: Lie groups and Lie algebras in Julia.

The package is named after the Norwegian mathematician Marius Sophus Lie (1842–1899).

📚 Documentation: juliamanifolds.github.io/LieGroups.jl/

📦 Repository: github.com/JuliaManifolds/LieGroups.jl

💬 Discussions: github.com/JuliaManifolds/LieGroups.jl/discussions

🎯 Issues: github.com/JuliaManifolds/LieGroups.jl/issues

The implemented Lie groups use the interface for manifolds in ManifoldsBase.jl together with an interface for Lie groups and Lie algebras as well as internally using the manifolds implemented in Manifolds.jl.

For more general information about the history of and contributions to the package see the About page.

Getting started

To install the package just type

using Pkg; Pkg.add("LieGroups")

Then you can directly start, for example consider the SpecialEuclideanGroup $\mathrm{SE}(3)$ representing all orientations and places an object can take in $ℝ^3$. These are characterised by a $3×3$ rotation matrix together with a point the object is at. For example. having such a point, we can use the Lie group logarithmic function log(G::SpecialEuclideanGroup, g) and the Lie group exponential function exp(G::SpecialEuclideanGroup, X) to create an orientation “half the way” from the origin pose.

The default representation is in homogeneous coordinates

using LieGroupsSE3 = SpecialEuclideanGroup(3)g = 1/sqrt(2) .* [1.0 -1.0 0.0 0.0; 1.0 1.0 0.0 3.0*sqrt(2); 0.0 0.0 sqrt(2) 0.0; 0.0 0.0 0.0 sqrt(2)]

4×4 Matrix{Float64}:
 0.707107  -0.707107  0.0  0.0
 0.707107   0.707107  0.0  3.0
 0.0        0.0       1.0  0.0
 0.0        0.0       0.0  1.0

Then half that pose is

h = exp(SE3, 0.5 .* log(SE3, g))

4×4 Matrix{Float64}:
 0.92388   -0.382683  0.0  0.298369
 0.382683   0.92388   0.0  1.5
 0.0        0.0       1.0  0.0
 0.0        0.0       0.0  1.0

To check, just “perform that movement” twice with the group operation compose of h with itself to get g back

compose(SE3, h, h)

4×4 Matrix{Float64}:
 0.707107  -0.707107  0.0  3.33067e-16
 0.707107   0.707107  0.0  3.0
 0.0        0.0       1.0  0.0
 0.0        0.0       0.0  1.0

for more details see the get started tutorial.

exp(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X) exp!(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g, X)
Computes the Lie group exponential on the complex CircleGroup, which coincides with the ordinary complex exponential.

exp(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X) exp!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g, X)
Compute the Lie group exponential on the CircleGroup, represented as two dimensional vectors in the real plane.

exp(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X) exp!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g, X)
Compute the Lie group exponential of a vector X of the LieAlgebra of the circle group, represented as angles in .

log(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g) log!(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X, g)
Compute the Lie group logarithm on the complex CircleGroup, which coincides with the ordinary complex logarithm.

log(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g) log!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X, g)
Compute the Lie group logarithm on the CircleGroup, represented as two dimensional vectors in the real plane.

log(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g) log!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X, g)
Compute the Lie group logarithm on the CircleGroup, represented as angles in .

exp(::GeneralLinearGroup, X) exp!(::GeneralLinearGroup, g, X)
Compute the Lie group exponential on the GeneralLinearGroup, which is given by the matrix exponential

exp(G::HeisenbergGroup, X) exp!(G::HeisenbergGroup, g, X)
Compute the Lie group exponential for the HeisenbergGroup G of the vector X.

log(G::HeisenbergGroup, g) log!(G::HeisenbergGroup, X, g)
Compute the Lie group logarithm for the HeisenbergGroup G.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

compose(L::LieGroup{𝔽,<:SemidirectProductGroupOperation{⋆,⋄,<:AbstractLeftGroupActionType,ActionActsOnLeft}}, g, h)
Let and be two Lie groups with group operations and , respectively.

compose(L::LieGroup{𝔽,<:SemidirectProductGroupOperation{⋆,⋄,<:AbstractLeftGroupActionType,ActionActsOnRight}}, g, h)
Let and be two Lie groups with group operations and , respectively.

compose(L::LieGroup{𝔽,SemidirectProductGroupOperation{⋄,⋆,<:AbstractRightGroupActionType,ActionActsOnLeft}}, g, h)
Let and be two Lie groups with group operations and , respectively.

compose(L::LieGroup{𝔽,LeftSemidirectProductGroupOperation{⋆,⋄,<:AbstractRightGroupActionType,ActionActsOnRight}}, g, h)
Let and be two Lie groups with group operations and , respectively.

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{2}} for dispatch.

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{3}} for dispatch.

log(G::SpecialEuclidean, g) log!(G::SpecialEuclidean, X, g)
Compute the Lie group logarithm function on the SpecialEuclideanGroup G, and G uses a TypeParameter{Tuple{2}} for dispatch.

log(G::SpecialEuclidean, g) log!(G::SpecialEuclidean, X, g)
Compute the Lie group logarithm function on the SpecialEuclideanGroup G, where e is the Identity on G uses a TypeParameter{Tuple{3}} for dispatch.

SpecialEuclideanGroup{T}
The special Euclidean group is the group of rigid motions of .

LieGroups.exp(M::SpecialGalileanGroup, X) LieGroups.exp!(M::SpecialGalileanGroup, h, X)
Compute the Lie group exponential function on the SpecialGalileanGroup(3), where X is an element of the Lie algebra.

LieGroups.log(M::SpecialGalileanGroup, g) LieGroups.log!(M::SpecialGalileanGroup, X, g)
Compute the Lie group logarithm function on the SpecialGalileanGroup(3), where g is a group element.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G::AbstractLieGroup, X::T) exp!(G::AbstractLieGroup, g, X)
Compute the (Lie group) exponential function

log(G::AbstractLieGroup, g, h) log(G::AbstractLieGroup, g) log(G::AbstractLieGroup, g::Identity, T) log!(G::AbstractLieGroup, X::T, g)
Compute the (Lie group) logarithmic function , which is the inverse of the Lie group exponential function.

compose(G::AbstractLieGroup, g, h) compose!(G::AbstractLieGroup, k, g, h)
Perform the group operation for two on the AbstractLieGroup G.

exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, X) exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, g, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, g, X)
Compute the Lie group exponential on a LieGroup at a point g or the Identity with an AbelianMultiplicationGroupOperation.

exp(G::LieGroup{𝔽,AdditionGroupOperation}, X) exp!(G::LieGroup{𝔽,AdditionGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AdditionGroupOperation.

exp(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, X) exp!(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AbstractMultiplicationGroupOperation, which simplifies to the matrix exponential.

log(G::LieGroup{𝔽,AdditionGroupOperation}, g) log!(G::LieGroup{𝔽,AdditionGroupOperation}, X, g)
Compute the Lie group logarithm on a LieGroup with an AdditionGroupOperation.

log(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, g) log!(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, X, g)
Compute the Lie group logarithm on a LieGroup with a concrete instance of AbstractMultiplicationGroupOperation, which simplifies to the (matrix) logarithm.

compose(G::LieGroup{𝔽,AdditionGroupOperation}, g, h) compose!(G::LieGroup{𝔽,AdditionGroupOperation}, k, g, h)
Compute the group operation composition of g and h with respect to the AdditionGroupOperation on G, which falls back to calling g+h, where + is assumed to be overloaded accordingly.

compose(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, g, h) compose!(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, k, g, h)
Compute the group operation composition of g and h on an LieGroup G.