The semidirect product Lie group

LieGroups.LeftSemidirectProductGroupOperation — Type

LeftSemidirectProductGroupOperation{O1,O2,A} <: AbstractGroupOperation

A struct to model a semidirect Lie group product.

Let $(\mathcal N, ⋄)$ and $(\mathcal H, ⋆)$ be two Lie groups with group operations $⋄$ and $⋆$, respectively, as well as a group action $σ: \mathcal H×\mathcal N → \mathcal N$, cf AbstractLeftGroupActionType.

We use here as well use the notation $σ_h: \mathcal N → \mathcal N$ as a family of maps on $\mathcal N$

Then we define a group operation $∘$ on the product manifold $\mathcal N×\mathcal H$ by

\[ (h_1,n_1) ∘ (h_2,n_2) := (h_1 ⋆ h_2, σ_{h_2}(n_1) ⋄ n_1).\]

See [HN12, Definition 9.2.22], second definition for more details.

Constructor

LeftSemidirectProductGroupOperation(
    op1::AbstractGroupOperation,
    op2::AbstractGroupOperation,
    action::AbstractGroupActionType
)

Parameters

op1::AbstractGroupOperation: The group operation $⋄$ on $\mathcal H$

op2::AbstractGroupOperation: The group operation $⋆$ on $\mathcal N$

action::AbstractGroupActionType The group action $σ$ of $\mathcal H$ on $\mathcal N$

LieGroups.RightSemidirectProductGroupOperation — Type

RightSemidirectProductGroupOperation{O1,O2,A} <: AbstractGroupOperation

A struct to model a semidirect Lie group product.

Let $(\mathcal N, ⋄)$ and $(\mathcal H, ⋆)$ be two Lie groups with group operations $⋄$ and $⋆$, respectively, as well as a group action $σ: \mathcal H×\mathcal N → \mathcal N$, cf AbstractGroupActionType.

We use here as well use the notation $σ_h: \mathcal N → \mathcal N$ as a family of maps on $\mathcal N$

Then we define a group operation $∘$ on the product manifold $\mathcal N×\mathcal H$ by

\[ (n_1,h_1) ∘ (n_2,h_2) := (n_1 ⋄ σ_{h_1}(n_2), h_1 ⋆ h_2)\]

See [HN12, Definition 9.2.22], first definition for more details.

Constructor

RightSemidirectProductGroupOperation(
    op1::AbstractGroupOperation,
    op2::AbstractGroupOperation,
    action::AbstractGroupActionType
)

Parameters

op1::AbstractGroupOperation: The group operation $⋆$ on $\mathcal N$

op2::AbstractGroupOperation: The group operation $⋄$ on $\mathcal H$

action::AbstractGroupActionType: The group action $σ$ of $\mathcal H$ on $\mathcal N$

LieGroups.:⋉ — Method

L1 ⋉ L2
⋉(L1, L2)

For two LieGroups L1, L2, generate the LeftSemidirectProductLieGroup(L1, L2), where the corresponding default_left_action(L1, L2) is used.

LieGroups.:⋊ — Method

L1 ⋊ L2
⋊(L1, L2)

For two LieGroups L1, L2, generate the RightSemidirectProductLieGroup(L1, L2), where the corresponding default_right_action(L1, L2) is used.

LieGroups.LeftSemidirectProductLieGroup — Function

LeftSemidirectProductLieGroup(
    N::LieGroup, H::LieGroup, action::AbstractGroupActionType=default_left_action(N, H)
)

Generate the semidirect product Lie Group $\mathcal G = N ⋉ H$ for an AbstractLeftGroupActionType using the LeftSemidirectProductGroupOperation for the group operation definition as well as [HN12, Definition 9.2.22], second definition, for more details.

The short form N⋉H can be used if the corresponding default_left_action(N,H) is the one you want to use.

LieGroups.RightSemidirectProductLieGroup — Function

RightSemidirectProductLieGroup(
    N::LieGroup, H::LieGroup, action::AbstractGroupActionType=default_right_action(N,H)
)

Generate the semidirect product Lie Group $\mathcal G = N ⋊ H$ for an AbstractLeftGroupActionType using the RightSemidirectProductGroupOperation for the group operation definition as well as [HN12, Definition 9.2.22], first definition, for more details.

The short form N⋊H can be used if the corresponding default_right_action(N,H) is the one you want to use.