The special orthogonal group

LieGroups.SpecialOrthogonalGroup — Type

SpecialOrthogonalGroup{T}

The special orthogonal group $\mathrm{SO}(n)$ is the Lie group consisting of the MatrixMultiplicationGroupOperation on the manifold of rotations Rotations.

Constructor

SpecialOrthogonalGroup(n::Int; kwargs...)

Generate special orthogonal group $\mathrm{SO}(n)$. All keyword arguments in kwargs... are passed on to Rotations as well.

For this Lie group, several implementations are already covered by the defaults in the generic (matrix) multiplication operation.

Functions

Base.exp — Method

exp(G, X)
exp!(G, g, X)

Compute the Lie group exponential function on the OrthogonalGroup $\mathrm{O}(2)$ or SpecialOrthogonalGroup $\mathrm{SO}(2)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

Since the Lie algebra of both groups agrees and consist of the set of skew symmetric matrices, these simplify for the case of $2×2$ matrices to $X=\begin{pmatrix} 0 & -α\\ α & 0\end{pmatrix}$, for some $α∈ℝ$.

Their exponential is

\[\exp_{\mathcal G}(X) = \begin{pmatrix} \cos(α) & -\sin(α)\\ \sin(α) & \cos(α)\end{pmatrix}.\]

This result can be computed in-place of g.

Note that since $\mathrm{O}(2)$ consists of two disjoint connected components and the exponential map is smooth, the result $g$ always lies in the connected component of the identity.

Base.exp — Method

exp(G, X)
exp!(G, g, X)

Compute the Lie group exponential function on the OrthogonalGroup $\mathrm{O}(3)$ or SpecialOrthogonalGroup $\mathrm{SO}(3)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

Since the Lie algebra of both groups agrees and consist of the set of skew symmetric matrices, the $3×3$ skew symmetric matrices are of the form

\[ X = \begin{pmatrix} 0 & -c & b\\ c & 0 & -a\\ -b & a & 0\end{pmatrix},\]

for some $a, b, c ∈ ℝ$. To compute the exponential, the Rodrigues' rotation formula can be used. With $α = \sqrt{a^2+b^2+c^2} = \frac{1}{\sqrt{2}}\lVert X \rVert_{}$ we obtain for $α ≠ 0$

\[\exp_{\mathcal G}(X) = I_3 + \frac{\sin}{α}X + \frac{(1 - \cos)}{α^2}X^2,\]

and \exp{\mathcal G}(X) = I3`` otherwise.

This result can be computed in-place of g.

Note that since $\mathrm{SO}(3)$ consists of two disjoint connected components and the exponential map is smooth, the result $g$ always lies in the connected component of the identity.

Base.exp — Method

exp(G, e, X)
exp!(G, e, g, X)

Compute the Lie group exponential function on the OrthogonalGroup $\mathrm{O}(4)$ or SpecialOrthogonalGroup $\mathrm{SO}(4)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

Similar to the $3×3$ case, an efficient computation is provided, adapted from [GX02], [AR13] with a few numerical stabilisations.

Base.log — Method

log(G, g)
log!(G, X, g)

Compute the Lie group logarithm function on the OrthogonalGroup $\mathrm{O}(2)$ or SpecialOrthogonalGroup $\mathrm{SO}(2)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

For the two-dimensional case, any rotation matrix $g$ can be represented as $\begin{pmatrix} \cos(α) & -\sin(α)\\ \sin(α) & \cos(α)\end{pmatrix}$. For the SpecialOrthogonalGroup, $g$ might also include reflections.

The logarithm is then

\[\log_{\mathcal G}(g) = \begin{pmatrix} 0 & α\\ -α & 0\end{pmatrix}.\]

This result can be computed in-place of X

Note the logarithmic map is only locally around the identity uniquely determined. Especially, since $\mathrm{O}(2)$ consists of two disjoint connected components and the exponential map is smooth, for any $g$ in the other component, the logarithmic map is defined, but not the inverse of the exponential map.

Base.log — Method

log(G, g)
log!(G, X, g)

Compute the Lie group logarithm function on the OrthogonalGroup $\mathrm{O}(3)$ or SpecialOrthogonalGroup $\mathrm{SO}(3)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

Here, $\exp_{\mathcal G}(X) = g$ is inverted using the Rodrigues' rotation formula

\[\exp_{\mathcal G}(X) = I_3 + \frac{\sin}{α}X + \frac{(1 - \cos)}{α^2}X^2,\]

For $α ∉ \{ 0, π \}$ we obtain $X$ from the observation that

\[\mathrm{tr}(g) = 1 + 2\cos(α) \qquad\text{ and }\qquad \frac{1}{2}(g-g^\mathrm{T}) = \sin(α)X.\]

For $α = 0$ we have $g = I_3$ and $X = 0$.

For $α = π$ we have to solve $X^2 = \frac{1}{2}(g-I_3)$, where $X$ is skew-symmetric and hence we have to solve for three unknowns.

\[\log_{\mathcal G}(g) = X.\]

This result can be computed in-place of X

Note the logarithmic map is only locally around the identity uniquely determined. Especially, since $\mathrm{O}(3)$ consists of two disjoint connected components and the exponential map is smooth, for any $g$ in the other component, the logarithmic map is defined, but not the inverse of the exponential map.

Base.log — Method

log(G, g)
log!(G, X, g)

Compute the Lie group logarithm function on the OrthogonalGroup $\mathrm{O}(4)$ or SpecialOrthogonalGroup $\mathrm{SO}(4)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

The implementation is based on a generalized variant of the Rodrigues' like formula. For details, see [GX02, Section 3].

This result can be computed in-place of X

Note the logarithmic map is only locally around the identity uniquely determined.

LieGroups.jacobian_exp — Function

jacobian_exp(M::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, X, ::DefaultLieAlgebraOrthogonalBasis)
jacobian_exp!(M::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, J, X, ::DefaultLieAlgebraOrthogonalBasis)

Compute Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialOrthogonalGroup(2) manifold.

It is equal to matrix $[1]$, see [SDA21], Appendix A.

LieGroups.jacobian_exp — Function

jacobian_exp(M::SpecialOrthogonalGroup{TypeParameter{Tuple{3}}}, X, ::DefaultLieAlgebraOrthogonalBasis)
jacobian_exp!(M::SpecialOrthogonalGroup{TypeParameter{Tuple{3}}}, J, X, ::DefaultLieAlgebraOrthogonalBasis)

Compute Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialOrthogonalGroup(3) manifold. The formula reads

\[J = 𝕀 + \frac{\cos(θ) - 1}{θ^2} X + \frac{θ - \sin(θ)}{θ^3} X^2,\]

where $θ$ is the norm of X. It is adapted from [Chi12], Eq. (10.86), to LieGroups.jl conventions.

LieGroups.jacobian_exp! — Method

jacobian_exp(M::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, X, ::DefaultLieAlgebraOrthogonalBasis)
jacobian_exp!(M::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, J, X, ::DefaultLieAlgebraOrthogonalBasis)

Compute Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialOrthogonalGroup(2) manifold.

It is equal to matrix $[1]$, see [SDA21], Appendix A.

LieGroups.jacobian_exp! — Method

jacobian_exp(M::SpecialOrthogonalGroup{TypeParameter{Tuple{3}}}, X, ::DefaultLieAlgebraOrthogonalBasis)
jacobian_exp!(M::SpecialOrthogonalGroup{TypeParameter{Tuple{3}}}, J, X, ::DefaultLieAlgebraOrthogonalBasis)

Compute Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialOrthogonalGroup(3) manifold. The formula reads

\[J = 𝕀 + \frac{\cos(θ) - 1}{θ^2} X + \frac{θ - \sin(θ)}{θ^3} X^2,\]

where $θ$ is the norm of X. It is adapted from [Chi12], Eq. (10.86), to LieGroups.jl conventions.

ManifoldsBase.exp! — Method

exp(G, X)
exp!(G, g, X)

Compute the Lie group exponential function on the OrthogonalGroup $\mathrm{O}(2)$ or SpecialOrthogonalGroup $\mathrm{SO}(2)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

Since the Lie algebra of both groups agrees and consist of the set of skew symmetric matrices, these simplify for the case of $2×2$ matrices to $X=\begin{pmatrix} 0 & -α\\ α & 0\end{pmatrix}$, for some $α∈ℝ$.

Their exponential is

\[\exp_{\mathcal G}(X) = \begin{pmatrix} \cos(α) & -\sin(α)\\ \sin(α) & \cos(α)\end{pmatrix}.\]

This result can be computed in-place of g.

Note that since $\mathrm{O}(2)$ consists of two disjoint connected components and the exponential map is smooth, the result $g$ always lies in the connected component of the identity.

ManifoldsBase.exp! — Method

exp(G, X)
exp!(G, g, X)

Compute the Lie group exponential function on the OrthogonalGroup $\mathrm{O}(3)$ or SpecialOrthogonalGroup $\mathrm{SO}(3)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

Since the Lie algebra of both groups agrees and consist of the set of skew symmetric matrices, the $3×3$ skew symmetric matrices are of the form

\[ X = \begin{pmatrix} 0 & -c & b\\ c & 0 & -a\\ -b & a & 0\end{pmatrix},\]

for some $a, b, c ∈ ℝ$. To compute the exponential, the Rodrigues' rotation formula can be used. With $α = \sqrt{a^2+b^2+c^2} = \frac{1}{\sqrt{2}}\lVert X \rVert_{}$ we obtain for $α ≠ 0$

\[\exp_{\mathcal G}(X) = I_3 + \frac{\sin}{α}X + \frac{(1 - \cos)}{α^2}X^2,\]

and \exp{\mathcal G}(X) = I3`` otherwise.

This result can be computed in-place of g.

Note that since $\mathrm{SO}(3)$ consists of two disjoint connected components and the exponential map is smooth, the result $g$ always lies in the connected component of the identity.

ManifoldsBase.exp! — Method

exp(G, e, X)
exp!(G, e, g, X)

Compute the Lie group exponential function on the OrthogonalGroup $\mathrm{O}(4)$ or SpecialOrthogonalGroup $\mathrm{SO}(4)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

Similar to the $3×3$ case, an efficient computation is provided, adapted from [GX02], [AR13] with a few numerical stabilisations.

ManifoldsBase.get_coordinates! — Method

get_coordinates(𝔤::OrthogonalLieAlgebra, X, ::DefaultLieAlgebraOrthogonalBasis)
get_coordinates(G::SpecialOrthogonalLieAlgebra, X, ::DefaultLieAlgebraOrthogonalBasis)
get_coordinates!(G::OrthogonalLieAlgebra, c, X ::DefaultLieAlgebraOrthogonalBasis)
get_coordinates!(G::SpecialOrthogonalLieAlgebra, c, X ::DefaultLieAlgebraOrthogonalBasis)

Compute the vector of coordinates $c ∈ ℝ^d$ from the Lie algebra tangent vector $X ∈ 𝔬(n)$ of the OrthogonalGroup O(n) in the DefaultLieAlgebraOrthogonalBasis, where $d$ is the dimension of the Lie algebra. This is also the version used in vee.

For $O(2)$ there is only one coefficient $α$ in the basis $\begin{pmatrix} 0 & -α\\ α & 0\end{pmatrix}$, which is returned as $c = (α)^\mathrm{T}$.

A usual basis representation of ``𝔬(3) is given by

\[ X = \begin{pmatrix} 0 & -γ & β\\ γ & 0 & -α\\ -β & α & 0\end{pmatrix},\]

hence the coordinate vector is $c = (α, β, γ)^\mathrm{T} ∈ ℝ^3$.

For n ≥ 4the lower triangular part is added toc` row-wise.

ManifoldsBase.get_coordinates — Method

get_coordinates(𝔤::OrthogonalLieAlgebra, X, ::DefaultLieAlgebraOrthogonalBasis)
get_coordinates(G::SpecialOrthogonalLieAlgebra, X, ::DefaultLieAlgebraOrthogonalBasis)
get_coordinates!(G::OrthogonalLieAlgebra, c, X ::DefaultLieAlgebraOrthogonalBasis)
get_coordinates!(G::SpecialOrthogonalLieAlgebra, c, X ::DefaultLieAlgebraOrthogonalBasis)

Compute the vector of coordinates $c ∈ ℝ^d$ from the Lie algebra tangent vector $X ∈ 𝔬(n)$ of the OrthogonalGroup O(n) in the DefaultLieAlgebraOrthogonalBasis, where $d$ is the dimension of the Lie algebra. This is also the version used in vee.

For $O(2)$ there is only one coefficient $α$ in the basis $\begin{pmatrix} 0 & -α\\ α & 0\end{pmatrix}$, which is returned as $c = (α)^\mathrm{T}$.

A usual basis representation of ``𝔬(3) is given by

\[ X = \begin{pmatrix} 0 & -γ & β\\ γ & 0 & -α\\ -β & α & 0\end{pmatrix},\]

hence the coordinate vector is $c = (α, β, γ)^\mathrm{T} ∈ ℝ^3$.

For n ≥ 4the lower triangular part is added toc` row-wise.

ManifoldsBase.get_vector! — Method

get_vector(G::OrthogonalLieAlgebra, e, c, ::DefaultLieAlgebraOrthogonalBasis)
get_vector(G::SpecialOrthogonalLieAlgebra, e, c, ::DefaultLieAlgebraOrthogonalBasis)
get_vector!(G::OrthogonalLieAlgebra, X, e, c, ::DefaultLieAlgebraOrthogonalBasis)
get_vector!(G::SpecialOrthogonalLieAlgebra, X, e, c, ::DefaultLieAlgebraOrthogonalBasis)

Compute the tangent vector $X ∈ 𝔬(n)$ from a vector of coordinates $c ∈ ℝ^d$.

Here $d$ is the dimension of the Lie algebra of the OrthogonalGroup O(n) and the coordinates are with respect to the DefaultLieAlgebraOrthogonalBasis. This is also the version used in hat.

For $O(2)$ there is only one coefficient ``$c = (α)^\mathrm{T}$ and hence $X = \begin{pmatrix} 0 & -α\\ α & 0\end{pmatrix}$ is returned.

For $n=3$ a usual representtion turns $c = (α, β, γ)^\mathrm{T} ∈ ℝ^3$ into

\[ X = \begin{pmatrix} 0 & -γ & β\\ γ & 0 & -α\\ -β & α & 0\end{pmatrix},\]

hence the coordinate vector is .

For n ≥ 4` all further coefficients are used to fill up the following rows of the lower triangular part – which determines the upper triangular part due to skew-symmetry

ManifoldsBase.get_vector — Method

get_vector(G::OrthogonalLieAlgebra, e, c, ::DefaultLieAlgebraOrthogonalBasis)
get_vector(G::SpecialOrthogonalLieAlgebra, e, c, ::DefaultLieAlgebraOrthogonalBasis)
get_vector!(G::OrthogonalLieAlgebra, X, e, c, ::DefaultLieAlgebraOrthogonalBasis)
get_vector!(G::SpecialOrthogonalLieAlgebra, X, e, c, ::DefaultLieAlgebraOrthogonalBasis)

Compute the tangent vector $X ∈ 𝔬(n)$ from a vector of coordinates $c ∈ ℝ^d$.

Here $d$ is the dimension of the Lie algebra of the OrthogonalGroup O(n) and the coordinates are with respect to the DefaultLieAlgebraOrthogonalBasis. This is also the version used in hat.

For $O(2)$ there is only one coefficient ``$c = (α)^\mathrm{T}$ and hence $X = \begin{pmatrix} 0 & -α\\ α & 0\end{pmatrix}$ is returned.

For $n=3$ a usual representtion turns $c = (α, β, γ)^\mathrm{T} ∈ ℝ^3$ into

\[ X = \begin{pmatrix} 0 & -γ & β\\ γ & 0 & -α\\ -β & α & 0\end{pmatrix},\]

hence the coordinate vector is .

For n ≥ 4` all further coefficients are used to fill up the following rows of the lower triangular part – which determines the upper triangular part due to skew-symmetry

ManifoldsBase.log! — Method

log(G, g)
log!(G, X, g)

Compute the Lie group logarithm function on the OrthogonalGroup $\mathrm{O}(2)$ or SpecialOrthogonalGroup $\mathrm{SO}(2)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

For the two-dimensional case, any rotation matrix $g$ can be represented as $\begin{pmatrix} \cos(α) & -\sin(α)\\ \sin(α) & \cos(α)\end{pmatrix}$. For the SpecialOrthogonalGroup, $g$ might also include reflections.

The logarithm is then

\[\log_{\mathcal G}(g) = \begin{pmatrix} 0 & α\\ -α & 0\end{pmatrix}.\]

This result can be computed in-place of X

Note the logarithmic map is only locally around the identity uniquely determined. Especially, since $\mathrm{O}(2)$ consists of two disjoint connected components and the exponential map is smooth, for any $g$ in the other component, the logarithmic map is defined, but not the inverse of the exponential map.

ManifoldsBase.log! — Method

log(G, g)
log!(G, X, g)

Compute the Lie group logarithm function on the OrthogonalGroup $\mathrm{O}(3)$ or SpecialOrthogonalGroup $\mathrm{SO}(3)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

Here, $\exp_{\mathcal G}(X) = g$ is inverted using the Rodrigues' rotation formula

\[\exp_{\mathcal G}(X) = I_3 + \frac{\sin}{α}X + \frac{(1 - \cos)}{α^2}X^2,\]

For $α ∉ \{ 0, π \}$ we obtain $X$ from the observation that

\[\mathrm{tr}(g) = 1 + 2\cos(α) \qquad\text{ and }\qquad \frac{1}{2}(g-g^\mathrm{T}) = \sin(α)X.\]

For $α = 0$ we have $g = I_3$ and $X = 0$.

For $α = π$ we have to solve $X^2 = \frac{1}{2}(g-I_3)$, where $X$ is skew-symmetric and hence we have to solve for three unknowns.

\[\log_{\mathcal G}(g) = X.\]

This result can be computed in-place of X

Note the logarithmic map is only locally around the identity uniquely determined. Especially, since $\mathrm{O}(3)$ consists of two disjoint connected components and the exponential map is smooth, for any $g$ in the other component, the logarithmic map is defined, but not the inverse of the exponential map.

ManifoldsBase.log! — Method

log(G, g)
log!(G, X, g)

Compute the Lie group logarithm function on the OrthogonalGroup $\mathrm{O}(4)$ or SpecialOrthogonalGroup $\mathrm{SO}(4)$, where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

The implementation is based on a generalized variant of the Rodrigues' like formula. For details, see [GX02, Section 3].

This result can be computed in-place of X

Note the logarithmic map is only locally around the identity uniquely determined.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

SpecialOrthogonalGroup{T}
The special orthogonal group is the Lie group consisting of the MatrixMultiplicationGroupOperation on the manifold of rotations Rotations.

jacobian_exp(M::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(M::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialOrthogonalGroup(2) manifold.

jacobian_exp(M::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(M::SpecialOrthogonalGroup{TypeParameter{Tuple{2}}}, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialOrthogonalGroup(2) manifold.

jacobian_exp(M::SpecialOrthogonalGroup{TypeParameter{Tuple{3}}}, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(M::SpecialOrthogonalGroup{TypeParameter{Tuple{3}}}, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialOrthogonalGroup(3) manifold.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

get_vector(G::OrthogonalLieAlgebra, e, c, ::DefaultLieAlgebraOrthogonalBasis) get_vector(G::SpecialOrthogonalLieAlgebra, e, c, ::DefaultLieAlgebraOrthogonalBasis) get_vector!(G::OrthogonalLieAlgebra, X, e, c, ::DefaultLieAlgebraOrthogonalBasis) get_vector!(G::SpecialOrthogonalLieAlgebra, X, e, c, ::DefaultLieAlgebraOrthogonalBasis)
Compute the tangent vector from a vector of coordinates .

get_vector(G::OrthogonalLieAlgebra, e, c, ::DefaultLieAlgebraOrthogonalBasis) get_vector(G::SpecialOrthogonalLieAlgebra, e, c, ::DefaultLieAlgebraOrthogonalBasis) get_vector!(G::OrthogonalLieAlgebra, X, e, c, ::DefaultLieAlgebraOrthogonalBasis) get_vector!(G::SpecialOrthogonalLieAlgebra, X, e, c, ::DefaultLieAlgebraOrthogonalBasis)
Compute the tangent vector from a vector of coordinates .

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

DefaultLieAlgebraOrthogonalBasis{𝔽} <: ManifoldsBase.AbstractOrthogonalBasis{𝔽,ManifoldsBase.TangentSpaceType}
Specify an orthogonal basis for a Lie algebra.

get_coordinates(𝔤::LieAlgebra, X::T, B::AbstractBasis) get_coordinates!(𝔤::LieAlgebra, c, X::T, B::AbstractBasis)
Return the vector of coordinates to the decomposition of X with respect to an AbstractBasis of the LieAlgebra 𝔤.

get_coordinates(𝔤::LieAlgebra, X::T, B::AbstractBasis) get_coordinates!(𝔤::LieAlgebra, c, X::T, B::AbstractBasis)
Return the vector of coordinates to the decomposition of X with respect to an AbstractBasis of the LieAlgebra 𝔤.

exp(G::AbstractLieGroup, g, X) exp!(G::AbstractLieGroup, h, g, X)
Compute the Lie group exponential map for and , where denotes the LieAlgebra of .

exp(G::AbstractLieGroup, X::T) exp!(G::AbstractLieGroup, g, X)
Compute the (Lie group) exponential function

log(G::AbstractLieGroup, g, h) log(G::AbstractLieGroup, g) log(G::AbstractLieGroup, g::Identity, T) log!(G::AbstractLieGroup, X::T, g)
Compute the (Lie group) logarithmic function , which is the inverse of the Lie group exponential function.

jacobian_exp(G::AbstractLieGroup, X, b) jacobian_exp!(G::AbstractLieGroup, J, X, b)
Compute the Jacobian of the Lie group exponential function at , represented in an AbstractBasis of the LieAlgebra .

jacobian_exp(G::AbstractLieGroup, X, b) jacobian_exp!(G::AbstractLieGroup, J, X, b)
Compute the Jacobian of the Lie group exponential function at , represented in an AbstractBasis of the LieAlgebra .

exp(G::AbstractLieGroup, g, X) exp!(G::AbstractLieGroup, h, g, X)
Compute the Lie group exponential map for and , where denotes the LieAlgebra of .

exp(G::AbstractLieGroup, X::T) exp!(G::AbstractLieGroup, g, X)
Compute the (Lie group) exponential function

log(G::AbstractLieGroup, g, h) log(G::AbstractLieGroup, g) log(G::AbstractLieGroup, g::Identity, T) log!(G::AbstractLieGroup, X::T, g)
Compute the (Lie group) logarithmic function , which is the inverse of the Lie group exponential function.

exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, X) exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, g, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, g, X)
Compute the Lie group exponential on a LieGroup at a point g or the Identity with an AbelianMultiplicationGroupOperation.

exp(G::LieGroup{𝔽,AdditionGroupOperation}, X) exp!(G::LieGroup{𝔽,AdditionGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AdditionGroupOperation.

exp(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, X) exp!(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AbstractMultiplicationGroupOperation, which simplifies to the matrix exponential.

log(G::LieGroup{𝔽,AdditionGroupOperation}, g) log!(G::LieGroup{𝔽,AdditionGroupOperation}, X, g)
Compute the Lie group logarithm on a LieGroup with an AdditionGroupOperation.

log(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, g) log!(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, X, g)
Compute the Lie group logarithm on a LieGroup with a concrete instance of AbstractMultiplicationGroupOperation, which simplifies to the (matrix) logarithm.

jacobian_exp(G::LieGroup{𝔽,AdditionGroupOperation}, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(G::LieGroup{𝔽,AdditionGroupOperation}, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute the Jacobian of the Lie group exponential in a basis of the Lie algebra on a LieGroup with an AdditionGroupOperation.

jacobian_exp(G::LieGroup{𝔽,AdditionGroupOperation}, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(G::LieGroup{𝔽,AdditionGroupOperation}, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute the Jacobian of the Lie group exponential in a basis of the Lie algebra on a LieGroup with an AdditionGroupOperation.

exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, X) exp(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, g, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, X) exp!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h, g, X)
Compute the Lie group exponential on a LieGroup at a point g or the Identity with an AbelianMultiplicationGroupOperation.

exp(G::LieGroup{𝔽,AdditionGroupOperation}, X) exp!(G::LieGroup{𝔽,AdditionGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AdditionGroupOperation.

exp(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, X) exp!(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, g, X)
Compute the Lie group exponential on a LieGroup with an AbstractMultiplicationGroupOperation, which simplifies to the matrix exponential.

log(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, h) log(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, g, h) log!(G::LieGroup{𝔽,<:AbeliantMultiplicationGroupOperation}, X, g) log!(G::LieGroup{𝔽,<:AbeliantMultiplicationGroupOperation}, X, g, h)
Compute the Lie group logarithm on a LieGroup at a point g or the Identity with a concrete instance of AbelianMultiplicationGroupOperation.

log(G::LieGroup{𝔽,AdditionGroupOperation}, g) log!(G::LieGroup{𝔽,AdditionGroupOperation}, X, g)
Compute the Lie group logarithm on a LieGroup with an AdditionGroupOperation.

log(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, g) log!(G::LieGroup{𝔽,<:AbstractMultiplicationGroupOperation}, X, g)
Compute the Lie group logarithm on a LieGroup with a concrete instance of AbstractMultiplicationGroupOperation, which simplifies to the (matrix) logarithm.

exp(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X) exp!(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g, X)
Computes the Lie group exponential on the complex CircleGroup, which coincides with the ordinary complex exponential.

exp(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X) exp!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g, X)
Compute the Lie group exponential on the CircleGroup, represented as two dimensional vectors in the real plane.

exp(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X) exp!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g, X)
Compute the Lie group exponential of a vector X of the LieAlgebra of the circle group, represented as angles in .

log(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g) log!(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X, g)
Compute the Lie group logarithm on the complex CircleGroup, which coincides with the ordinary complex logarithm.

log(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g) log!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X, g)
Compute the Lie group logarithm on the CircleGroup, represented as two dimensional vectors in the real plane.

log(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g) log!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X, g)
Compute the Lie group logarithm on the CircleGroup, represented as angles in .

exp(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X) exp!(::LieGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g, X)
Computes the Lie group exponential on the complex CircleGroup, which coincides with the ordinary complex exponential.

exp(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X) exp!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g, X)
Compute the Lie group exponential on the CircleGroup, represented as two dimensional vectors in the real plane.

exp(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X) exp!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g, X)
Compute the Lie group exponential of a vector X of the LieAlgebra of the circle group, represented as angles in .

log(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, g) log!(::CircleGroup{ℂ, AbelianMultiplicationGroupOperation, Circle{ℂ}}, X, g)
Compute the Lie group logarithm on the complex CircleGroup, which coincides with the ordinary complex logarithm.

log(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, g) log!(::LieGroup{ℝ, AbelianMultiplicationGroupOperation, Sphere}, X, g)
Compute the Lie group logarithm on the CircleGroup, represented as two dimensional vectors in the real plane.

log(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, g) log!(::LieGroup{ℝ, AdditionGroupOperation, Circle{ℝ}}, X, g)
Compute the Lie group logarithm on the CircleGroup, represented as angles in .

exp(::GeneralLinearGroup, X) exp!(::GeneralLinearGroup, g, X)
Compute the Lie group exponential on the GeneralLinearGroup, which is given by the matrix exponential

exp(::GeneralLinearGroup, X) exp!(::GeneralLinearGroup, g, X)
Compute the Lie group exponential on the GeneralLinearGroup, which is given by the matrix exponential

exp(G::HeisenbergGroup, g, X)
Exponential map on the HeisenbergGroup G with the left-invariant metric.

exp(G::HeisenbergGroup, X) exp!(G::HeisenbergGroup, g, X)
Compute the Lie group exponential for the HeisenbergGroup G of the vector X.

log(G::HeisenbergGroup, g) log!(G::HeisenbergGroup, X, g)
Compute the Lie group logarithm for the HeisenbergGroup G.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, X) exp!(G, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

exp(G, e, X) exp!(G, e, g, X)
Compute the Lie group exponential function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

get_coordinates(𝔤::OrthogonalLieAlgebra, X, ::DefaultLieAlgebraOrthogonalBasis) get_coordinates(G::SpecialOrthogonalLieAlgebra, X, ::DefaultLieAlgebraOrthogonalBasis) get_coordinates!(G::OrthogonalLieAlgebra, c, X ::DefaultLieAlgebraOrthogonalBasis) get_coordinates!(G::SpecialOrthogonalLieAlgebra, c, X ::DefaultLieAlgebraOrthogonalBasis)
Compute the vector of coordinates from the Lie algebra tangent vector of the OrthogonalGroup O(n) in the DefaultLieAlgebraOrthogonalBasis, where is the dimension of the Lie algebra.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

log(G, g) log!(G, X, g)
Compute the Lie group logarithm function on the OrthogonalGroup or SpecialOrthogonalGroup , where e is the Identity{MatrixMultiplicationGroupOperation} and G uses a TypeParameter for dispatch.

jacobian_exp(G::ProductLieGroup, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(G::ProductLieGroup, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute the Jacobian of the Lie group exponential in a basis of the Lie algebra on a ProductLieGroup.

jacobian_exp(G::ProductLieGroup, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(G::ProductLieGroup, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute the Jacobian of the Lie group exponential in a basis of the Lie algebra on a ProductLieGroup.

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{2}} for dispatch.

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{3}} for dispatch.

log(G::SpecialEuclidean, g) log!(G::SpecialEuclidean, X, g)
Compute the Lie group logarithm function on the SpecialEuclideanGroup G, and G uses a TypeParameter{Tuple{2}} for dispatch.

log(G::SpecialEuclidean, g) log!(G::SpecialEuclidean, X, g)
Compute the Lie group logarithm function on the SpecialEuclideanGroup G, where e is the Identity on G uses a TypeParameter{Tuple{3}} for dispatch.

jacobian_exp(G::SpecialEuclideanGroup{TypeParameter{Tuple{2}}}, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(G::SpecialEuclideanGroup{TypeParameter{Tuple{2}}}, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute the Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialEuclideanGroup(2).

jacobian_exp(G::SpecialEuclideanGroup{TypeParameter{Tuple{2}}}, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(G::SpecialEuclideanGroup{TypeParameter{Tuple{2}}}, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute the Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialEuclideanGroup(2).

jacobian_exp(G::SpecialEuclideanGroup{TypeParameter{Tuple{3}}}, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(G::SpecialEuclideanGroup{TypeParameter{Tuple{3}}}, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute the Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialEuclideanGroup(3).

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{3}} for dispatch.

exp(G::SpecialEuclidean, X) exp!(G::SpecialEuclidean, g, X)
Compute the Lie group exponential function on the SpecialEuclideanGroup G using a TypeParameter{Tuple{2}} for dispatch.

log(G::SpecialEuclidean, g) log!(G::SpecialEuclidean, X, g)
Compute the Lie group logarithm function on the SpecialEuclideanGroup G, and G uses a TypeParameter{Tuple{2}} for dispatch.

log(G::SpecialEuclidean, g) log!(G::SpecialEuclidean, X, g)
Compute the Lie group logarithm function on the SpecialEuclideanGroup G, where e is the Identity on G uses a TypeParameter{Tuple{3}} for dispatch.

LieGroups.exp(M::SpecialGalileanGroup, X) LieGroups.exp!(M::SpecialGalileanGroup, h, X)
Compute the Lie group exponential function on the SpecialGalileanGroup(3), where X is an element of the Lie algebra.

LieGroups.log(M::SpecialGalileanGroup, g) LieGroups.log!(M::SpecialGalileanGroup, X, g)
Compute the Lie group logarithm function on the SpecialGalileanGroup(3), where g is a group element.

jacobian_exp(G::SpecialGalileanGroup, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(G::SpecialGalileanGroup, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute the Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialGalileanGroup(3).

jacobian_exp(G::SpecialGalileanGroup, X, ::DefaultLieAlgebraOrthogonalBasis) jacobian_exp!(G::SpecialGalileanGroup, J, X, ::DefaultLieAlgebraOrthogonalBasis)
Compute the Jacobian of the Lie group exponential in a basis of the Lie algebra on the SpecialGalileanGroup(3).

LieGroups.exp(M::SpecialGalileanGroup, X) LieGroups.exp!(M::SpecialGalileanGroup, h, X)
Compute the Lie group exponential function on the SpecialGalileanGroup(3), where X is an element of the Lie algebra.

LieGroups.log(M::SpecialGalileanGroup, g) LieGroups.log!(M::SpecialGalileanGroup, X, g)
Compute the Lie group logarithm function on the SpecialGalileanGroup(3), where g is a group element.