An Interface for Lie group actions

LieGroups.AbstractGroupAction — Type

AbstractGroupAction{T<:AbstractGroupActionType, L<:LieGroup, M<:AbstractManifold}

An abstract group action of AbstractGroupActionType T of a LieGroup of type L acting on an AbstractManifold of type M.

See [HN12, Section 9.1.3] for more details.

LieGroups.AbstractGroupActionType — Type

AbstractGroupActionType

An abstract supertype for group actions.

LieGroups.AbstractLeftGroupActionType — Type

AbstractLeftGroupActionType <: AbstractGroupActionType

A type representing a (smooth) group action $σ: \mathcal G × \mathcal M → \mathcal M$ of a LieGroup $\mathcal G$ acting (from the left) on an AbstractManifold $\mathcal M$. with the following properties

$σ(\mathrm e, p) = p$ holds for all $p ∈ \mathcal M$

$σ(g, σ(h, p)) = σ(g∘h, p)$ holds for all $g,h ∈ \mathcal G$, $p ∈ \mathcal M$

where $∘$ denotes the group operation of the LieGroup $\mathcal G$. See also [HN12, Definition 9.1.11].

The type of action can be seen a bit better when writing the action as a family $σ_g(p)$: we obtain from the second property as

\[ σ_g(σ_h(p)) = σ_{gh}(p)\]

and see that $g$ appears on the left.

To emphasize the side the group operation is acting from, we sometimes write $σ^{\mathrm L}$. If the action is clear from context we write $σ(g, p) = g ⋅ p$.

One notable example of a left action is the inverse of an action of AbstractRightGroupActionType $σ^{\mathrm R}$, which is given by $τ_g = (σ^{\mathrm R}_g)^{-1} = σ^{\mathrm R}_{g^{-1}}$. We obtain

\[τ_g(τ_h(p)) = σ^{\mathrm R}_{g^{-1}}(σ^{\mathrm R}_{h^{-1}}(p)) = σ^{\mathrm R}_{h^{-1}g^{-1}}(p) = σ^{\mathrm R}_{(gh)^{-1}}(p) τ_{gh}(p).\]

Note

In function definitions we follow the idea of the family of actions and use the order (M, g, p) in function signatures.

LieGroups.AbstractRightGroupActionType — Type

AbstractRightGroupActionType <: AbstractGroupActionType

A type representing a (smooth) group action $σ: \mathcal M × \mathcal G → \mathcal M$ of a LieGroup $\mathcal G$ acting (from the right) on an AbstractManifold $\mathcal M$. with the following properties

$σ(p, \mathrm e) = p$ holds for all $p ∈ \mathcal M$

$σ(σ(p, g), h) = σ(p, g∘h)$ holds for all $g,h ∈ \mathcal G$, $p ∈ \mathcal M$

where $∘$ denotes the group operation of the LieGroup $\mathcal G$. See also [HN12, Remark 9.1.12].

The type of action can be seen a bit better when writing the action as a family $σ_g(p)$: we obtain from the second property as

\[ σ_g(σ_h(p)) = σ_{hg}(p)\]

and see that $g$ appears on the right.

To emphasize the side the group operation is acting from, we sometimes write $σ^{\mathrm R}$. If the action is clear from context we write $σ(p, g) = p ⋅ g$.

One notable example of a right action is the inverse of an action of AbstractLeftGroupActionType $σ^{\mathrm L}$, which is given by $τ_g = (σ^{\mathrm L}_g)^{-1} = σ^{\mathrm L}_{g^{-1}}$. We obtain

\[τ_g(τ_h(p)) = σ^{\mathrm L}_{g^{-1}}(σ^{\mathrm L}_{h^{-1}}(p)) = σ^{\mathrm L}_{g^{-1}h^{-1}}(p) = σ^{\mathrm L}_{(hg)^{-1}}(p) τ_{hg}(p).\]

Note

In function definitions we follow the idea of the family of actions and use the order (M, g, p) in function signatures.

Functions for Lie group actions

Base.inv — Method

 inv(T::AbstractGroupActionType)

Return the type representing the inverse of an AbstractGroupActionType.

Base.inv — Method

 inv(A::AbstractGroupAction{T})

Return the tuple representing the inverse of an AbstractGroupAction of AbstractGroupActionType T. This is usually done by returning the group action with the inverse type of T.

LieGroups.base_lie_group — Method

base_lie_group(A::AbstractGroupAction)

Return the LieGroup of the AbstractGroupAction specifying the action.

LieGroups.diff_apply! — Method

diff_apply(A::AbstractGroupAction{T, L, M}, g, p, X)
diff_apply!(A::AbstractGroupAction{T, L, M}, Y, g, p, X)

Compute the differential $D_p σ_g(p): T_p\mathcal M → T_{σ_g(p)}\mathcal M$, where for a left group action we have $σ_g(p) = σ(g,p)$, for a right action $σ_g(p) = σ(p, g)$.

LieGroups.diff_apply — Method

diff_apply(A::AbstractGroupAction{T, L, M}, g, p, X)
diff_apply!(A::AbstractGroupAction{T, L, M}, Y, g, p, X)

Compute the differential $D_p σ_g(p): T_p\mathcal M → T_{σ_g(p)}\mathcal M$, where for a left group action we have $σ_g(p) = σ(g,p)$, for a right action $σ_g(p) = σ(p, g)$.

LieGroups.diff_group_apply! — Method

diff_group_apply(A::AbstractGroupAction{T, L, M}, g, p, X)
diff_group_apply!(A::AbstractGroupAction{T, L, M}, Y, g, p, X)

Compute the differential $D_g σ_g(p): \mathfrak g → \mathfrak g$, where we use the short hand notation $σ_p(g) = σ(g,p)$ for a left action, and for a right action $σ_p(g) = σ(p, g)$.

LieGroups.diff_group_apply — Method

diff_group_apply(A::AbstractGroupAction{T, L, M}, g, p, X)
diff_group_apply!(A::AbstractGroupAction{T, L, M}, Y, g, p, X)

Compute the differential $D_g σ_g(p): \mathfrak g → \mathfrak g$, where we use the short hand notation $σ_p(g) = σ(g,p)$ for a left action, and for a right action $σ_p(g) = σ(p, g)$.

LieGroups.switch — Method

switch(T::AbstractGroupActionType)

Return the object representing an AbstractGroupActionType related to a group operation action that switched the side, that is it turns a left action type into its corresponding right action type.

LieGroups.switch — Method

 switch(A::AbstractGroupAction{T})

Return the group operation action representing the similar AbstractGroupAction of AbstractGroupActionType T but acting from the other side. It switches left to right and vice versa. This is done by returning the group action with the “switched” type of T.

Manifolds.apply! — Method

apply(A::AbstractGroupAction{T, L, M}, g, p)
apply!(A::AbstractGroupAction{T, L, M}, q, g, p)

Apply the group action induced by $g ∈ \mathcal G$ to $p ∈ \mathcal M$, where the kind of group action is indicated by the AbstractGroupActionType T. This can be perfomed in-place of q.

Manifolds.apply — Method

apply(A::AbstractGroupAction{T, L, M}, g, p)
apply!(A::AbstractGroupAction{T, L, M}, q, g, p)

Apply the group action induced by $g ∈ \mathcal G$ to $p ∈ \mathcal M$, where the kind of group action is indicated by the AbstractGroupActionType T. This can be perfomed in-place of q.

ManifoldsBase.base_manifold — Method

base_manifold(A::AbstractGroupAction)

Return the AbstractManifold the group action acts upon.

Generic Lie group actions

LieGroups.GroupOperationAction — Type

GroupOperationAction{T<:AbstractLeftGroupActionType,G<:LieGroup} <: AbstractGroupAction{T,G,G}

The AbstractGroupAction.

LieGroups.InverseLeftGroupOperation — Type

InverseLeftGroupOperation <: AbstractRightGroupActionType

A type for the AbstractLeftGroupActionType when acting on the group itself given by the inverse of a LeftGroupOperation $σ_h$ as

\[τ_h(g) \coloneqq σ_h^{-1}(g) = σ(h^{-1},g) = h^{-1}∘g\]

Note that while in naming it is the inverse of the left action, it's properties yield that is is an AbstractRightGroupActionType, since

\[τ_g(τ_h(p)) = σ^{\mathrm L}_{g^{-1}}(σ^{\mathrm L}_{h^{-1}}(p)) = σ^{\mathrm L}_{g^{-1}h^{-1}}(p) = σ^{\mathrm L}_{(hg)^{-1}}(p) τ_{hg}(p).\]

for its inverse $(σ_h)^{-1}$ see InverseLeftGroupOperation.

Note

Some literature also calls this by itself the right group operation action.

LieGroups.InverseRightGroupOperation — Type

InverseRightGroupOperation <: AbstractLeftGroupActionType

A type for the AbstractLeftGroupActionType when acting on the group itself given by the inverse of a RightGroupOperation $σ_h$ as

\[τ_h(g) \coloneqq σ_h^{-1}(g) = σ(h^{-1},g) = g∘h^{-1}\]

Note that while in naming it is the inverse of the right action, it's properties yield that is is an AbstractLeftGroupActionType, since

\[τ_g(τ_h(p)) = σ^{\mathrm R}_{g^{-1}}(σ^{\mathrm R}_{h^{-1}}(p)) = σ^{\mathrm R}_{h^{-1}g^{-1}}(p) = σ^{\mathrm R}_{(gh)^{-1}}(p) τ_{gh}(p).\]

for its inverse $(σ_h)^{-1}$ see InverseLeftGroupOperation.

LieGroups.LeftGroupOperation — Type

LeftGroupOperation <: AbstractLeftGroupActionType

A type for the AbstractLeftGroupActionType when acting on the group itself from the left, that is

\[σ_h(g) = σ(h,g) = h∘g\]

for its inverse $(σ_h)^{-1}$ see InverseLeftGroupOperation.

LieGroups.RightGroupOperation — Type

RightGroupOperation <: AbstractRightGroupActionType

A type for the AbstractLeftGroupActionType when acting on the group itself gfrom the right

\[σ_h(g) = σ(h,g) = g∘h\]

for its inverse $(σ_h)^{-1}$ see InverseRightGroupOperation.

Base.inv — Method

inv(::GroupOperationAction)

return the inverse group operation action, that is, use the type representing the inverse operation.

Base.inv — Method

inv(::InverseLeftGroupOperation)

Return the inverse of the InverseLeftGroupOperation, that is the LeftGroupOperation.

Base.inv — Method

inv(::InverseRightGroupOperation)

Return the inverse of the InverseRightGroupOperation, that is the RightGroupOperation.

Base.inv — Method

inv(::LeftGroupOperation)

Return the inverse of the LeftGroupOperation, that is the InverseLeftGroupOperation.

Base.inv — Method

inv(::RightGroupOperation)

Return the inverse of the RightGroupOperation, that is the InverseRightGroupOperation.

LieGroups.diff_apply! — Method

diff_apply(A::GroupOperationAction, g, p, X)

For the group operation action $σ_g(p)$, compute the differential $D_p σ_g(p): T_p\mathcal G → T_{σ_g(p)}\mathcal M$, that is

for the LeftGroupOperation this calls diff_right_compose(G, g, p, X), since here $σ_g(p) = g∘p$

for the RightGroupOperation this calls diff_left_compose(G, p, g, X), since here $σ_g(p) = p∘g$

for the InverseLeftGroupOperation this calls diff_right_compose with $g^{-1}$, since here $σ_g(p) = g^{-1}∘p$

for the InverseRightGroupOperation this calls diff_left_compose with $g^{-1}$, since here $σ_g(p) = p∘g^{-1}$

LieGroups.diff_apply — Method

diff_apply(A::GroupOperationAction, g, p, X)

For the group operation action $σ_g(p)$, compute the differential $D_p σ_g(p): T_p\mathcal G → T_{σ_g(p)}\mathcal M$, that is

for the LeftGroupOperation this calls diff_right_compose(G, g, p, X), since here $σ_g(p) = g∘p$

for the RightGroupOperation this calls diff_left_compose(G, p, g, X), since here $σ_g(p) = p∘g$

for the InverseLeftGroupOperation this calls diff_right_compose with $g^{-1}$, since here $σ_g(p) = g^{-1}∘p$

for the InverseRightGroupOperation this calls diff_left_compose with $g^{-1}$, since here $σ_g(p) = p∘g^{-1}$

LieGroups.diff_group_apply! — Method

diff_apply(A::GroupOperationAction, g, p, X)

For the group operation action $σ_g(p)$, compute the differential $D_p σ_g(p): T_p\mathcal G → T_{σ_g(p)}\mathcal M$, that is

for the LeftGroupOperation this calls diff_right_compose(G, g, p, X), since here $σ_g(p) = g∘p$

for the RightGroupOperation this calls diff_left_compose(G, p, g, X), since here $σ_g(p) = p∘g$

for the InverseLeftGroupOperation this calls diff_right_compose with $g^{-1}$, since here $σ_g(p) = g^{-1}∘p$

for the InverseRightGroupOperation this calls diff_left_compose with $g^{-1}$, since here $σ_g(p) = p∘g^{-1}$

LieGroups.diff_group_apply — Method

diff_apply(A::GroupOperationAction, g, p, X)

For the group operation action $σ_g(p)$, compute the differential $D_p σ_g(p): T_p\mathcal G → T_{σ_g(p)}\mathcal M$, that is

for the LeftGroupOperation this calls diff_right_compose(G, g, p, X), since here $σ_g(p) = g∘p$

for the RightGroupOperation this calls diff_left_compose(G, p, g, X), since here $σ_g(p) = p∘g$

for the InverseLeftGroupOperation this calls diff_right_compose with $g^{-1}$, since here $σ_g(p) = g^{-1}∘p$

for the InverseRightGroupOperation this calls diff_left_compose with $g^{-1}$, since here $σ_g(p) = p∘g^{-1}$

LieGroups.switch — Method

switch(::InverseLeftGroupOperation)

Return the InverseRightGroupOperation, that is, turns $σ_g = g^{-1}∘h$ into $τ_g(h) = h∘g^{-1}$

LieGroups.switch — Method

switch(::InverseRightGroupOperation)

Return the InverseLeftGroupOperation, that is, turns $σ_g = h∘g^{-1}$ into $τ_g(h) = g^{-1}∘h$

LieGroups.switch — Method

switch(::LeftGroupOperation)

Return the RightGroupOperation, that is, turns $σ_g = g∘h$ into $τ_g(h) = h∘g$

LieGroups.switch — Method

switch(::RightGroupOperation)

Return the LeftGroupOperation, that is, turns $σ_g = h∘g$ into $τ_g(h) = g∘h$

Manifolds.apply! — Method

apply(A::GroupOperationAction, g, h)
apply!(A::GroupOperationAction, k, g, h)

apply the stored group operation action, using compose $∘$. this can be done in-place of k.

Manifolds.apply — Method

apply(A::GroupOperationAction, g, h)
apply!(A::GroupOperationAction, k, g, h)

apply the stored group operation action, using compose $∘$. this can be done in-place of k.

Literature

[HN12]

J. Hilgert and K.-H. Neeb. Structure and Geometry of Lie Groups (Springer Monographs in Mathematics, 2012).