Lie group operations

LieGroups.AbstractGroupOperation — Type

AbstractGroupOperation

Represent a type of group operation for a LieGroup $\mathcal G$, that is a smooth binary operation $∘ : \mathcal G × \mathcal G → \mathcal G$ on elements of a Lie group $\mathcal G$.

LieGroups.Identity — Type

Identity{O<:AbstractGroupOperation}

Represent the group identity element $e ∈ \mathcal G$ on a LieGroup $\mathcal G$ with AbstractGroupOperation of type O.

Similar to the philosophy that points are agnostic of their group at hand, the identity does not store the group $\mathcal G$ it belongs to. However it depends on the type of the AbstractGroupOperation used.

See also identity_element on how to obtain the corresponding AbstractManifoldPoint or array representation.

Constructors

Identity(::LieGroup{𝔽,O}) where {𝔽,M,O<:AbstractGroupOperation}
Identity(o::AbstractGroupOperation)
Identity(::Type{AbstractGroupOperation})

create the identity of the corresponding subtype O<:AbstractGroupOperation

For some generic group operations, generic implementations are available.

Additive group operation

LieGroups.AdditionGroupOperation — Type

AdditionGroupOperation <: AbstractGroupOperation

A group operation that is realised introducing defaults that fall back to + and - being overloaded, for example _compose(G::LieGroup{𝔽,AdditionGroupOperation}, a,b) = a+b

Base.exp — Method

exp(G::LieGroup{𝔽,AdditionGroupOperation}, e::Identity{AdditionGroupOperation}, X, t=1)
exp!(G::LieGroup{𝔽,AdditionGroupOperation}, g, e::Identity{AdditionGroupOperation}, X, t)

Compute the Lie group exponential on a LieGroup with an AdditionGroupOperation. This can be computed in-place of g.

Since e is just the zero-element with respect to the corresponding +, the formula reads $g=0+X=X$.

Base.inv — Method

inv(G::LieGroup{𝔽,AdditionGroupOperation}, g)
inv!(G::LieGroup{𝔽,AdditionGroupOperation}, h, g)

Compute the inverse group element $g^{-1}$, which for the AdditionGroupOperation simplifies to $-g$. This can be done in-place of h.

Base.log — Method

log(G::LieGroup{𝔽,AdditionGroupOperation}, e::Identity{AdditionGroupOperation}, g)
log!(G::LieGroup{𝔽,AdditionGroupOperation}, X, e::Identity{AdditionGroupOperation}, g)

Compute the Lie group logarithm on a LieGroup with an AdditionGroupOperation. This can be computed in-place of X.

Since e is just the zero-element with respect to the corresponding +, the formula reads $X=g-0=g$.

LieGroups.compose! — Method

compose(G::LieGroup{𝔽,AdditionGroupOperation}, g, h)
compose!(G::LieGroup{𝔽,AdditionGroupOperation}, k, g, h)

Copmute the group operation composition of g and h with respect to the AdditionGroupOperation on G, which falls back to calling g+h, where + is assumed to be overloaded accordingly.

This can be computed in-place of k.

LieGroups.diff_conjugate! — Method

diff_conjugate((G::LieGroup{𝔽,AdditionGroupOperation}, Y, g, h, X))
diff_conjugate!(G::LieGroup{𝔽,AdditionGroupOperation}, Y, g, h, X)

Compute the differential of the conjutage ``$c_g(h) = g∘h∘g^{-1} = g+h-g = h$, which simplifies for [AdditionGroupOperation] to$D(c_g(h))[X] = X$.

LieGroups.diff_conjugate — Method

diff_conjugate((G::LieGroup{𝔽,AdditionGroupOperation}, Y, g, h, X))
diff_conjugate!(G::LieGroup{𝔽,AdditionGroupOperation}, Y, g, h, X)

Compute the differential of the conjutage ``$c_g(h) = g∘h∘g^{-1} = g+h-g = h$, which simplifies for [AdditionGroupOperation] to$D(c_g(h))[X] = X$.

LieGroups.diff_inv! — Method

diff_inv(G::LieGroup{𝔽,AdditionGroupOperation}, g, X)
diff_inv!(G::LieGroup{𝔽,AdditionGroupOperation}, Y, g, X)

Compute the differential of the inverse operation $ι_{\mathcal G}(g) = g^-1$ = -g$, which simplifies for [`AdditionGroupOperation`] to$Dι_{\mathcal G}(g)[X] = -X``

LieGroups.diff_inv — Method

diff_inv(G::LieGroup{𝔽,AdditionGroupOperation}, g, X)
diff_inv!(G::LieGroup{𝔽,AdditionGroupOperation}, Y, g, X)

Compute the differential of the inverse operation $ι_{\mathcal G}(g) = g^-1$ = -g$, which simplifies for [`AdditionGroupOperation`] to$Dι_{\mathcal G}(g)[X] = -X``

LieGroups.diff_left_compose! — Method

diff_left_compose(G::LieGroup{𝔽,AdditionGroupOperation}, g, h, X)
diff_left_compose!(G::LieGroup{𝔽,AdditionGroupOperation}, Y, g, h, X)

Compute the differential of the left group multiplication $λ_g(h) = g∘h$, which simplifies for [AdditionGroupOperation] to $Dλ_g(h)[X] = X$.

LieGroups.diff_left_compose — Method

diff_left_compose(G::LieGroup{𝔽,AdditionGroupOperation}, g, h, X)
diff_left_compose!(G::LieGroup{𝔽,AdditionGroupOperation}, Y, g, h, X)

Compute the differential of the left group multiplication $λ_g(h) = g∘h$, which simplifies for [AdditionGroupOperation] to $Dλ_g(h)[X] = X$.

LieGroups.diff_right_compose! — Method

diff_right_compose(G::LieGroup{𝔽,AdditionGroupOperation}, h, g, X)
diff_right_compose!(G::LieGroup{𝔽,AdditionGroupOperation}, Y, h, g, X)

Compute the differential of the right group multiplication $ρ_g(h) = h∘g$, which simplifies for [AdditionGroupOperation] to $Dρ_g(h)[X] = X$.

LieGroups.diff_right_compose — Method

diff_right_compose(G::LieGroup{𝔽,AdditionGroupOperation}, h, g, X)
diff_right_compose!(G::LieGroup{𝔽,AdditionGroupOperation}, Y, h, g, X)

Compute the differential of the right group multiplication $ρ_g(h) = h∘g$, which simplifies for [AdditionGroupOperation] to $Dρ_g(h)[X] = X$.

LieGroups.identity_element! — Method

identity_element(G::LieGroup{𝔽,AdditionGroupOperation})
identity_element!(G::LieGroup{𝔽,AdditionGroupOperation}, e)

Return the a point representation of the Identity, which for the AdditionGroupOperation is the zero element or array.

LieGroups.inv! — Method

inv(G::LieGroup{𝔽,AdditionGroupOperation}, g)
inv!(G::LieGroup{𝔽,AdditionGroupOperation}, h, g)

Compute the inverse group element $g^{-1}$, which for the AdditionGroupOperation simplifies to $-g$. This can be done in-place of h.

LieGroups.lie_bracket! — Method

lie_bracket!(𝔤::LieAlgebra{𝔽,<:LieGroup{𝔽,AdditionGroupOperation}}, X, Y)
lie_bracket!(𝔤::LieAlgebra{𝔽,<:LieGroup{𝔽,AdditionGroupOperation}}, Z, X, Y)

Compute the Lie bracket $[⋅,⋅]: \mathfrak g×\mathfrak g → \mathfrak g$, which for the for the AdditionGroupOperation simplifies to the correspondinzero_vector. The computation can be done in-place of Z.

LieGroups.lie_bracket — Method

lie_bracket!(𝔤::LieAlgebra{𝔽,<:LieGroup{𝔽,AdditionGroupOperation}}, X, Y)
lie_bracket!(𝔤::LieAlgebra{𝔽,<:LieGroup{𝔽,AdditionGroupOperation}}, Z, X, Y)

Compute the Lie bracket $[⋅,⋅]: \mathfrak g×\mathfrak g → \mathfrak g$, which for the for the AdditionGroupOperation simplifies to the correspondinzero_vector. The computation can be done in-place of Z.

Manifolds.compose — Method

compose(G::LieGroup{𝔽,AdditionGroupOperation}, g, h)
compose!(G::LieGroup{𝔽,AdditionGroupOperation}, k, g, h)

Copmute the group operation composition of g and h with respect to the AdditionGroupOperation on G, which falls back to calling g+h, where + is assumed to be overloaded accordingly.

This can be computed in-place of k.

Manifolds.identity_element — Method

identity_element(G::LieGroup{𝔽,AdditionGroupOperation})
identity_element!(G::LieGroup{𝔽,AdditionGroupOperation}, e)

Return the a point representation of the Identity, which for the AdditionGroupOperation is the zero element or array.

ManifoldsBase.exp! — Method

exp(G::LieGroup{𝔽,AdditionGroupOperation}, e::Identity{AdditionGroupOperation}, X, t=1)
exp!(G::LieGroup{𝔽,AdditionGroupOperation}, g, e::Identity{AdditionGroupOperation}, X, t)

Compute the Lie group exponential on a LieGroup with an AdditionGroupOperation. This can be computed in-place of g.

Since e is just the zero-element with respect to the corresponding +, the formula reads $g=0+X=X$.

ManifoldsBase.log! — Method

log(G::LieGroup{𝔽,AdditionGroupOperation}, e::Identity{AdditionGroupOperation}, g)
log!(G::LieGroup{𝔽,AdditionGroupOperation}, X, e::Identity{AdditionGroupOperation}, g)

Compute the Lie group logarithm on a LieGroup with an AdditionGroupOperation. This can be computed in-place of X.

Since e is just the zero-element with respect to the corresponding +, the formula reads $X=g-0=g$.